Dashboard module's public now uses Gulp
