+++ app/dashboard/public/css/style.css
@@ -1,0 +1,1 @@
+div.tagsinput{border:1px solid #CCC;background:#FFF;padding:5px;width:300px;height:100px;overflow-y:auto}div.tagsinput span.tag{border:1px solid #a5d24a;-moz-border-radius:2px;-webkit-border-radius:2px;display:block;float:left;padding:5px;text-decoration:none;background:#cde69c;color:#638421;margin-right:5px;margin-bottom:5px;font-family:helvetica;font-size:13px}div.tagsinput span.tag a{font-weight:bold;color:#82ad2b;text-decoration:none;font-size:11px}div.tagsinput input{width:80px;margin:0px;font-family:helvetica;font-size:13px;border:1px solid transparent;padding:5px;background:transparent;color:#000;outline:0px;margin-right:5px;margin-bottom:5px}div.tagsinput div{display:block;float:left}.tags_clear{clear:both;width:100%;height:0px}.not_valid{background:#FBD8DB !important;color:#90111A !important}.container{width:1170px}h1{border-bottom:1px solid #e5e5e5;padding-bottom:15px}h3{font-size:20px}body{padding-bottom:96px}#mainFooter{border-top:1px solid #e5e5e5;background:#f2f2f2;padding:10px 0;font-size:90%;position:fixed;bottom:0;left:0;width:100%;z-index:999}#mainFooter p:last-child{margin-bottom:0}#new_events a{color:orange}#new_events a:hover{color:#fff}#check_summary li strong{font-size:2.5em}#check_summary li .label{font-size:1em;margin:10px;vertical-align:37%}div.btn-toolbar{height:20px}#welcome{text-align:center;font-size:1.5em;margin:180px 0}#events ul{margin-top:1em;margin-left:0;list-style:none;padding-left:0}#events li{padding-bottom:0.5em}#events li.new{background-color:#ffecca}#events li.new:first-child{padding-top:0.5em}#singleStat .row{text-align:center;font-size:4em}#singleStat .row h3{margin:30px 0 10px 0}.uptimeBar{height:15px;position:relative;background:#bfd255;background:-webkit-gradient(linear, left top, left bottom, color-stop(0%, #bfd255), color-stop(50%, #8eb92a), color-stop(51%, #72aa00), color-stop(100%, #9ecb2d));background:-webkit-linear-gradient(top, #bfd255 0%, #8eb92a 50%, #72aa00 51%, #9ecb2d 100%);background:linear-gradient(to bottom, #bfd255 0%, #8eb92a 50%, #72aa00 51%, #9ecb2d 100%);filter:progid:DXImageTransform.Microsoft.gradient( startColorstr='#bfd255', endColorstr='#9ecb2d',GradientType=0 )}.uptimeBar div{position:absolute;height:inherit;margin:0}.uptimeBar .down{min-width:1px;background:#efc5ca;background:-webkit-gradient(linear, left top, left bottom, color-stop(0%, #efc5ca), color-stop(50%, #d24b5a), color-stop(51%, #ba2737), color-stop(100%, #f18e99));background:-webkit-linear-gradient(top, #efc5ca 0%, #d24b5a 50%, #ba2737 51%, #f18e99 100%);background:linear-gradient(to bottom, #efc5ca 0%, #d24b5a 50%, #ba2737 51%, #f18e99 100%);filter:progid:DXImageTransform.Microsoft.gradient( startColorstr='#efc5ca', endColorstr='#f18e99',GradientType=0 )}.uptimeBar .down:hover{background:#b94a48}.uptimeBar .pause{background:#6db3f2;background:-webkit-gradient(linear, left top, left bottom, color-stop(0%, #6db3f2), color-stop(50%, #54a3ee), color-stop(51%, #3690f0), color-stop(100%, #1e69de));background:-webkit-linear-gradient(top, #6db3f2 0%, #54a3ee 50%, #3690f0 51%, #1e69de 100%);background:linear-gradient(to bottom, #6db3f2 0%, #54a3ee 50%, #3690f0 51%, #1e69de 100%);filter:progid:DXImageTransform.Microsoft.gradient( startColorstr='#6db3f2', endColorstr='#1e69de',GradientType=0 )}.uptimeBar .pause:hover{background:#3a87ad}.timeline .uptimeBar{height:22px}.week-end{background-color:#eee}tfoot{color:#DDD}table .header{cursor:pointer}table .header:after{content:"";float:right;margin-top:7px;border-width:0 4px 4px;border-style:solid;border-color:#000 transparent;visibility:hidden}table .header:hover:after{visibility:visible}table .headerSortUp,table .headerSortDown{background-color:#f7f7f9;text-shadow:0 1px 1px rgba(255,255,255,0.75)}table .headerSortDown:after,table .headerSortDown:hover:after{visibility:visible;opacity:0.6}table .headerSortUp:after{border-bottom:none;border-left:4px solid transparent;border-right:4px solid transparent;border-top:4px solid #000;visibility:visible;box-shadow:none;opacity:0.6}div.tagsinput span.tag span{cursor:pointer}table.chart td{width:12%}table.chart th.date{width:30%}table.chart th.uptimeCol{width:22%}table.ping .error{max-width:350px;word-wrap:break-word}#sectionList section{padding-top:88px}#sectionList strong{font-size:1.5em}#dateNavigationContainer{min-height:88px}#dateNavigation.affix{top:0;background-color:#fff;z-index:50;width:1140px}#secondaryNav.affix{top:91px;width:140px}#dateNavigation .title{text-align:center;margin:5px}#dateNavigation .title .btn-link{padding:0;border:none;vertical-align:inherit}#dateNavigation .timeline{margin-left:19px;height:22px}#dateNavigation .btn.month,#dateNavigation .btn.day,#dateNavigation .btn.hour,#dateNavigation .btn.quarter{padding-right:0;padding-left:0}#dateNavigation .btn.month{width:76px}#dateNavigation .btn.day.nb31{width:30px}#dateNavigation .btn.day.nb30{width:35px}#dateNavigation .btn.day.nb29{width:36px}#dateNavigation .btn.day.nb28{width:47px}#dateNavigation .btn.hour{width:38px}#dateNavigation .btn.tenminutes{width:152px}ul.nav-tabs li span{font-weight:bold}.graph{margin:20px 0 0 25px;width:600px;height:200px}#checkUptimeBars li{padding:0 0 1em 4px;list-style-type:none}#checkUptimeBars li .name{margin-bottom:2px}#check_24h+a.btn{margin-left:10px}.uneditable-input{font-weight:normal;color:#737373}.form-actions{margin-top:26px}.help,.help-block{font-size:90%}.btn-default.disabled,.btn-default.disabled.active,.btn-default.disabled.focus,.btn-default.disabled:active,.btn-default.disabled:focus,.btn-default.disabled:hover,.btn-default[disabled],.btn-default[disabled].active,.btn-default[disabled].focus,.btn-default[disabled]:active,.btn-default[disabled]:focus,.btn-default[disabled]:hover,fieldset[disabled] .btn-default,fieldset[disabled] .btn-default.active,fieldset[disabled] .btn-default.focus,fieldset[disabled] .btn-default:active,fieldset[disabled] .btn-default:focus,fieldset[disabled] .btn-default:hover{background:#ededed}#checkEvents .badge{display:inline-block;font-weight:normal}#checkEvents time{display:inline-block;min-width:64px;font-size:80%}#checkEvents span.msg{margin-left:8px}li.event-item{padding:.5em;border-bottom:1px solid #eee}
