+++ app/dashboard/public/img/apple-touch-icon-ipad.png
