+++ app/dashboard/public/img/favicon.ico
