+++ app/dashboard/public/js/scripts.js
@@ -1,0 +1,5 @@
+function require(t){var e=require.resolve(t),i=require.modules[e];if(!i)throw new Error('failed to require "'+t+'"');return i.exports||(i.exports={},i.call(i.exports,i,i.exports,require.relative(e))),i.exports}!function(t){"use strict";var e=function(e,i){this.options=t.extend({},t.fn.affix.defaults,i),this.$window=t(window).on("scroll.affix.data-api",t.proxy(this.checkPosition,this)).on("click.affix.data-api",t.proxy(function(){setTimeout(t.proxy(this.checkPosition,this),1)},this)),this.$element=t(e),this.checkPosition()};e.prototype.checkPosition=function(){if(this.$element.is(":visible")){var e,i=t(document).height(),n=this.$window.scrollTop(),a=this.$element.offset(),r=this.options.offset,o=r.bottom,s=r.top,l="affix affix-top affix-bottom";"object"!=typeof r&&(o=s=r),"function"==typeof s&&(s=r.top()),"function"==typeof o&&(o=r.bottom()),e=null!=this.unpin&&n+this.unpin<=a.top?!1:null!=o&&a.top+this.$element.height()>=i-o?"bottom":null!=s&&s>=n?"top":!1,this.affixed!==e&&(this.affixed=e,this.unpin="bottom"==e?a.top-n:null,this.$element.removeClass(l).addClass("affix"+(e?"-"+e:"")))}},t.fn.affix=function(i){return this.each(function(){var n=t(this),a=n.data("affix"),r="object"==typeof i&&i;a||n.data("affix",a=new e(this,r)),"string"==typeof i&&a[i]()})},t.fn.affix.Constructor=e,t.fn.affix.defaults={offset:0},t(window).on("load",function(){t('[data-spy="affix"]').each(function(){var e=t(this),i=e.data();i.offset=i.offset||{},i.offsetBottom&&(i.offset.bottom=i.offsetBottom),i.offsetTop&&(i.offset.top=i.offsetTop),e.affix(i)})})}(window.jQuery),!function(t){"use strict";var e=function(e,i){this.$element=t(e),this.options=t.extend({},t.fn.button.defaults,i)};e.prototype={constructor:e,setState:function(t){var e="disabled",i=this.$element,n=i.data(),a=i.is("input")?"val":"html";t+="Text",n.resetText||i.data("resetText",i[a]()),i[a](n[t]||this.options[t]),setTimeout(function(){"loadingText"==t?i.addClass(e).attr(e,e):i.removeClass(e).removeAttr(e)},0)},toggle:function(){var t=this.$element.parent('[data-toggle="buttons-radio"]');t&&t.find(".active").removeClass("active"),this.$element.toggleClass("active")}},t.fn.button=function(i){return this.each(function(){var n=t(this),a=n.data("button"),r="object"==typeof i&&i;a||n.data("button",a=new e(this,r)),"toggle"==i?a.toggle():i&&a.setState(i)})},t.fn.button.defaults={loadingText:"loading..."},t.fn.button.Constructor=e,t(function(){t("body").on("click.button.data-api","[data-toggle^=button]",function(e){var i=t(e.target);i.hasClass("btn")||(i=i.closest(".btn")),i.button("toggle")})})}(window.jQuery);var updateCheckState=function(t){var e="";if("undefined"==typeof t.isUp)e+=t.lastTested?'<span class="label label-info">unknown</span>':'<span class="label label-danger">down</span> <span class="label label-warning">new</span>';else{var i={};t.isPaused?(i.label="info",i.color="blue",i.text="paused",i.date=t.lastTested):t.isUp?(i.label="success",i.color="green",i.text="up",i.date=t.lastChanged):(i.label="danger",i.color="red",i.text="down",i.date=t.lastChanged),e+='<span class="label label-'+i.label+'">'+i.text+"</span>",e+='<span class="'+i.color+'"> for <span title="'+new Date(i.date)+'">'+moment(i.date).fromNow(!0)+"</span></span>"}$("#check_24h").html(e)},DateInterval=function(t,e,i,n,a){this.type=t,this.origin=i,this.url=n,this.maxZoom=a||"hour",this.listeners={},this.stat={},this.stats=[],this.setDate(e)};DateInterval.prototype.setDate=function(t){this.date=Math.max(Math.min(t,Date.now()),this.origin),this.momentForDate=moment(this.date),this.begin=this.momentForDate.clone().startOf(this.type),this.end=this.momentForDate.clone().endOf(this.type),this.trigger("change-date"),this.refreshData()},DateInterval.prototype.update=function(t,e){this.type=t,this.setDate(e)},DateInterval.prototype.beginsAfterOrigin=function(){return this.begin.valueOf()>this.origin.valueOf()},DateInterval.prototype.endsBeforeNow=function(){return this.end.valueOf()<Date.now()},DateInterval.prototype.getNextDate=function(t){return t=t||this.type,this.momentForDate.clone().add(t+"s",1)},DateInterval.prototype.getPreviousDate=function(t){return t=t||this.type,this.momentForDate.clone().subtract(t+"s",1)},DateInterval.prototype.refreshData=function(){var t=this;$.getJSON(this.url+"stats/"+this.type+this.getIntervalQueryString(),function(e){t.stats=e,t.trigger("refresh-stats")}),$.getJSON(this.url+"stat/"+this.type+"/"+this.date,function(e){t.stat=e,t.trigger("refresh-stat")})},DateInterval.prototype.getIntervalQueryString=function(){return"?begin="+this.begin.valueOf()+"&end="+this.end.valueOf()},DateInterval.prototype.types=["year","month","day","hour","tenminutes"],DateInterval.prototype.isMaxZoom=function(){return this.type===this.maxZoom},DateInterval.prototype.subType=function(t){t=t||this.type;var e=this.types.indexOf(t);return-1===e||4===e?!1:this.types[e+1]},DateInterval.prototype.superType=function(t){var e=this.types.indexOf(t);return 0>=e?!1:this.types[e-1]},DateInterval.prototype.getSubTypeDuration=function(){var t=new moment;return t.endOf(this.subType(this.type)).valueOf()-t.startOf(this.subType(this.type)).valueOf()},DateInterval.prototype.on=function(t,e){this.listeners[t]||(this.listeners[t]=[]),this.listeners[t].push(e)},DateInterval.prototype.trigger=function(t,e){if(this.listeners[t])for(var i=0;i<this.listeners[t].length;i++)this.listeners[t][i].apply(this,e||[])};var NAV_BTN_WIDTH=19,DateNavigation=function(t,e,i){this.interval=t,this.initialType=t.type,this.initialDate=t.date,this.init(e,i)};DateNavigation.prototype.init=function(t,e){this.redraw();var i=this.interval;i.on("change-date",this.redraw.bind(this)),$("#dateNavigation").on("click","a",function(){var t=$(this).data();i.update(t.type,parseInt(t.date))}),setInterval(this.redrawPeriods.bind(this),3e5),i.on("refresh-stat",function(){var n=this.stat?this.stat.outages||[]:[],a={from:i.begin.valueOf(),to:i.end.valueOf(),periods:n};"check"==t?a.check=e:"tag"==t&&(a.origin=this.origin.valueOf()),$("#dateNavigation .timeline").html(uptimeBar(t,a))}),$("#dateNavigation").affix({offset:$("#dateNavigation").position()});var n=window.history.state,a=location.href;this.pushStateEnabled=!1;var r=this;i.on("change-date",function(){r.pushStateEnabled&&history.pushState&&history.pushState({type:this.type,date:this.date,stat:this.stat,stats:this.stats},null,"?type="+this.type+"&date="+this.date+location.hash)}),window.addEventListener("popstate",function(t){var e=!n&&location.href==a;if(n=!0,!e&&r.pushStateEnabled){var o=t.state;if(!o){if(location.href!=a)return;o={type:r.initialType,date:r.initialDate}}r.pushStateEnabled=!1,i.update(o.type,parseInt(o.date)),r.pushStateEnabled=!0}}),this.pushStateEnabled=!0},DateNavigation.prototype.redraw=function(){this.redrawTitle(),this.redrawPeriods(),this.redrawZoom(),this.adjustWidth()},DateNavigation.prototype.titleForPeriod=function(t,e){switch(e){case"month":return t.format("MMMM");case"day":return t.format("D");case"hour":return t.format("ha");case"tenminutes":return t.format("h:mma")+" to "+t.clone().add("minutes",10).subtract("seconds",1).format("h:mma")}},DateNavigation.prototype.tooltipForPeriod=function(t,e){switch(e){case"month":return t.format("MMMM YYYY");case"day":return t.format("dddd, LL");case"hour":return"from "+t.format("h:mma")+" to "+t.clone().endOf("hour").format("h:mma");case"tenminutes":return""}},DateNavigation.prototype.redrawPeriods=function(){var t=this.interval.type,e='<div class="btn-group btn-group-xs">';e+=this.interval.beginsAfterOrigin()?'<a class="btn btn-default" data-type="'+t+'" data-date="'+this.interval.getPreviousDate()+'">&lt;</a>':'<a class="btn btn-default" disabled="disabled">&lt;</a>';for(var i=this.interval.begin,n=this.interval.end,a=this.interval.subType(t),r=i.clone();r.valueOf()<n.valueOf();)e+=r.valueOf()<Date.now()&&r.clone().endOf(a).valueOf()>this.interval.origin&&!this.interval.isMaxZoom()?'<a class="btn btn-default '+a+" nb"+n.date()+'" data-type="'+a+'" data-date="'+r.valueOf()+'" title="'+this.tooltipForPeriod(r,a)+'">'+this.titleForPeriod(r,a)+"</a>":'<a class="btn btn-default '+a+" nb"+n.date()+'" disabled="disabled">'+this.titleForPeriod(r,a)+"</a>","tenminutes"==a?r.add("minutes",10):r.add(a+"s",1);e+=this.interval.endsBeforeNow()?'<a class="btn btn-default" data-type="'+t+'" data-date="'+this.interval.getNextDate()+'">&gt;</a>':'<a class="btn btn-default" disabled="disabled">&gt;</a>',e+="</div>",$("#dateNavigation .periods").html(e)},DateNavigation.prototype.redrawTitle=function(){var t="",e=this.interval.momentForDate;switch(this.interval.type){case"year":t+=e.year();break;case"month":t+=e.format("MMMM"),t+=' <div class="btn-group btn-group-xs"><a class="btn btn-link" data-type="year" data-date="'+this.interval.date+'">',t+=e.year(),t+="</a></div>";break;case"day":t+=e.format("dddd "),t+='<div class="btn-group btn-group-xs"><a class="btn btn-link" data-type="month" data-date="'+this.interval.date+'">',t+=e.format("MMMM"),t+="</a></div> ",t+=e.format("Do, YYYY");break;case"hour":t+='<div class="btn-group btn-group-xs"><a class="btn btn-link" data-type="day" data-date="'+this.interval.date+'">',t+=e.format("dddd MMMM Do"),t+="</a></div>, ",t+=e.clone().startOf("hour").format("ha")+" to "+e.clone().endOf("hour").format("h:mma")}$("#dateNavigation .title").html(t)},DateNavigation.prototype.redrawZoom=function(){var t="",e=this.interval.subType(this.interval.type);t+=e===!1||this.interval.isMaxZoom()?'<a class="btn btn-sm btn-default" disabled><i class="fa fa-search-plus"></i></a>':'<a class="btn btn-sm btn-default" data-type="'+e+'" data-date="'+this.interval.date+'"><i class="fa fa-search-plus"></i></a>';var i=this.interval.superType(this.interval.type);t+=i!==!1?'<a class="btn btn-sm btn-default" data-type="'+i+'" data-date="'+this.interval.date+'"><li class="fa fa-search-minus"></li></a>':'<a class="btn btn-sm btn-default" disabled><i class="fa fa-search-minus"></i></a>',$("#dateNavigation .zoom").html(t)},DateNavigation.prototype.adjustWidth=function(){var t=$("#dateNavigation .periods .btn-group").width()-2*NAV_BTN_WIDTH;$("#dateNavigation .timeline").width(t),$(".adjustedOnDateNavigation").width(t)},require.modules={},require.resolve=function(t){var e=t,i=t+".js",n=t+"/index.js";return require.modules[i]&&i||require.modules[n]&&n||e},require.register=function(t,e){require.modules[t]=e},require.relative=function(t){return function(e){if("."!=e.substr(0,1))return require(e);var i=t.split("/"),n=e.split("/");i.pop();for(var a=0;a<n.length;a++){var r=n[a];".."==r?i.pop():"."!=r&&i.push(r)}return require(i.join("/"))}},require.register("ejs.js",function(t,e,i){function n(t){return t.substr(1).split("|").reduce(function(t,e){var i=e.split(":"),n=i.shift(),a=i.shift()||"";return a&&(a=", "+a),"filters."+n+"("+t+a+")"})}function a(t,e,i,n){var a=e.split("\n"),r=Math.max(n-3,0),o=Math.min(a.length,n+3),s=a.slice(r,o).map(function(t,e){var i=e+r+1;return(i==n?" >> ":"    ")+i+"| "+t}).join("\n");throw t.path=i,t.message=(i||"ejs")+":"+n+"\n"+s+"\n\n"+t.message,t}var r=i("./utils");e.version="0.5.0";var o=e.filters=i("./filters"),s={};e.clearCache=function(){s={}};var l=(e.parse=function(t,i){for(var i=i||{},a=i.open||e.open||"<%",r=i.close||e.close||"%>",o=["var buf = [];","\nwith (locals) {","\n  buf.push('"],s=1,l=0,h=t.length;h>l;++l)if(t.slice(l,a.length+l)==a){l+=a.length;var c,u,d="__stack.lineno="+s;switch(t.substr(l,1)){case"=":c="', escape(("+d+", ",u=")), '",++l;break;case"-":c="', ("+d+", ",u="), '",++l;break;default:c="');"+d+";",u="; buf.push('"}for(var f=t.indexOf(r,l),p=t.substring(l,f),g=l,m=0;~(m=p.indexOf("\n",m));)m++,s++;":"==p.substr(0,1)&&(p=n(p)),o.push(c,p,u),l+=f-g+r.length-1}else"\\"==t.substr(l,1)?o.push("\\\\"):"'"==t.substr(l,1)?o.push("\\'"):"\r"==t.substr(l,1)?o.push(" "):"\n"==t.substr(l,1)?(o.push("\\n"),s++):o.push(t.substr(l,1));return o.push("');\n}\nreturn buf.join('');"),o.join("")},e.compile=function(t,i){i=i||{};var n=JSON.stringify(t),s=i.filename?JSON.stringify(i.filename):"undefined";t=["var __stack = { lineno: 1, input: "+n+", filename: "+s+" };",a.toString(),"try {",e.parse(t,i),"} catch (err) {","  rethrow(err, __stack.input, __stack.filename, __stack.lineno);","}"].join("\n"),i.debug&&console.log(t);var l=new Function("locals, filters, escape",t);return function(t){return l.call(this,t,o,r.escape)}});e.render=function(t,e){var i,e=e||{};if(e.cache){if(!e.filename)throw new Error('"cache" option requires "filename".');i=s[e.filename]||(s[e.filename]=l(t,e))}else i=l(t,e);return e.__proto__=e.locals,i.call(e.scope,e)},e.renderFile=function(t,i,n){var a=t+":string";"function"==typeof i&&(n=i,i={}),i.filename=t;try{var r=i.cache?e.cache[a]||(e.cache[a]=fs.readFileSync(t,"utf8")):fs.readFileSync(t,"utf8");n(null,e.render(r,i))}catch(o){n(o)}},e.__express=e.renderFile,i.extensions?i.extensions[".ejs"]=function(t,e){source=i("fs").readFileSync(e,"utf-8"),t._compile(l(source,{}),e)}:i.registerExtension&&i.registerExtension(".ejs",function(t){return l(t,{})})}),require.register("filters.js",function(t,e){e.first=function(t){return t[0]},e.last=function(t){return t[t.length-1]},e.capitalize=function(t){return t=String(t),t[0].toUpperCase()+t.substr(1,t.length)},e.downcase=function(t){return String(t).toLowerCase()},e.upcase=function(t){return String(t).toUpperCase()},e.sort=function(t){return Object.create(t).sort()},e.sort_by=function(t,e){return Object.create(t).sort(function(t,i){return t=t[e],i=i[e],t>i?1:i>t?-1:0})},e.size=e.length=function(t){return t.length},e.plus=function(t,e){return Number(t)+Number(e)},e.minus=function(t,e){return Number(t)-Number(e)},e.times=function(t,e){return Number(t)*Number(e)},e.divided_by=function(t,e){return Number(t)/Number(e)},e.join=function(t,e){return t.join(e||", ")},e.truncate=function(t,e){return t=String(t),t.substr(0,e)},e.truncate_words=function(t,e){var t=String(t),i=t.split(/ +/);return i.slice(0,e).join(" ")},e.replace=function(t,e,i){return String(t).replace(e,i||"")},e.prepend=function(t,e){return Array.isArray(t)?[e].concat(t):e+t},e.append=function(t,e){return Array.isArray(t)?t.concat(e):t+e},e.map=function(t,e){return t.map(function(t){return t[e]})},e.reverse=function(t){return Array.isArray(t)?t.reverse():String(t).split("").reverse().join("")},e.get=function(t,e){return t[e]},e.json=function(t){return JSON.stringify(t)}}),require.register("utils.js",function(t,e){e.escape=function(t){return String(t).replace(/&(?!\w+;)/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;")}}),!function(t,e,i){"undefined"!=typeof module?module.exports=i(t,e):"function"==typeof define&&"object"==typeof define.amd?define(i):e[t]=i(t,e)}("bean",this,function(t,e){var i=window,n=e[t],a=/over|out/,r=/[^\.]*(?=\..*)\.|.*/,o=/\..*/,s="addEventListener",l="attachEvent",h="removeEventListener",c="detachEvent",u=document||{},d=u.documentElement||{},f=d[s],p=f?s:l,g=Array.prototype.slice,m=/click|mouse|menu|drag|drop/i,v=/^touch|^gesture/i,x={one:1},y=function(t,e,i){for(i=0;i<e.length;i++)t[e[i]]=1;return t}({},("click dblclick mouseup mousedown contextmenu mousewheel DOMMouseScroll mouseover mouseout mousemove selectstart selectend keydown keypress keyup orientationchange focus blur change reset select submit load unload beforeunload resize move DOMContentLoaded readystatechange error abort scroll "+(f?"show input invalid touchstart touchmove touchend touchcancel gesturestart gesturechange gestureend message readystatechange pageshow pagehide popstate hashchange offline online afterprint beforeprint dragstart dragenter dragover dragleave drag drop dragend loadstart progress suspend emptied stalled loadmetadata loadeddata canplay canplaythrough playing waiting seeking seeked ended durationchange timeupdate play pause ratechange volumechange cuechange checking noupdate downloading cached updateready obsolete ":"")).split(" ")),b=function(){function t(t,e){for(;null!==(e=e.parentNode);)if(e===t)return!0;return!1}function e(e){var i=e.relatedTarget;return i?i!==this&&"xul"!==i.prefix&&!/document/.test(this.toString())&&!t(this,i):null===i}return{mouseenter:{base:"mouseover",condition:e},mouseleave:{base:"mouseout",condition:e},mousewheel:{base:/Firefox/.test(navigator.userAgent)?"DOMMouseScroll":"mousewheel"}}}(),w=function(){var t="altKey attrChange attrName bubbles cancelable ctrlKey currentTarget detail eventPhase getModifierState isTrusted metaKey relatedNode relatedTarget shiftKey srcElement target timeStamp type view which".split(" "),e=t.concat("button buttons clientX clientY dataTransfer fromElement offsetX offsetY pageX pageY screenX screenY toElement".split(" ")),i=t.concat("char charCode key keyCode".split(" ")),n=t.concat("touches targetTouches changedTouches scale rotation".split(" ")),r="preventDefault",o=function(t){return function(){t[r]?t[r]():t.returnValue=!1}},s="stopPropagation",l=function(t){return function(){t[s]?t[s]():t.cancelBubble=!0}},h=function(t){return function(){t[r](),t[s](),t.stopped=!0}},c=function(t,e,i){var n,a;for(n=i.length;n--;)a=i[n],!(a in e)&&a in t&&(e[a]=t[a])};return function(f,p){var g={originalEvent:f,isNative:p};if(!f)return g;var x,y=f.type,b=f.target||f.srcElement;return g[r]=o(f),g[s]=l(f),g.stop=h(g),g.target=b&&3===b.nodeType?b.parentNode:b,p&&(-1!==y.indexOf("key")?(x=i,g.keyCode=f.which||f.keyCode):m.test(y)?(x=e,g.rightClick=3===f.which||2===f.button,g.pos={x:0,y:0},f.pageX||f.pageY?(g.clientX=f.pageX,g.clientY=f.pageY):(f.clientX||f.clientY)&&(g.clientX=f.clientX+u.body.scrollLeft+d.scrollLeft,g.clientY=f.clientY+u.body.scrollTop+d.scrollTop),a.test(y)&&(g.relatedTarget=f.relatedTarget||f[("mouseover"===y?"from":"to")+"Element"])):v.test(y)&&(x=n),c(f,g,x||t)),g}}(),M=function(t,e){return f||e||t!==u&&t!==i?t:d},k=function(){function t(t,e,i,n,a){this.element=t,this.type=e,this.handler=i,this.original=n,this.namespaces=a,this.custom=b[e],this.isNative=y[e]&&t[p],this.eventType=f||this.isNative?e:"propertychange",this.customType=!f&&!this.isNative&&e,this.target=M(t,this.isNative),this.eventSupport=this.target[p]}return t.prototype={inNamespaces:function(t){var e,i;if(!t)return!0;if(!this.namespaces)return!1;for(e=t.length;e--;)for(i=this.namespaces.length;i--;)if(t[e]===this.namespaces[i])return!0;return!1},matches:function(t,e,i){return!(this.element!==t||e&&this.original!==e||i&&this.handler!==i)}},t}(),T=function(){var t={},e=function(i,n,a,r,o){if(n&&"*"!==n){var s,l=0,h=t["$"+n],c="*"===i;if(!h)return;for(s=h.length;s>l;l++)if((c||h[l].matches(i,a,r))&&!o(h[l],h,l,n))return}else for(var u in t)"$"===u.charAt(0)&&e(i,u.substr(1),a,r,o)},i=function(e,i,n){var a,r=t["$"+i];if(r)for(a=r.length;a--;)if(r[a].matches(e,n,null))return!0;return!1},n=function(t,i,n){var a=[];return e(t,i,n,null,function(t){return a.push(t)}),a},a=function(e){return(t["$"+e.type]||(t["$"+e.type]=[])).push(e),e},r=function(i){e(i.element,i.type,null,i.handler,function(e,i,n){return i.splice(n,1),0===i.length&&delete t["$"+e.type],!1})},o=function(){var e,i=[];for(e in t)"$"===e.charAt(0)&&(i=i.concat(t[e]));return i};return{has:i,get:n,put:a,del:r,entries:o}}(),S=f?function(t,e,i,n){t[n?s:h](e,i,!1)}:function(t,e,i,n,a){a&&n&&null===t["_on"+a]&&(t["_on"+a]=0),t[n?l:c]("on"+e,i)},F=function(t,e,n){return function(a){return a=w(a||((this.ownerDocument||this.document||this).parentWindow||i).event,!0),e.apply(t,[a].concat(n))}},_=function(t,e,n,a,r,o){return function(s){(a?a.apply(this,arguments):f?!0:s&&s.propertyName==="_on"+n||!s)&&(s&&(s=w(s||((this.ownerDocument||this.document||this).parentWindow||i).event,o)),e.apply(t,!s||r&&0!==r.length?g.call(arguments,s?0:1).concat(r):arguments))}},C=function(t,e,i,n,a){return function(){t(e,i,a),n.apply(this,arguments)}},D=function(t,e,i,n){var a,r,s,l=e&&e.replace(o,""),h=T.get(t,l,i);for(a=0,r=h.length;r>a;a++)h[a].inNamespaces(n)&&((s=h[a]).eventSupport&&S(s.target,s.eventType,s.handler,!1,s.type),T.del(s))},A=function(t,e,i,n,a){var s,l=e.replace(o,""),h=e.replace(r,"").split(".");return T.has(t,l,i)?t:("unload"===l&&(i=C(D,t,l,i,n)),b[l]&&(b[l].condition&&(i=_(t,i,l,b[l].condition,!0)),l=b[l].base||l),s=T.put(new k(t,l,i,n,h[0]&&h)),s.handler=s.isNative?F(t,s.handler,a):_(t,s.handler,l,!1,a,!1),s.eventSupport&&S(s.target,s.eventType,s.handler,!0,s.customType),void 0)},O=function(t,e,i){return function(n){var a,r,o="string"==typeof t?i(t,this):t;for(a=n.target;a&&a!==this;a=a.parentNode)for(r=o.length;r--;)if(o[r]===a)return e.apply(a,arguments)}},z=function(t,e,i){var n,a,s,l,h=D,c=e&&"string"==typeof e;if(c&&e.indexOf(" ")>0){for(e=e.split(" "),l=e.length;l--;)z(t,e[l],i);return t}if(a=c&&e.replace(o,""),a&&b[a]&&(a=b[a].type),!e||c)(s=c&&e.replace(r,""))&&(s=s.split(".")),h(t,a,i,s);else if("function"==typeof e)h(t,null,e);else for(n in e)e.hasOwnProperty(n)&&z(t,n,e[n]);return t},L=function(t,e,i,n,a){var r,o,s,l,h=i,c=i&&"string"==typeof i;if(e&&!i&&"object"==typeof e)for(r in e)e.hasOwnProperty(r)&&L.apply(this,[t,r,e[r]]);else for(l=arguments.length>3?g.call(arguments,3):[],o=(c?i:e).split(" "),c&&(i=O(e,h=n,a))&&(l=g.call(l,1)),this===x&&(i=C(z,t,e,i,h)),s=o.length;s--;)A(t,o[s],i,h,l);return t},I=function(){return L.apply(x,arguments)},H=f?function(t,e,n){var a=u.createEvent(t?"HTMLEvents":"UIEvents");a[t?"initEvent":"initUIEvent"](e,!0,!0,i,1),n.dispatchEvent(a)}:function(t,e,i){i=M(i,t),t?i.fireEvent("on"+e,u.createEventObject()):i["_on"+e]++},P=function(t,e,i){var n,a,s,l,h,c=e.split(" ");for(n=c.length;n--;)if(e=c[n].replace(o,""),(l=c[n].replace(r,""))&&(l=l.split(".")),l||i||!t[p])for(h=T.get(t,e),i=[!1].concat(i),a=0,s=h.length;s>a;a++)h[a].inNamespaces(l)&&h[a].handler.apply(t,i);else H(y[e],e,t);return t},W=function(t,e,i){for(var n=0,a=T.get(e,i),r=a.length;r>n;n++)a[n].original&&L(t,a[n].type,a[n].original);return t},E={add:L,one:I,remove:z,clone:W,fire:P,noConflict:function(){return e[t]=n,this}};if(i[l]){var N=function(){var t,e=T.entries();for(t in e)e[t].type&&"unload"!==e[t].type&&z(e[t].element,e[t].type);i[c]("onunload",N),i.CollectGarbage&&i.CollectGarbage()};i[l]("onunload",N)}return E}),function(){var t=this,e=t._,i={},n=Array.prototype,a=Object.prototype,r=Function.prototype,o=n.slice,s=n.unshift,l=a.toString,h=a.hasOwnProperty,c=n.forEach,u=n.map,d=n.reduce,f=n.reduceRight,p=n.filter,g=n.every,m=n.some,v=n.indexOf,x=n.lastIndexOf,y=Array.isArray,b=Object.keys,w=r.bind,M=function(t){return new _(t)};"undefined"!=typeof module&&module.exports?(module.exports=M,M._=M):t._=M,M.VERSION="1.1.7";var k=M.each=M.forEach=function(t,e,n){if(null!=t)if(c&&t.forEach===c)t.forEach(e,n);else if(t.length===+t.length){for(var a=0,r=t.length;r>a;a++)if(a in t&&e.call(n,t[a],a,t)===i)return}else for(var o in t)if(h.call(t,o)&&e.call(n,t[o],o,t)===i)return};M.map=function(t,e,i){var n=[];return null==t?n:u&&t.map===u?t.map(e,i):(k(t,function(t,a,r){n[n.length]=e.call(i,t,a,r)}),n)},M.reduce=M.foldl=M.inject=function(t,e,i,n){var a=void 0!==i;if(null==t&&(t=[]),d&&t.reduce===d)return n&&(e=M.bind(e,n)),a?t.reduce(e,i):t.reduce(e);if(k(t,function(t,r,o){a?i=e.call(n,i,t,r,o):(i=t,a=!0)}),!a)throw new TypeError("Reduce of empty array with no initial value");return i},M.reduceRight=M.foldr=function(t,e,i,n){if(null==t&&(t=[]),f&&t.reduceRight===f)return n&&(e=M.bind(e,n)),void 0!==i?t.reduceRight(e,i):t.reduceRight(e);var a=(M.isArray(t)?t.slice():M.toArray(t)).reverse();return M.reduce(a,e,i,n)},M.find=M.detect=function(t,e,i){var n;return T(t,function(t,a,r){return e.call(i,t,a,r)?(n=t,!0):void 0}),n},M.filter=M.select=function(t,e,i){var n=[];return null==t?n:p&&t.filter===p?t.filter(e,i):(k(t,function(t,a,r){e.call(i,t,a,r)&&(n[n.length]=t)}),n)},M.reject=function(t,e,i){var n=[];return null==t?n:(k(t,function(t,a,r){e.call(i,t,a,r)||(n[n.length]=t)}),n)},M.every=M.all=function(t,e,n){var a=!0;return null==t?a:g&&t.every===g?t.every(e,n):(k(t,function(t,r,o){return(a=a&&e.call(n,t,r,o))?void 0:i}),a)};var T=M.some=M.any=function(t,e,n){e=e||M.identity;var a=!1;return null==t?a:m&&t.some===m?t.some(e,n):(k(t,function(t,r,o){return(a|=e.call(n,t,r,o))?i:void 0}),!!a)};M.include=M.contains=function(t,e){var i=!1;return null==t?i:v&&t.indexOf===v?-1!=t.indexOf(e):(T(t,function(t){return(i=t===e)?!0:void 0}),i)},M.invoke=function(t,e){var i=o.call(arguments,2);return M.map(t,function(t){return(e.call?e||t:t[e]).apply(t,i)})},M.pluck=function(t,e){return M.map(t,function(t){return t[e]})},M.max=function(t,e,i){if(!e&&M.isArray(t))return Math.max.apply(Math,t);var n={computed:-1/0};return k(t,function(t,a,r){var o=e?e.call(i,t,a,r):t;o>=n.computed&&(n={value:t,computed:o})}),n.value},M.min=function(t,e,i){if(!e&&M.isArray(t))return Math.min.apply(Math,t);var n={computed:1/0};return k(t,function(t,a,r){var o=e?e.call(i,t,a,r):t;o<n.computed&&(n={value:t,computed:o})}),n.value},M.sortBy=function(t,e,i){return M.pluck(M.map(t,function(t,n,a){return{value:t,criteria:e.call(i,t,n,a)}}).sort(function(t,e){var i=t.criteria,n=e.criteria;return n>i?-1:i>n?1:0}),"value")},M.groupBy=function(t,e){var i={};return k(t,function(t,n){var a=e(t,n);(i[a]||(i[a]=[])).push(t)}),i},M.sortedIndex=function(t,e,i){i||(i=M.identity);for(var n=0,a=t.length;a>n;){var r=n+a>>1;i(t[r])<i(e)?n=r+1:a=r}return n},M.toArray=function(t){return t?t.toArray?t.toArray():M.isArray(t)?o.call(t):M.isArguments(t)?o.call(t):M.values(t):[]},M.size=function(t){return M.toArray(t).length},M.first=M.head=function(t,e,i){return null==e||i?t[0]:o.call(t,0,e)},M.rest=M.tail=function(t,e,i){return o.call(t,null==e||i?1:e)},M.last=function(t){return t[t.length-1]},M.compact=function(t){return M.filter(t,function(t){return!!t})},M.flatten=function(t){return M.reduce(t,function(t,e){return M.isArray(e)?t.concat(M.flatten(e)):(t[t.length]=e,t)},[])},M.without=function(t){return M.difference(t,o.call(arguments,1))},M.uniq=M.unique=function(t,e){return M.reduce(t,function(t,i,n){return 0!=n&&(e===!0?M.last(t)==i:M.include(t,i))||(t[t.length]=i),t},[])},M.union=function(){return M.uniq(M.flatten(arguments))},M.intersection=M.intersect=function(t){var e=o.call(arguments,1);return M.filter(M.uniq(t),function(t){return M.every(e,function(e){return M.indexOf(e,t)>=0})})},M.difference=function(t,e){return M.filter(t,function(t){return!M.include(e,t)})},M.zip=function(){for(var t=o.call(arguments),e=M.max(M.pluck(t,"length")),i=new Array(e),n=0;e>n;n++)i[n]=M.pluck(t,""+n);return i},M.indexOf=function(t,e,i){if(null==t)return-1;var n,a;if(i)return n=M.sortedIndex(t,e),t[n]===e?n:-1;if(v&&t.indexOf===v)return t.indexOf(e);for(n=0,a=t.length;a>n;n++)if(t[n]===e)return n;return-1},M.lastIndexOf=function(t,e){if(null==t)return-1;if(x&&t.lastIndexOf===x)return t.lastIndexOf(e);for(var i=t.length;i--;)if(t[i]===e)return i;return-1},M.range=function(t,e,i){arguments.length<=1&&(e=t||0,t=0),i=arguments[2]||1;for(var n=Math.max(Math.ceil((e-t)/i),0),a=0,r=new Array(n);n>a;)r[a++]=t,t+=i;return r},M.bind=function(t,e){if(t.bind===w&&w)return w.apply(t,o.call(arguments,1));var i=o.call(arguments,2);return function(){return t.apply(e,i.concat(o.call(arguments)))}},M.bindAll=function(t){var e=o.call(arguments,1);return 0==e.length&&(e=M.functions(t)),k(e,function(e){t[e]=M.bind(t[e],t)}),t},M.memoize=function(t,e){var i={};return e||(e=M.identity),function(){var n=e.apply(this,arguments);return h.call(i,n)?i[n]:i[n]=t.apply(this,arguments)}},M.delay=function(t,e){var i=o.call(arguments,2);return setTimeout(function(){return t.apply(t,i)},e)},M.defer=function(t){return M.delay.apply(M,[t,1].concat(o.call(arguments,1)))};var S=function(t,e,i){var n;return function(){var a=this,r=arguments,o=function(){n=null,t.apply(a,r)};i&&clearTimeout(n),(i||!n)&&(n=setTimeout(o,e))}};M.throttle=function(t,e){return S(t,e,!1)},M.debounce=function(t,e){return S(t,e,!0)},M.once=function(t){var e,i=!1;return function(){return i?e:(i=!0,e=t.apply(this,arguments))}},M.wrap=function(t,e){return function(){var i=[t].concat(o.call(arguments));return e.apply(this,i)}},M.compose=function(){var t=o.call(arguments);return function(){for(var e=o.call(arguments),i=t.length-1;i>=0;i--)e=[t[i].apply(this,e)];return e[0]}},M.after=function(t,e){return function(){return--t<1?e.apply(this,arguments):void 0}},M.keys=b||function(t){if(t!==Object(t))throw new TypeError("Invalid object");var e=[];for(var i in t)h.call(t,i)&&(e[e.length]=i);return e},M.values=function(t){return M.map(t,M.identity)},M.functions=M.methods=function(t){var e=[];for(var i in t)M.isFunction(t[i])&&e.push(i);return e.sort()},M.extend=function(t){return k(o.call(arguments,1),function(e){for(var i in e)void 0!==e[i]&&(t[i]=e[i])}),t},M.defaults=function(t){return k(o.call(arguments,1),function(e){for(var i in e)null==t[i]&&(t[i]=e[i])}),t},M.clone=function(t){return M.isArray(t)?t.slice():M.extend({},t)},M.tap=function(t,e){return e(t),t},M.isEqual=function(t,e){if(t===e)return!0;var i=typeof t,n=typeof e;if(i!=n)return!1;if(t==e)return!0;if(!t&&e||t&&!e)return!1;if(t._chain&&(t=t._wrapped),e._chain&&(e=e._wrapped),t.isEqual)return t.isEqual(e);if(e.isEqual)return e.isEqual(t);if(M.isDate(t)&&M.isDate(e))return t.getTime()===e.getTime();if(M.isNaN(t)&&M.isNaN(e))return!1;if(M.isRegExp(t)&&M.isRegExp(e))return t.source===e.source&&t.global===e.global&&t.ignoreCase===e.ignoreCase&&t.multiline===e.multiline;if("object"!==i)return!1;if(t.length&&t.length!==e.length)return!1;var a=M.keys(t),r=M.keys(e);if(a.length!=r.length)return!1;for(var o in t)if(!(o in e&&M.isEqual(t[o],e[o])))return!1;return!0},M.isEmpty=function(t){if(M.isArray(t)||M.isString(t))return 0===t.length;for(var e in t)if(h.call(t,e))return!1;return!0},M.isElement=function(t){return!!t&&1==t.nodeType},M.isArray=y||function(t){return"[object Array]"===l.call(t)},M.isObject=function(t){return t===Object(t)},M.isArguments=function(t){return!!t&&!!h.call(t,"callee")},M.isFunction=function(t){return!!(t&&t.constructor&&t.call&&t.apply)},M.isString=function(t){return!!(""===t||t&&t.charCodeAt&&t.substr)},M.isNumber=function(t){return!!(0===t||t&&t.toExponential&&t.toFixed)},M.isNaN=function(t){return t!==t},M.isBoolean=function(t){return t===!0||t===!1},M.isDate=function(t){return!!(t&&t.getTimezoneOffset&&t.setUTCFullYear)},M.isRegExp=function(t){return!(!(t&&t.test&&t.exec)||!t.ignoreCase&&t.ignoreCase!==!1)},M.isNull=function(t){return null===t},M.isUndefined=function(t){return void 0===t},M.noConflict=function(){return t._=e,this},M.identity=function(t){return t},M.times=function(t,e,i){for(var n=0;t>n;n++)e.call(i,n)},M.mixin=function(t){k(M.functions(t),function(e){D(e,M[e]=t[e])})};var F=0;M.uniqueId=function(t){var e=F++;return t?t+e:e},M.templateSettings={evaluate:/<%([\s\S]+?)%>/g,interpolate:/<%=([\s\S]+?)%>/g},M.template=function(t,e){var i=M.templateSettings,n="var __p=[],print=function(){__p.push.apply(__p,arguments);};with(obj||{}){__p.push('"+t.replace(/\\/g,"\\\\").replace(/'/g,"\\'").replace(i.interpolate,function(t,e){return"',"+e.replace(/\\'/g,"'")+",'"}).replace(i.evaluate||null,function(t,e){return"');"+e.replace(/\\'/g,"'").replace(/[\r\n\t]/g," ")+"__p.push('"}).replace(/\r/g,"\\r").replace(/\n/g,"\\n").replace(/\t/g,"\\t")+"');}return __p.join('');",a=new Function("obj",n);return e?a(e):a};var _=function(t){this._wrapped=t};M.prototype=_.prototype;var C=function(t,e){return e?M(t).chain():t},D=function(t,e){_.prototype[t]=function(){var t=o.call(arguments);return s.call(t,this._wrapped),C(e.apply(M,t),this._chain)}};M.mixin(M),k(["pop","push","reverse","shift","sort","splice","unshift"],function(t){var e=n[t];_.prototype[t]=function(){return e.apply(this._wrapped,arguments),C(this._wrapped,this._chain)}}),k(["concat","join","slice"],function(t){var e=n[t];_.prototype[t]=function(){return C(e.apply(this._wrapped,arguments),this._chain)}}),_.prototype.chain=function(){return this._chain=!0,this},_.prototype.value=function(){return this._wrapped}}(),function(){var t,e=this,i=this.Flotr;t={_:_,bean:bean,isIphone:/iphone/i.test(navigator.userAgent),isIE:-1!=navigator.appVersion.indexOf("MSIE")?parseFloat(navigator.appVersion.split("MSIE")[1]):!1,graphTypes:{},plugins:{},addType:function(e,i){t.graphTypes[e]=i,t.defaultOptions[e]=i.options||{},t.defaultOptions.defaultType=t.defaultOptions.defaultType||e
+},addPlugin:function(e,i){t.plugins[e]=i,t.defaultOptions[e]=i.options||{}},draw:function(e,i,n,a){return new(a=a||t.Graph)(e,i,n)},merge:function(e,i){var n,a,r=i||{};for(n in e)a=e[n],r[n]=a&&"object"==typeof a?a.constructor===Array?this._.clone(a):a.constructor===RegExp||this._.isElement(a)||a.jquery?a:t.merge(a,i?i[n]:void 0):a;return r},clone:function(e){return t.merge(e,{})},getTickSize:function(e,i,n,a){var r=(n-i)/e,o=t.getMagnitude(r),s=10,l=r/o;return 1.5>l?s=1:2.25>l?s=2:3>l?s=0===a?2:2.5:7.5>l&&(s=5),s*o},defaultTickFormatter:function(t){return t+""},defaultTrackFormatter:function(t){return"("+t.x+", "+t.y+")"},engineeringNotation:function(t,e,i){var n=["Y","Z","E","P","T","G","M","k",""],a=["y","z","a","f","p","n","µ","m",""],r=n.length;if(i=i||1e3,e=Math.pow(10,e||2),0===t)return 0;if(t>1)for(;r--&&t>=i;)t/=i;else for(n=a,r=n.length;r--&&1>t;)t*=i;return Math.round(t*e)/e+n[r]},getMagnitude:function(t){return Math.pow(10,Math.floor(Math.log(t)/Math.LN10))},toPixel:function(t){return Math.floor(t)+.5},toRad:function(t){return-t*(Math.PI/180)},floorInBase:function(t,e){return e*Math.floor(t/e)},drawText:function(e,i,n,a,r){return e.fillText?(r=this._.extend({size:t.defaultOptions.fontSize,color:"#000000",textAlign:"left",textBaseline:"bottom",weight:1,angle:0},r),e.save(),e.translate(n,a),e.rotate(r.angle),e.fillStyle=r.color,e.font=(r.weight>1?"bold ":"")+1.3*r.size+"px sans-serif",e.textAlign=r.textAlign,e.textBaseline=r.textBaseline,e.fillText(i,0,0),e.restore(),void 0):void e.drawText(i,n,a,r)},getBestTextAlign:function(e,i){return i=i||{textAlign:"center",textBaseline:"middle"},e+=t.getTextAngleFromAlign(i),Math.abs(Math.cos(e))>.01&&(i.textAlign=Math.cos(e)>0?"right":"left"),Math.abs(Math.sin(e))>.01&&(i.textBaseline=Math.sin(e)>0?"top":"bottom"),i},alignTable:{"right middle":0,"right top":Math.PI/4,"center top":Math.PI/2,"left top":3*(Math.PI/4),"left middle":Math.PI,"left bottom":-3*(Math.PI/4),"center bottom":-Math.PI/2,"right bottom":-Math.PI/4,"center middle":0},getTextAngleFromAlign:function(e){return t.alignTable[e.textAlign+" "+e.textBaseline]||0},noConflict:function(){return e.Flotr=i,this}},e.Flotr=t}(),Flotr.defaultOptions={colors:["#00A8F0","#C0D800","#CB4B4B","#4DA74D","#9440ED"],ieBackgroundColor:"#FFFFFF",title:null,subtitle:null,shadowSize:4,defaultType:null,HtmlText:!0,fontColor:"#545454",fontSize:7.5,resolution:1,parseFloat:!0,preventDefault:!0,xaxis:{ticks:null,minorTicks:null,showLabels:!0,showMinorLabels:!1,labelsAngle:0,title:null,titleAngle:0,noTicks:5,minorTickFreq:null,tickFormatter:Flotr.defaultTickFormatter,tickDecimals:null,min:null,max:null,autoscale:!1,autoscaleMargin:0,color:null,mode:"normal",timeFormat:null,timeMode:"UTC",timeUnit:"millisecond",scaling:"linear",base:Math.E,titleAlign:"center",margin:!0},x2axis:{},yaxis:{ticks:null,minorTicks:null,showLabels:!0,showMinorLabels:!1,labelsAngle:0,title:null,titleAngle:90,noTicks:5,minorTickFreq:null,tickFormatter:Flotr.defaultTickFormatter,tickDecimals:null,min:null,max:null,autoscale:!1,autoscaleMargin:0,color:null,scaling:"linear",base:Math.E,titleAlign:"center",margin:!0},y2axis:{titleAngle:270},grid:{color:"#545454",backgroundColor:null,backgroundImage:null,watermarkAlpha:.4,tickColor:"#DDDDDD",labelMargin:3,verticalLines:!0,minorVerticalLines:null,horizontalLines:!0,minorHorizontalLines:null,outlineWidth:1,outline:"nsew",circular:!1},mouse:{track:!1,trackAll:!1,position:"se",relative:!1,trackFormatter:Flotr.defaultTrackFormatter,margin:5,lineColor:"#FF3F19",trackDecimals:1,sensibility:2,trackY:!0,radius:3,fillColor:null,fillOpacity:.4}},function(){function t(){this.rgba=["r","g","b","a"];for(var t=4;-1<--t;)this[this.rgba[t]]=arguments[t]||(3==t?1:0);this.normalize()}var e=Flotr._,i={aqua:[0,255,255],azure:[240,255,255],beige:[245,245,220],black:[0,0,0],blue:[0,0,255],brown:[165,42,42],cyan:[0,255,255],darkblue:[0,0,139],darkcyan:[0,139,139],darkgrey:[169,169,169],darkgreen:[0,100,0],darkkhaki:[189,183,107],darkmagenta:[139,0,139],darkolivegreen:[85,107,47],darkorange:[255,140,0],darkorchid:[153,50,204],darkred:[139,0,0],darksalmon:[233,150,122],darkviolet:[148,0,211],fuchsia:[255,0,255],gold:[255,215,0],green:[0,128,0],indigo:[75,0,130],khaki:[240,230,140],lightblue:[173,216,230],lightcyan:[224,255,255],lightgreen:[144,238,144],lightgrey:[211,211,211],lightpink:[255,182,193],lightyellow:[255,255,224],lime:[0,255,0],magenta:[255,0,255],maroon:[128,0,0],navy:[0,0,128],olive:[128,128,0],orange:[255,165,0],pink:[255,192,203],purple:[128,0,128],violet:[128,0,128],red:[255,0,0],silver:[192,192,192],white:[255,255,255],yellow:[255,255,0]};t.prototype={scale:function(){for(var t=4;-1<--t;)e.isUndefined(arguments[t])||(this[this.rgba[t]]*=arguments[t]);return this.normalize()},alpha:function(t){return!e.isUndefined(t)&&!e.isNull(t)&&(this.a=t),this.normalize()},clone:function(){return new t(this.r,this.b,this.g,this.a)},limit:function(t,e,i){return Math.max(Math.min(t,i),e)},normalize:function(){var t=this.limit;return this.r=t(parseInt(this.r,10),0,255),this.g=t(parseInt(this.g,10),0,255),this.b=t(parseInt(this.b,10),0,255),this.a=t(this.a,0,1),this},distance:function(e){if(e){e=new t.parse(e);for(var i=0,n=3;-1<--n;)i+=Math.abs(this[this.rgba[n]]-e[this.rgba[n]]);return i}},toString:function(){return this.a>=1?"rgb("+[this.r,this.g,this.b].join(",")+")":"rgba("+[this.r,this.g,this.b,this.a].join(",")+")"},contrast:function(){var t=1-(.299*this.r+.587*this.g+.114*this.b)/255;return.5>t?"#000000":"#ffffff"}},e.extend(t,{parse:function(e){if(e instanceof t)return e;var n;if(n=/#([a-fA-F0-9]{2})([a-fA-F0-9]{2})([a-fA-F0-9]{2})/.exec(e))return new t(parseInt(n[1],16),parseInt(n[2],16),parseInt(n[3],16));if(n=/rgb\(\s*([0-9]{1,3})\s*,\s*([0-9]{1,3})\s*,\s*([0-9]{1,3})\s*\)/.exec(e))return new t(parseInt(n[1],10),parseInt(n[2],10),parseInt(n[3],10));if(n=/#([a-fA-F0-9])([a-fA-F0-9])([a-fA-F0-9])/.exec(e))return new t(parseInt(n[1]+n[1],16),parseInt(n[2]+n[2],16),parseInt(n[3]+n[3],16));if(n=/rgba\(\s*([0-9]{1,3})\s*,\s*([0-9]{1,3})\s*,\s*([0-9]{1,3})\s*,\s*([0-9]+(?:\.[0-9]+)?)\s*\)/.exec(e))return new t(parseInt(n[1],10),parseInt(n[2],10),parseInt(n[3],10),parseFloat(n[4]));if(n=/rgb\(\s*([0-9]+(?:\.[0-9]+)?)\%\s*,\s*([0-9]+(?:\.[0-9]+)?)\%\s*,\s*([0-9]+(?:\.[0-9]+)?)\%\s*\)/.exec(e))return new t(2.55*parseFloat(n[1]),2.55*parseFloat(n[2]),2.55*parseFloat(n[3]));if(n=/rgba\(\s*([0-9]+(?:\.[0-9]+)?)\%\s*,\s*([0-9]+(?:\.[0-9]+)?)\%\s*,\s*([0-9]+(?:\.[0-9]+)?)\%\s*,\s*([0-9]+(?:\.[0-9]+)?)\s*\)/.exec(e))return new t(2.55*parseFloat(n[1]),2.55*parseFloat(n[2]),2.55*parseFloat(n[3]),parseFloat(n[4]));var a=(e+"").replace(/^\s*([\S\s]*?)\s*$/,"$1").toLowerCase();return"transparent"==a?new t(255,255,255,0):(n=i[a])?new t(n[0],n[1],n[2]):new t(0,0,0,0)},processColor:function(i,n){var a=n.opacity;if(!i)return"rgba(0, 0, 0, 0)";if(i instanceof t)return i.alpha(a).toString();if(e.isString(i))return t.parse(i).alpha(a).toString();var r=i.colors?i:{colors:i};if(!n.ctx)return e.isArray(r.colors)?t.parse(e.isArray(r.colors[0])?r.colors[0][1]:r.colors[0]).alpha(a).toString():"rgba(0, 0, 0, 0)";r=e.extend({start:"top",end:"bottom"},r),/top/i.test(r.start)&&(n.x1=0),/left/i.test(r.start)&&(n.y1=0),/bottom/i.test(r.end)&&(n.x2=0),/right/i.test(r.end)&&(n.y2=0);var o,s,l,h=n.ctx.createLinearGradient(n.x1,n.y1,n.x2,n.y2);for(o=0;o<r.colors.length;o++)s=r.colors[o],e.isArray(s)?(l=s[0],s=s[1]):l=o/(r.colors.length-1),h.addColorStop(l,t.parse(s).alpha(a));return h}}),Flotr.Color=t}(),Flotr.Date={set:function(t,e,i,n){i=i||"UTC",e="set"+("UTC"===i?"UTC":"")+e,t[e](n)},get:function(t,e,i){return i=i||"UTC",e="get"+("UTC"===i?"UTC":"")+e,t[e]()},format:function(t,e,i){function n(t){return t+="",1==t.length?"0"+t:t}if(t){for(var a,r=this.get,o={h:r(t,"Hours",i).toString(),H:n(r(t,"Hours",i)),M:n(r(t,"Minutes",i)),S:n(r(t,"Seconds",i)),s:r(t,"Milliseconds",i),d:r(t,"Date",i).toString(),m:(r(t,"Month",i)+1).toString(),y:r(t,"FullYear",i).toString(),b:Flotr.Date.monthNames[r(t,"Month",i)]},s=[],l=!1,h=0;h<e.length;++h)a=e.charAt(h),l?(s.push(o[a]||a),l=!1):"%"==a?l=!0:s.push(a);return s.join("")}},getFormat:function(t,e){var i=Flotr.Date.timeUnits;return t<i.second?"%h:%M:%S.%s":t<i.minute?"%h:%M:%S":t<i.day?e<2*i.day?"%h:%M":"%b %d %h:%M":t<i.month?"%b %d":t<i.year?e<i.year?"%b":"%b %y":"%y"},formatter:function(t,e){var i=e.options,n=Flotr.Date.timeUnits[i.timeUnit],a=new Date(t*n);if(e.options.timeFormat)return Flotr.Date.format(a,i.timeFormat,i.timeMode);var r=(e.max-e.min)*n,o=e.tickSize*Flotr.Date.timeUnits[e.tickUnit];return Flotr.Date.format(a,Flotr.Date.getFormat(o,r),i.timeMode)},generator:function(t){function e(t){r(v,t,c,Flotr.floorInBase(o(v,t,c),m))}var i,n,a,r=this.set,o=this.get,s=this.timeUnits,l=this.spec,h=t.options,c=h.timeMode,u=s[h.timeUnit],d=t.min*u,f=t.max*u,p=(f-d)/h.noTicks,g=[],m=t.tickSize;for(n=h.tickFormatter===Flotr.defaultTickFormatter?this.formatter:h.tickFormatter,a=0;a<l.length-1;++a){var v=l[a][0]*s[l[a][1]];if(p<(v+l[a+1][0]*s[l[a+1][1]])/2&&v>=m)break}m=l[a][0],i=l[a][1],"year"==i&&(m=Flotr.getTickSize(h.noTicks*s.year,d,f,0),.5==m&&(i="month",m=6)),t.tickUnit=i,t.tickSize=m;var x=m*s[i];switch(v=new Date(d),i){case"millisecond":e("Milliseconds");break;case"second":e("Seconds");break;case"minute":e("Minutes");break;case"hour":e("Hours");break;case"month":e("Month");break;case"year":e("FullYear")}x>=s.second&&r(v,"Milliseconds",c,0),x>=s.minute&&r(v,"Seconds",c,0),x>=s.hour&&r(v,"Minutes",c,0),x>=s.day&&r(v,"Hours",c,0),x>=4*s.day&&r(v,"Date",c,1),x>=s.year&&r(v,"Month",c,0);var y,b=0,w=0/0;do if(y=w,w=v.getTime(),g.push({v:w/u,label:n(w/u,t)}),"month"==i)if(1>m){r(v,"Date",c,1);var M=v.getTime();r(v,"Month",c,o(v,"Month",c)+1);var k=v.getTime();v.setTime(w+b*s.hour+(k-M)*m),b=o(v,"Hours",c),r(v,"Hours",c,0)}else r(v,"Month",c,o(v,"Month",c)+m);else"year"==i?r(v,"FullYear",c,o(v,"FullYear",c)+m):v.setTime(w+x);while(f>w&&w!=y);return g},timeUnits:{millisecond:1,second:1e3,minute:6e4,hour:36e5,day:864e5,month:2592e6,year:31556952e3},spec:[[1,"millisecond"],[20,"millisecond"],[50,"millisecond"],[100,"millisecond"],[200,"millisecond"],[500,"millisecond"],[1,"second"],[2,"second"],[5,"second"],[10,"second"],[30,"second"],[1,"minute"],[2,"minute"],[5,"minute"],[10,"minute"],[30,"minute"],[1,"hour"],[2,"hour"],[4,"hour"],[8,"hour"],[12,"hour"],[1,"day"],[2,"day"],[3,"day"],[.25,"month"],[.5,"month"],[1,"month"],[2,"month"],[3,"month"],[6,"month"],[1,"year"]],monthNames:["Jan","Feb","Mar","Apr","May","Jun","Jul","Aug","Sep","Oct","Nov","Dec"]},function(){function t(t){return t&&t.jquery?t[0]:t}var e=Flotr._;Flotr.DOM={addClass:function(i,n){i=t(i);var a=i.className?i.className:"";e.include(a.split(/\s+/g),n)||(i.className=(a?a+" ":"")+n)},create:function(t){return document.createElement(t)},node:function(t){var e,i=Flotr.DOM.create("div");return i.innerHTML=t,e=i.children[0],i.innerHTML="",e},empty:function(e){e=t(e),e.innerHTML=""},remove:function(e){e=t(e),e.parentNode.removeChild(e)},hide:function(e){e=t(e),Flotr.DOM.setStyles(e,{display:"none"})},insert:function(i,n){i=t(i),e.isString(n)?i.innerHTML+=n:e.isElement(n)&&i.appendChild(n)},opacity:function(e,i){e=t(e),e.style.opacity=i},position:function(e,i){return e=t(e),e.offsetParent?(i=this.position(e.offsetParent),i.left+=e.offsetLeft,i.top+=e.offsetTop,i):{left:e.offsetLeft||0,top:e.offsetTop||0}},removeClass:function(i,n){var a=i.className?i.className:"";i=t(i),i.className=e.filter(a.split(/\s+/g),function(t){return t!=n?!0:void 0}).join(" ")},setStyles:function(i,n){i=t(i),e.each(n,function(t,e){i.style[e]=t})},show:function(e){e=t(e),Flotr.DOM.setStyles(e,{display:""})},size:function(e){return e=t(e),{height:e.offsetHeight,width:e.offsetWidth}}}}(),function(){var t=Flotr,e=t.bean;t.EventAdapter={observe:function(t,i,n){return e.add(t,i,n),this},fire:function(t,i,n){return e.fire(t,i,n),"undefined"!=typeof Prototype&&Event.fire(t,i,n),this},stopObserving:function(t,i,n){return e.remove(t,i,n),this},eventPointer:function(e){if(!t._.isUndefined(e.touches)&&e.touches.length>0)return{x:e.touches[0].pageX,y:e.touches[0].pageY};if(!t._.isUndefined(e.changedTouches)&&e.changedTouches.length>0)return{x:e.changedTouches[0].pageX,y:e.changedTouches[0].pageY};if(e.pageX||e.pageY)return{x:e.pageX,y:e.pageY};if(e.clientX||e.clientY){var i=document,n=i.body,a=i.documentElement;return{x:e.clientX+n.scrollLeft+a.scrollLeft,y:e.clientY+n.scrollTop+a.scrollTop}}}}}(),function(){var t=Flotr,e=t.DOM,i=t._,n=function(t){this.o=t};n.prototype={dimensions:function(t,e,i,n){return t?this.o.html?this.html(t,this.o.element,i,n):this.canvas(t,e):{width:0,height:0}},canvas:function(e,i){if(this.o.textEnabled){i=i||{};var n,a=this.measureText(e,i),r=a.width,o=i.size||t.defaultOptions.fontSize,s=i.angle||0,l=Math.cos(s),h=Math.sin(s),c=2,u=6;return n={width:Math.abs(l*r)+Math.abs(h*o)+c,height:Math.abs(h*r)+Math.abs(l*o)+u}}},html:function(t,i,n,a){var r=e.create("div");return e.setStyles(r,{position:"absolute",top:"-10000px"}),e.insert(r,'<div style="'+n+'" class="'+a+' flotr-dummy-div">'+t+"</div>"),e.insert(this.o.element,r),e.size(r)},measureText:function(e,n){var a,r=this.o.ctx;return!r.fillText||t.isIphone&&r.measure?{width:r.measure(e,n)}:(n=i.extend({size:t.defaultOptions.fontSize,weight:1,angle:0},n),r.save(),r.font=(n.weight>1?"bold ":"")+1.3*n.size+"px sans-serif",a=r.measureText(e),r.restore(),a)}},Flotr.Text=n}(),function(){function t(){return i.observe.apply(this,arguments),this._handles.push(arguments),this}var e=Flotr.DOM,i=Flotr.EventAdapter,n=Flotr._,a=Flotr;Graph=function(t,e,r){this._setEl(t),this._initMembers(),this._initPlugins(),i.fire(this.el,"flotr:beforeinit",[this]),this.data=e,this.series=a.Series.getSeries(e),this._initOptions(r),this._initGraphTypes(),this._initCanvas(),this._text=new a.Text({element:this.el,ctx:this.ctx,html:this.options.HtmlText,textEnabled:this.textEnabled}),i.fire(this.el,"flotr:afterconstruct",[this]),this._initEvents(),this.findDataRanges(),this.calculateSpacing(),this.draw(n.bind(function(){i.fire(this.el,"flotr:afterinit",[this])},this))},Graph.prototype={destroy:function(){i.fire(this.el,"flotr:destroy"),n.each(this._handles,function(t){i.stopObserving.apply(this,t)}),this._handles=[],this.el.graph=null},observe:t,_observe:t,processColor:function(t,e){var i={x1:0,y1:0,x2:this.plotWidth,y2:this.plotHeight,opacity:1,ctx:this.ctx};return n.extend(i,e),a.Color.processColor(t,i)},findDataRanges:function(){var t,e,i,r=this.axes;n.each(this.series,function(n){i=n.getRange(),i&&(t=n.xaxis,e=n.yaxis,t.datamin=Math.min(i.xmin,t.datamin),t.datamax=Math.max(i.xmax,t.datamax),e.datamin=Math.min(i.ymin,e.datamin),e.datamax=Math.max(i.ymax,e.datamax),t.used=t.used||i.xused,e.used=e.used||i.yused)},this),!r.x.used&&!r.x2.used&&(r.x.used=!0),!r.y.used&&!r.y2.used&&(r.y.used=!0),n.each(r,function(t){t.calculateRange()});var o=n.keys(a.graphTypes),s=!1;n.each(this.series,function(t){t.hide||(n.each(o,function(e){t[e]&&t[e].show&&(this.extendRange(e,t),s=!0)},this),s||this.extendRange(this.options.defaultType,t))},this)},extendRange:function(t,e){this[t].extendRange&&this[t].extendRange(e,e.data,e[t],this[t]),this[t].extendYRange&&this[t].extendYRange(e.yaxis,e.data,e[t],this[t]),this[t].extendXRange&&this[t].extendXRange(e.xaxis,e.data,e[t],this[t])},calculateSpacing:function(){var t,e,i=this.axes,a=this.options,r=this.series,o=a.grid.labelMargin,s=this._text,l=i.x,h=i.x2,c=i.y,u=i.y2,d=a.grid.outlineWidth;for(n.each(i,function(t){t.calculateTicks(),t.calculateTextDimensions(s,a)}),e=s.dimensions(a.title,{size:1.5*a.fontSize},"font-size:1em;font-weight:bold;","flotr-title"),this.titleHeight=e.height,e=s.dimensions(a.subtitle,{size:a.fontSize},"font-size:smaller;","flotr-subtitle"),this.subtitleHeight=e.height,t=0;t<a.length;++t)r[t].points.show&&(d=Math.max(d,r[t].points.radius+r[t].points.lineWidth/2));var f=this.plotOffset;l.options.margin===!1?(f.bottom=0,f.top=0):(f.bottom+=(a.grid.circular?0:l.used&&l.options.showLabels?l.maxLabel.height+o:0)+(l.used&&l.options.title?l.titleSize.height+o:0)+d,f.top+=(a.grid.circular?0:h.used&&h.options.showLabels?h.maxLabel.height+o:0)+(h.used&&h.options.title?h.titleSize.height+o:0)+this.subtitleHeight+this.titleHeight+d),c.options.margin===!1?(f.left=0,f.right=0):(f.left+=(a.grid.circular?0:c.used&&c.options.showLabels?c.maxLabel.width+o:0)+(c.used&&c.options.title?c.titleSize.width+o:0)+d,f.right+=(a.grid.circular?0:u.used&&u.options.showLabels?u.maxLabel.width+o:0)+(u.used&&u.options.title?u.titleSize.width+o:0)+d),f.top=Math.floor(f.top),this.plotWidth=this.canvasWidth-f.left-f.right,this.plotHeight=this.canvasHeight-f.bottom-f.top,l.length=h.length=this.plotWidth,c.length=u.length=this.plotHeight,c.offset=u.offset=this.plotHeight,l.setScale(),h.setScale(),c.setScale(),u.setScale()},draw:function(t){var e,n=this.ctx;if(i.fire(this.el,"flotr:beforedraw",[this.series,this]),this.series.length){for(n.save(),n.translate(this.plotOffset.left,this.plotOffset.top),e=0;e<this.series.length;e++)this.series[e].hide||this.drawSeries(this.series[e]);n.restore(),this.clip()}i.fire(this.el,"flotr:afterdraw",[this.series,this]),t&&t()},drawSeries:function(t){function e(t,e){var i=this.getOptions(t,e);this[e].draw(i)}var i=!1;t=t||this.series,n.each(a.graphTypes,function(n,a){t[a]&&t[a].show&&this[a]&&(i=!0,e.call(this,t,a))},this),i||e.call(this,t,this.options.defaultType)},getOptions:function(t,e){var i=t[e],n=(this[e],t.xaxis),r=t.yaxis,o={context:this.ctx,width:this.plotWidth,height:this.plotHeight,fontSize:this.options.fontSize,fontColor:this.options.fontColor,textEnabled:this.textEnabled,htmlText:this.options.HtmlText,text:this._text,element:this.el,data:t.data,color:t.color,shadowSize:t.shadowSize,xScale:n.d2p,yScale:r.d2p,xInverse:n.p2d,yInverse:r.p2d};return o=a.merge(i,o),o.fillStyle=this.processColor(i.fillColor||t.color,{opacity:i.fillOpacity}),o},getEventPosition:function(t){var n,a,r,o=document,s=o.body,l=o.documentElement,h=this.axes,c=this.plotOffset,u=this.lastMousePos,d=i.eventPointer(t),f=d.x-u.pageX,p=d.y-u.pageY;return"ontouchstart"in this.el?(n=e.position(this.overlay),a=d.x-n.left-c.left,r=d.y-n.top-c.top):(n=this.overlay.getBoundingClientRect(),a=t.clientX-n.left-c.left-s.scrollLeft-l.scrollLeft,r=t.clientY-n.top-c.top-s.scrollTop-l.scrollTop),{x:h.x.p2d(a),x2:h.x2.p2d(a),y:h.y.p2d(r),y2:h.y2.p2d(r),relX:a,relY:r,dX:f,dY:p,absX:d.x,absY:d.y,pageX:d.x,pageY:d.y}},clickHandler:function(t){return this.ignoreClick?(this.ignoreClick=!1,this.ignoreClick):void i.fire(this.el,"flotr:click",[this.getEventPosition(t),this])},mouseMoveHandler:function(t){if(!this.mouseDownMoveHandler){var e=this.getEventPosition(t);i.fire(this.el,"flotr:mousemove",[t,e,this]),this.lastMousePos=e}},mouseDownHandler:function(t){this.mouseUpHandler||(this.mouseUpHandler=n.bind(function(t){i.stopObserving(document,"mouseup",this.mouseUpHandler),i.stopObserving(document,"mousemove",this.mouseDownMoveHandler),this.mouseDownMoveHandler=null,this.mouseUpHandler=null,i.fire(this.el,"flotr:mouseup",[t,this])},this),this.mouseDownMoveHandler=n.bind(function(e){var n=this.getEventPosition(e);i.fire(this.el,"flotr:mousemove",[t,n,this]),this.lastMousePos=n},this),i.observe(document,"mouseup",this.mouseUpHandler),i.observe(document,"mousemove",this.mouseDownMoveHandler),i.fire(this.el,"flotr:mousedown",[t,this]),this.ignoreClick=!1)},drawTooltip:function(t,i,n,a){var r=this.getMouseTrack(),o="opacity:0.7;background-color:#000;color:#fff;display:none;position:absolute;padding:2px 8px;-moz-border-radius:4px;border-radius:4px;white-space:nowrap;",s=a.position,l=a.margin,h=this.plotOffset;null!==i&&null!==n?(a.relative?("n"==s.charAt(0)?o+="bottom:"+(l-h.top-n+this.canvasHeight)+"px;top:auto;":"s"==s.charAt(0)&&(o+="top:"+(l+h.top+n)+"px;bottom:auto;"),"e"==s.charAt(1)?o+="left:"+(l+h.left+i)+"px;right:auto;":"w"==s.charAt(1)&&(o+="right:"+(l-h.left-i+this.canvasWidth)+"px;left:auto;")):("n"==s.charAt(0)?o+="top:"+(l+h.top)+"px;bottom:auto;":"s"==s.charAt(0)&&(o+="bottom:"+(l+h.bottom)+"px;top:auto;"),"e"==s.charAt(1)?o+="right:"+(l+h.right)+"px;left:auto;":"w"==s.charAt(1)&&(o+="left:"+(l+h.left)+"px;right:auto;")),r.style.cssText=o,e.empty(r),e.insert(r,t),e.show(r)):e.hide(r)},clip:function(t){var e=this.plotOffset,i=this.canvasWidth,n=this.canvasHeight;t=t||this.ctx,a.isIE&&a.isIE<9?(t.save(),t.fillStyle=this.processColor(this.options.ieBackgroundColor),t.fillRect(0,0,i,e.top),t.fillRect(0,0,e.left,n),t.fillRect(0,n-e.bottom,i,e.bottom),t.fillRect(i-e.right,0,e.right,n),t.restore()):(t.clearRect(0,0,i,e.top),t.clearRect(0,0,e.left,n),t.clearRect(0,n-e.bottom,i,e.bottom),t.clearRect(i-e.right,0,e.right,n))},_initMembers:function(){this._handles=[],this.lastMousePos={pageX:null,pageY:null},this.plotOffset={left:0,right:0,top:0,bottom:0},this.ignoreClick=!0,this.prevHit=null},_initGraphTypes:function(){n.each(a.graphTypes,function(t,e){this[e]=a.clone(t)},this)},_initEvents:function(){var t,e,a,r=this.el;"ontouchstart"in r?(t=n.bind(function(n){a=!0,i.stopObserving(document,"touchend",t),i.fire(r,"flotr:mouseup",[event,this]),this.multitouches=null,e||this.clickHandler(n)},this),this.observe(this.overlay,"touchstart",n.bind(function(n){e=!1,a=!1,this.ignoreClick=!1,n.touches&&n.touches.length>1&&(this.multitouches=n.touches),i.fire(r,"flotr:mousedown",[event,this]),this.observe(document,"touchend",t)},this)),this.observe(this.overlay,"touchmove",n.bind(function(t){var n=this.getEventPosition(t);this.options.preventDefault&&t.preventDefault(),e=!0,this.multitouches||t.touches&&t.touches.length>1?this.multitouches=t.touches:a||i.fire(r,"flotr:mousemove",[event,n,this]),this.lastMousePos=n},this))):this.observe(this.overlay,"mousedown",n.bind(this.mouseDownHandler,this)).observe(r,"mousemove",n.bind(this.mouseMoveHandler,this)).observe(this.overlay,"click",n.bind(this.clickHandler,this)).observe(r,"mouseout",function(){i.fire(r,"flotr:mouseout")})},_initCanvas:function(){function t(t,i){return t||(t=e.create("canvas"),"undefined"!=typeof FlashCanvas&&"function"==typeof t.getContext&&FlashCanvas.initElement(t),t.className="flotr-"+i,t.style.cssText="position:absolute;left:0px;top:0px;",e.insert(s,t)),n.each(o,function(n,a){e.show(t),("canvas"!=i||t.getAttribute(a)!==n)&&(t.setAttribute(a,n*l.resolution),t.style[a]=n+"px")}),t.context_=null,t}function i(t){window.G_vmlCanvasManager&&window.G_vmlCanvasManager.initElement(t);var e=t.getContext("2d");return window.G_vmlCanvasManager||e.scale(l.resolution,l.resolution),e}var a,r,o,s=this.el,l=this.options,h=s.children,c=[];for(r=h.length;r--;)a=h[r],this.canvas||"flotr-canvas"!==a.className?this.overlay||"flotr-overlay"!==a.className?c.push(a):this.overlay=a:this.canvas=a;for(r=c.length;r--;)s.removeChild(c[r]);if(e.setStyles(s,{position:"relative"}),o={},o.width=s.clientWidth,o.height=s.clientHeight,o.width<=0||o.height<=0||l.resolution<=0)throw"Invalid dimensions for plot, width = "+o.width+", height = "+o.height+", resolution = "+l.resolution;this.canvas=t(this.canvas,"canvas"),this.overlay=t(this.overlay,"overlay"),this.ctx=i(this.canvas),this.ctx.clearRect(0,0,this.canvas.width,this.canvas.height),this.octx=i(this.overlay),this.octx.clearRect(0,0,this.overlay.width,this.overlay.height),this.canvasHeight=o.height,this.canvasWidth=o.width,this.textEnabled=!!this.ctx.drawText||!!this.ctx.fillText},_initPlugins:function(){n.each(a.plugins,function(t,e){n.each(t.callbacks,function(t,e){this.observe(this.el,e,n.bind(t,this))},this),this[e]=a.clone(t),n.each(this[e],function(t,i){n.isFunction(t)&&(this[e][i]=n.bind(t,this))},this)},this)},_initOptions:function(t){var e=a.clone(a.defaultOptions);e.x2axis=n.extend(n.clone(e.xaxis),e.x2axis),e.y2axis=n.extend(n.clone(e.yaxis),e.y2axis),this.options=a.merge(t||{},e),null===this.options.grid.minorVerticalLines&&"logarithmic"===this.options.xaxis.scaling&&(this.options.grid.minorVerticalLines=!0),null===this.options.grid.minorHorizontalLines&&"logarithmic"===this.options.yaxis.scaling&&(this.options.grid.minorHorizontalLines=!0),i.fire(this.el,"flotr:afterinitoptions",[this]),this.axes=a.Axis.getAxes(this.options);var r,o,s,l,h=[],c=[],u=this.series.length,d=this.series.length,f=this.options.colors,p=[],g=0;for(o=d-1;o>-1;--o)r=this.series[o].color,r&&(--d,n.isNumber(r)?h.push(r):p.push(a.Color.parse(r)));for(o=h.length-1;o>-1;--o)d=Math.max(d,h[o]+1);for(o=0;c.length<d;){r=f.length==o?new a.Color(100,100,100):a.Color.parse(f[o]);var m=g%2==1?-1:1,v=1+m*Math.ceil(g/2)*.2;r.scale(v,v,v),c.push(r),++o>=f.length&&(o=0,++g)}for(o=0,s=0;u>o;++o){l=this.series[o],l.color?n.isNumber(l.color)&&(l.color=c[l.color].toString()):l.color=c[s++].toString(),l.xaxis||(l.xaxis=this.axes.x),1==l.xaxis?l.xaxis=this.axes.x:2==l.xaxis&&(l.xaxis=this.axes.x2),l.yaxis||(l.yaxis=this.axes.y),1==l.yaxis?l.yaxis=this.axes.y:2==l.yaxis&&(l.yaxis=this.axes.y2);for(var x in a.graphTypes)l[x]=n.extend(n.clone(this.options[x]),l[x]);l.mouse=n.extend(n.clone(this.options.mouse),l.mouse),n.isUndefined(l.shadowSize)&&(l.shadowSize=this.options.shadowSize)}},_setEl:function(t){if(!t)throw"The target container doesn't exist";if(t.graph instanceof Graph)t.graph.destroy();else if(!t.clientWidth)throw"The target container must be visible";t.graph=this,this.el=t}},Flotr.Graph=Graph}(),function(){function t(t){this.orientation=1,this.offset=0,this.datamin=Number.MAX_VALUE,this.datamax=-Number.MAX_VALUE,a.extend(this,t)}function e(t,e){return t=Math.log(Math.max(t,Number.MIN_VALUE)),e!==Math.E&&(t/=Math.log(e)),t}function n(t,e){return e===Math.E?Math.exp(t):Math.pow(e,t)}var a=Flotr._,r="logarithmic";t.prototype={setScale:function(){var t,i=this.length,a=this.max,o=this.min,s=this.offset,l=this.orientation,h=this.options,c=h.scaling===r;t=c?i/(e(a,h.base)-e(o,h.base)):i/(a-o),this.scale=t,c?(this.d2p=function(i){return s+l*(e(i,h.base)-e(o,h.base))*t},this.p2d=function(i){return n((s+l*i)/t+e(o,h.base),h.base)}):(this.d2p=function(e){return s+l*(e-o)*t},this.p2d=function(e){return(s+l*e)/t+o})},calculateTicks:function(){var t=this.options;this.ticks=[],this.minorTicks=[],t.ticks?(this._cleanUserTicks(t.ticks,this.ticks),this._cleanUserTicks(t.minorTicks||[],this.minorTicks)):"time"==t.mode?this._calculateTimeTicks():"logarithmic"===t.scaling?this._calculateLogTicks():this._calculateTicks(),a.each(this.ticks,function(t){t.label+=""}),a.each(this.minorTicks,function(t){t.label+=""})},calculateRange:function(){if(this.used){var t=this,e=t.options,i=null!==e.min?e.min:t.datamin,n=null!==e.max?e.max:t.datamax,a=e.autoscaleMargin;if("logarithmic"==e.scaling&&(0>=i&&(i=t.datamin),0>=n&&(n=i)),n==i){var r=n?.01:1;null===e.min&&(i-=r),null===e.max&&(n+=r)}if("logarithmic"===e.scaling){0>i&&(i=n/e.base);var o=Math.log(n);e.base!=Math.E&&(o/=Math.log(e.base)),o=Math.ceil(o);var s=Math.log(i);e.base!=Math.E&&(s/=Math.log(e.base)),s=Math.ceil(s),t.tickSize=Flotr.getTickSize(e.noTicks,s,o,null===e.tickDecimals?0:e.tickDecimals),null===e.minorTickFreq&&(e.minorTickFreq=o-s>10?0:o-s>5?2:5)}else t.tickSize=Flotr.getTickSize(e.noTicks,i,n,e.tickDecimals);t.min=i,t.max=n,null===e.min&&e.autoscale&&(t.min-=t.tickSize*a,t.min<0&&t.datamin>=0&&(t.min=0),t.min=t.tickSize*Math.floor(t.min/t.tickSize)),null===e.max&&e.autoscale&&(t.max+=t.tickSize*a,t.max>0&&t.datamax<=0&&t.datamax!=t.datamin&&(t.max=0),t.max=t.tickSize*Math.ceil(t.max/t.tickSize)),t.min==t.max&&(t.max=t.min+1)}},calculateTextDimensions:function(t,e){var i,n,a="";if(this.options.showLabels)for(n=0;n<this.ticks.length;++n)i=this.ticks[n].label.length,i>a.length&&(a=this.ticks[n].label);this.maxLabel=t.dimensions(a,{size:e.fontSize,angle:Flotr.toRad(this.options.labelsAngle)},"font-size:smaller;","flotr-grid-label"),this.titleSize=t.dimensions(this.options.title,{size:1.2*e.fontSize,angle:Flotr.toRad(this.options.titleAngle)},"font-weight:bold;","flotr-axis-title")},_cleanUserTicks:function(t,e){var i,n,r,o,s=this,l=this.options;for(a.isFunction(t)&&(t=t({min:s.min,max:s.max})),n=0;n<t.length;++n)o=t[n],"object"==typeof o?(i=o[0],r=o.length>1?o[1]:l.tickFormatter(i,{min:s.min,max:s.max})):(i=o,r=l.tickFormatter(i,{min:this.min,max:this.max})),e[n]={v:i,label:r}},_calculateTimeTicks:function(){this.ticks=Flotr.Date.generator(this)},_calculateLogTicks:function(){var t,e,n=this,a=n.options,r=Math.log(n.max);a.base!=Math.E&&(r/=Math.log(a.base)),r=Math.ceil(r);var o=Math.log(n.min);for(a.base!=Math.E&&(o/=Math.log(a.base)),o=Math.ceil(o),i=o;i<r;i+=n.tickSize){e=a.base==Math.E?Math.exp(i):Math.pow(a.base,i);var s=e*(a.base==Math.E?Math.exp(n.tickSize):Math.pow(a.base,n.tickSize)),l=(s-e)/a.minorTickFreq;for(n.ticks.push({v:e,label:a.tickFormatter(e,{min:n.min,max:n.max})}),t=e+l;s>t;t+=l)n.minorTicks.push({v:t,label:a.tickFormatter(t,{min:n.min,max:n.max})})}e=a.base==Math.E?Math.exp(i):Math.pow(a.base,i),n.ticks.push({v:e,label:a.tickFormatter(e,{min:n.min,max:n.max})})},_calculateTicks:function(){var t,e,i,n,a,r,o=this,s=o.options,l=o.tickSize,h=o.min,c=o.max,u=l*Math.ceil(h/l);for(s.minorTickFreq&&(e=l/s.minorTickFreq),a=0;(i=n=u+a*l)<=c;++a)if(t=s.tickDecimals,null===t&&(t=1-Math.floor(Math.log(l)/Math.LN10)),0>t&&(t=0),i=i.toFixed(t),o.ticks.push({v:i,label:s.tickFormatter(i,{min:o.min,max:o.max})}),s.minorTickFreq)for(r=0;r<s.minorTickFreq&&c>a*l+r*e;++r)i=n+r*e,o.minorTicks.push({v:i,label:s.tickFormatter(i,{min:o.min,max:o.max})})}},a.extend(t,{getAxes:function(e){return{x:new t({options:e.xaxis,n:1,length:this.plotWidth}),x2:new t({options:e.x2axis,n:2,length:this.plotWidth}),y:new t({options:e.yaxis,n:1,length:this.plotHeight,offset:this.plotHeight,orientation:-1}),y2:new t({options:e.y2axis,n:2,length:this.plotHeight,offset:this.plotHeight,orientation:-1})}}}),Flotr.Axis=t}(),function(){function t(t){e.extend(this,t)}var e=Flotr._;t.prototype={getRange:function(){var t,e,i,n=this.data,a=n.length,r=Number.MAX_VALUE,o=Number.MAX_VALUE,s=-Number.MAX_VALUE,l=-Number.MAX_VALUE,h=!1,c=!1;if(0>a||this.hide)return!1;for(i=0;a>i;i++)t=n[i][0],e=n[i][1],null!==t&&(r>t&&(r=t,h=!0),t>s&&(s=t,h=!0)),null!==e&&(o>e&&(o=e,c=!0),e>l&&(l=e,c=!0));return{xmin:r,xmax:s,ymin:o,ymax:l,xused:h,yused:c}}},e.extend(t,{getSeries:function(i){return e.map(i,function(i){var n;return i.data?(n=new t,e.extend(n,i)):n=new t({data:i}),n})}}),Flotr.Series=t}(),Flotr.addType("lines",{options:{show:!1,lineWidth:2,fill:!1,fillBorder:!1,fillColor:null,fillOpacity:.4,steps:!1,stacked:!1},stack:{values:[]},draw:function(t){var e,i=t.context,n=t.lineWidth,a=t.shadowSize;i.save(),i.lineJoin="round",a&&(i.lineWidth=a/2,e=n/2+i.lineWidth/2,i.strokeStyle="rgba(0,0,0,0.1)",this.plot(t,e+a/2,!1),i.strokeStyle="rgba(0,0,0,0.2)",this.plot(t,e,!1)),i.lineWidth=n,i.strokeStyle=t.color,this.plot(t,0,!0),i.restore()},plot:function(t,e,i){function n(){!e&&t.fill&&M&&(a=p(M[0]),u.fillStyle=t.fillStyle,u.lineTo(r,w),u.lineTo(a,w),u.lineTo(a,g(M[1])),u.fill(),t.fillBorder&&u.stroke())}var a,r,o,s,l,h,c,u=t.context,d=t.width,f=t.height,p=t.xScale,g=t.yScale,m=t.data,v=t.stacked?this.stack:!1,x=m.length-1,y=null,b=null,w=g(0),M=null;if(!(1>x)){for(u.beginPath(),c=0;x>c;++c)null!==m[c][1]&&null!==m[c+1][1]?(a=p(m[c][0]),r=p(m[c+1][0]),null===M&&(M=m[c]),v?(l=v.values[m[c][0]]||0,h=v.values[m[c+1][0]]||v.values[m[c][0]]||0,o=g(m[c][1]+l),s=g(m[c+1][1]+h),i&&(v.values[m[c][0]]=m[c][1]+l,c==x-1&&(v.values[m[c+1][0]]=m[c+1][1]+h))):(o=g(m[c][1]),s=g(m[c+1][1])),o>f&&s>f||0>o&&0>s||0>a&&0>r||a>d&&r>d||((y!=a||b!=o+e)&&u.moveTo(a,o+e),y=r,b=s+e,t.steps?(u.lineTo(y+e/2,o+e),u.lineTo(y+e/2,b)):u.lineTo(y,b))):t.fill&&c>0&&m[c][1]&&(u.stroke(),n(),M=null,u.closePath(),u.beginPath());(!t.fill||t.fill&&!t.fillBorder)&&u.stroke(),n(),u.closePath()}},extendYRange:function(t,e,i,n){var a=t.options;if(i.stacked&&(!a.max&&0!==a.max||!a.min&&0!==a.min)){var r,o,s=t.max,l=t.min,h=n.positiveSums||{},c=n.negativeSums||{};for(o=0;o<e.length;o++)r=e[o][0]+"",e[o][1]>0?(h[r]=(h[r]||0)+e[o][1],s=Math.max(s,h[r])):(c[r]=(c[r]||0)+e[o][1],l=Math.min(l,c[r]));n.negativeSums=c,n.positiveSums=h,t.max=s,t.min=l}i.steps&&(this.hit=function(t){var e,i=t.data,n=t.args,a=t.yScale,r=n[0],o=i.length,s=n[1],l=t.xInverse(r.relX),h=r.relY;
+for(e=0;o-1>e;e++)if(l>=i[e][0]&&l<=i[e+1][0]){Math.abs(a(i[e][1])-h)<8&&(s.x=i[e][0],s.y=i[e][1],s.index=e,s.seriesIndex=t.index);break}},this.drawHit=function(t){var e,i=t.context,n=t.args,a=t.data,r=t.xScale,o=n.index,s=r(n.x),l=t.yScale(n.y);a.length-1>o&&(e=t.xScale(a[o+1][0]),i.save(),i.strokeStyle=t.color,i.lineWidth=t.lineWidth,i.beginPath(),i.moveTo(s,l),i.lineTo(e,l),i.stroke(),i.closePath(),i.restore())},this.clearHit=function(t){var e,i=t.context,n=t.args,a=t.data,r=t.xScale,o=t.lineWidth,s=n.index,l=r(n.x),h=t.yScale(n.y);a.length-1>s&&(e=t.xScale(a[s+1][0]),i.clearRect(l-o,h-o,e-l+2*o,2*o))})}}),Flotr.addType("bars",{options:{show:!1,lineWidth:2,barWidth:1,fill:!0,fillColor:null,fillOpacity:.4,horizontal:!1,stacked:!1,centered:!0,topPadding:.1,grouped:!1},stack:{positive:[],negative:[],_positive:[],_negative:[]},draw:function(t){var e=t.context;this.current+=1,e.save(),e.lineJoin="miter",e.lineWidth=t.lineWidth,e.strokeStyle=t.color,t.fill&&(e.fillStyle=t.fillStyle),this.plot(t),e.restore()},plot:function(t){var e,i,n,a,r,o,s=t.data,l=t.context,h=t.shadowSize;if(!(s.length<1))for(this.translate(l,t.horizontal),e=0;e<s.length;e++)i=this.getBarGeometry(s[e][0],s[e][1],t),null!==i&&(n=i.left,a=i.top,r=i.width,o=i.height,t.fill&&l.fillRect(n,a,r,o),h&&(l.save(),l.fillStyle="rgba(0,0,0,0.05)",l.fillRect(n+h,a+h,r,o),l.restore()),t.lineWidth&&l.strokeRect(n,a,r,o))},translate:function(t,e){e&&(t.rotate(-Math.PI/2),t.scale(-1,1))},getBarGeometry:function(t,e,i){var n,a,r,o,s,l=i.horizontal,h=i.barWidth,c=i.centered,u=i.stacked?this.stack:!1,d=i.lineWidth,f=c?h/2:0,p=l?i.yScale:i.xScale,g=l?i.xScale:i.yScale,m=l?e:t,v=l?t:e,x=0;return i.grouped&&(this.current/this.groups,m-=f,h/=this.groups,f=h/2,m=m+h*this.current-f),u&&(n=v>0?u.positive:u.negative,x=n[m]||x,n[m]=x+v),a=p(m-f),r=p(m+h-f),o=g(v+x),s=g(x),0>s&&(s=0),null===t||null===e?null:{x:m,y:v,xScale:p,yScale:g,top:o,left:Math.min(a,r)-d/2,width:Math.abs(r-a)-d,height:s-o}},hit:function(t){var e,i,n=t.data,a=t.args,r=a[0],o=a[1],s=t.xInverse(r.relX),l=t.yInverse(r.relY),h=this.getBarGeometry(s,l,t),c=h.width/2,u=h.left,d=h.y;for(i=n.length;i--;)e=this.getBarGeometry(n[i][0],n[i][1],t),(d>0&&d<e.y||0>d&&d>e.y)&&Math.abs(u-e.left)<c&&(o.x=n[i][0],o.y=n[i][1],o.index=i,o.seriesIndex=t.index)},drawHit:function(t){var e=t.context,i=t.args,n=this.getBarGeometry(i.x,i.y,t),a=n.left,r=n.top,o=n.width,s=n.height;e.save(),e.strokeStyle=t.color,e.lineWidth=t.lineWidth,this.translate(e,t.horizontal),e.beginPath(),e.moveTo(a,r+s),e.lineTo(a,r),e.lineTo(a+o,r),e.lineTo(a+o,r+s),t.fill&&(e.fillStyle=t.fillStyle,e.fill()),e.stroke(),e.closePath(),e.restore()},clearHit:function(t){var e=t.context,i=t.args,n=this.getBarGeometry(i.x,i.y,t),a=n.left,r=n.width,o=n.top,s=n.height,l=2*t.lineWidth;e.save(),this.translate(e,t.horizontal),e.clearRect(a-l,Math.min(o,o+s)-l,r+2*l,Math.abs(s)+2*l),e.restore()},extendXRange:function(t,e,i,n){this._extendRange(t,e,i,n),this.groups=this.groups+1||1,this.current=0},extendYRange:function(t,e,i,n){this._extendRange(t,e,i,n)},_extendRange:function(t,e,i){var n=t.options.max;if(!_.isNumber(n)&&!_.isString(n)){var a,r,o,s=t.min,l=t.max,h=i.horizontal,c=t.orientation,u=this.positiveSums||{},d=this.negativeSums||{};if((1==c&&!h||-1==c&&h)&&i.centered&&(l=Math.max(t.datamax+i.barWidth,l),s=Math.min(t.datamin-i.barWidth,s)),i.stacked&&(1==c&&h||-1==c&&!h))for(o=e.length;o--;)a=e[o][1==c?1:0]+"",r=e[o][1==c?0:1],r>0?(u[a]=(u[a]||0)+r,l=Math.max(l,u[a])):(d[a]=(d[a]||0)+r,s=Math.min(s,d[a]));(1==c&&h||-1==c&&!h)&&i.topPadding&&(t.max===t.datamax||i.stacked&&this.stackMax!==l)&&(l+=i.topPadding*(l-s)),this.stackMin=s,this.stackMax=l,this.negativeSums=d,this.positiveSums=u,t.max=l,t.min=s}}}),Flotr.addType("bubbles",{options:{show:!1,lineWidth:2,fill:!0,fillOpacity:.4,baseRadius:2},draw:function(t){var e=t.context,i=t.shadowSize;e.save(),e.lineWidth=t.lineWidth,e.fillStyle="rgba(0,0,0,0.05)",e.strokeStyle="rgba(0,0,0,0.05)",this.plot(t,i/2),e.strokeStyle="rgba(0,0,0,0.1)",this.plot(t,i/4),e.strokeStyle=t.color,e.fillStyle=t.fillStyle,this.plot(t),e.restore()},plot:function(t,e){var i,n,a=t.data,r=t.context;for(e=e||0,n=0;n<a.length;++n)i=this.getGeometry(a[n],t),r.beginPath(),r.arc(i.x+e,i.y+e,i.z,0,2*Math.PI,!0),r.stroke(),t.fill&&r.fill(),r.closePath()},getGeometry:function(t,e){return{x:e.xScale(t[0]),y:e.yScale(t[1]),z:t[2]*e.baseRadius}},hit:function(t){var e,n,a,r,o=t.data,s=t.args,l=s[0],h=s[1],c=l.relX,u=l.relY;for(h.best=h.best||Number.MAX_VALUE,i=o.length;i--;)n=this.getGeometry(o[i],t),a=n.x-c,r=n.y-u,e=Math.sqrt(a*a+r*r),e<n.z&&n.z<h.best&&(h.x=o[i][0],h.y=o[i][1],h.index=i,h.seriesIndex=t.index,h.best=n.z)},drawHit:function(t){var e=t.context,i=this.getGeometry(t.data[t.args.index],t);e.save(),e.lineWidth=t.lineWidth,e.fillStyle=t.fillStyle,e.strokeStyle=t.color,e.beginPath(),e.arc(i.x,i.y,i.z,0,2*Math.PI,!0),e.fill(),e.stroke(),e.closePath(),e.restore()},clearHit:function(t){var e=t.context,i=this.getGeometry(t.data[t.args.index],t),n=i.z+t.lineWidth;e.save(),e.clearRect(i.x-n,i.y-n,2*n,2*n),e.restore()}}),Flotr.addType("candles",{options:{show:!1,lineWidth:1,wickLineWidth:1,candleWidth:.6,fill:!0,upFillColor:"#00A8F0",downFillColor:"#CB4B4B",fillOpacity:.5,barcharts:!1},draw:function(t){var e=t.context;e.save(),e.lineJoin="miter",e.lineCap="butt",e.lineWidth=t.wickLineWidth||t.lineWidth,this.plot(t),e.restore()},plot:function(t){var e,i,n,a,r,o,s,l,h,c,u,d,f,p,g,m=t.data,v=t.context,x=t.xScale,y=t.yScale,b=t.candleWidth/2,w=t.shadowSize,M=t.lineWidth,k=t.wickLineWidth,T=k%2/2;if(!(m.length<1))for(g=0;g<m.length;g++)i=m[g],n=i[0],r=i[1],o=i[2],s=i[3],l=i[4],h=x(n-b),c=x(n+b),u=y(s),d=y(o),f=y(Math.min(r,l)),p=y(Math.max(r,l)),e=t[r>l?"downFillColor":"upFillColor"],t.fill&&!t.barcharts&&(v.fillStyle="rgba(0,0,0,0.05)",v.fillRect(h+w,p+w,c-h,f-p),v.save(),v.globalAlpha=t.fillOpacity,v.fillStyle=e,v.fillRect(h,p+M,c-h,f-p),v.restore()),(M||k)&&(n=Math.floor((h+c)/2)+T,v.strokeStyle=e,v.beginPath(),t.barcharts?(v.moveTo(n,Math.floor(d+b)),v.lineTo(n,Math.floor(u+b)),a=Math.floor(r+b)+.5,v.moveTo(Math.floor(h)+T,a),v.lineTo(n,a),a=Math.floor(l+b)+.5,v.moveTo(Math.floor(c)+T,a),v.lineTo(n,a)):(v.strokeRect(h,p+M,c-h,f-p),v.moveTo(n,Math.floor(p+M)),v.lineTo(n,Math.floor(d+M)),v.moveTo(n,Math.floor(f+M)),v.lineTo(n,Math.floor(u+M))),v.closePath(),v.stroke())},extendXRange:function(t){null===t.options.max&&(t.max=Math.max(t.datamax+.5,t.max),t.min=Math.min(t.datamin-.5,t.min))}}),Flotr.addType("gantt",{options:{show:!1,lineWidth:2,barWidth:1,fill:!0,fillColor:null,fillOpacity:.4,centered:!0},draw:function(t){var e=this.ctx,i=t.gantt.barWidth,n=Math.min(t.gantt.lineWidth,i);if(e.save(),e.translate(this.plotOffset.left,this.plotOffset.top),e.lineJoin="miter",e.lineWidth=n,e.strokeStyle=t.color,e.save(),this.gantt.plotShadows(t,i,0,t.gantt.fill),e.restore(),t.gantt.fill){var a=t.gantt.fillColor||t.color;e.fillStyle=this.processColor(a,{opacity:t.gantt.fillOpacity})}this.gantt.plot(t,i,0,t.gantt.fill),e.restore()},plot:function(t,e,i,n){var a=t.data;if(!(a.length<1)){var r,o=t.xaxis,s=t.yaxis,l=this.ctx;for(r=0;r<a.length;r++){var h=a[r][0],c=a[r][1],u=a[r][2],d=!0,f=!0,p=!0;if(null!==c&&null!==u){var g=c,m=c+u,v=h-(t.gantt.centered?e/2:0),x=h+e-(t.gantt.centered?e/2:0);m<o.min||g>o.max||x<s.min||v>s.max||(g<o.min&&(g=o.min,d=!1),m>o.max&&(m=o.max,o.lastSerie!=t&&(f=!1)),v<s.min&&(v=s.min),x>s.max&&(x=s.max,s.lastSerie!=t&&(f=!1)),n&&(l.beginPath(),l.moveTo(o.d2p(g),s.d2p(v)+i),l.lineTo(o.d2p(g),s.d2p(x)+i),l.lineTo(o.d2p(m),s.d2p(x)+i),l.lineTo(o.d2p(m),s.d2p(v)+i),l.fill(),l.closePath()),t.gantt.lineWidth&&(d||p||f)&&(l.beginPath(),l.moveTo(o.d2p(g),s.d2p(v)+i),l[d?"lineTo":"moveTo"](o.d2p(g),s.d2p(x)+i),l[f?"lineTo":"moveTo"](o.d2p(m),s.d2p(x)+i),l[p?"lineTo":"moveTo"](o.d2p(m),s.d2p(v)+i),l.stroke(),l.closePath()))}}}},plotShadows:function(t,e){var i=t.data;if(!(i.length<1)){var n,a,r,o,s=t.xaxis,l=t.yaxis,h=this.ctx,c=this.options.shadowSize;for(n=0;n<i.length;n++)if(a=i[n][0],r=i[n][1],o=i[n][2],null!==r&&null!==o){var u=r,d=r+o,f=a-(t.gantt.centered?e/2:0),p=a+e-(t.gantt.centered?e/2:0);if(!(d<s.min||u>s.max||p<l.min||f>l.max)){u<s.min&&(u=s.min),d>s.max&&(d=s.max),f<l.min&&(f=l.min),p>l.max&&(p=l.max);var g=s.d2p(d)-s.d2p(u)-(s.d2p(d)+c<=this.plotWidth?0:c),m=l.d2p(f)-l.d2p(p)-(l.d2p(f)+c<=this.plotHeight?0:c);h.fillStyle="rgba(0,0,0,0.05)",h.fillRect(Math.min(s.d2p(u)+c,this.plotWidth),Math.min(l.d2p(p)+c,this.plotHeight),g,m)}}}},extendXRange:function(t){if(null===t.options.max){var e,i,n,a,r=t.min,o=t.max,s={},l=null;for(e=0;e<this.series.length;++e)if(n=this.series[e],a=n.gantt,a.show&&n.xaxis==t){for(i=0;i<n.data.length;i++)a.show&&(y=n.data[i][0]+"",s[y]=Math.max(s[y]||0,n.data[i][1]+n.data[i][2]),l=n);for(i in s)o=Math.max(s[i],o)}t.lastSerie=l,t.max=o,t.min=r}},extendYRange:function(t){if(null===t.options.max){var e,i,n,a,r=Number.MIN_VALUE,o=Number.MAX_VALUE,s=null;for(e=0;e<this.series.length;++e)if(n=this.series[e],a=n.gantt,a.show&&!n.hide&&n.yaxis==t){var l=Number.MIN_VALUE,h=Number.MAX_VALUE;for(i=0;i<n.data.length;i++)l=Math.max(l,n.data[i][0]),h=Math.min(h,n.data[i][0]);a.centered?(r=Math.max(l+.5,r),o=Math.min(h-.5,o)):(r=Math.max(l+1,r),o=Math.min(h,o)),a.barWidth+l>r&&(r=t.max+a.barWidth)}t.lastSerie=s,t.max=r,t.min=o,t.tickSize=Flotr.getTickSize(t.options.noTicks,o,r,t.options.tickDecimals)}}}),function(){function t(t){return"object"==typeof t&&t.constructor&&(Image?!0:t.constructor===Image)}Flotr.defaultMarkerFormatter=function(t){return Math.round(100*t.y)/100+""},Flotr.addType("markers",{options:{show:!1,lineWidth:1,color:"#000000",fill:!1,fillColor:"#FFFFFF",fillOpacity:.4,stroke:!1,position:"ct",verticalMargin:0,labelFormatter:Flotr.defaultMarkerFormatter,fontSize:Flotr.defaultOptions.fontSize,stacked:!1,stackingType:"b",horizontal:!1},stack:{positive:[],negative:[],values:[]},draw:function(t){function e(t,e){return n=u.negative[t]||0,i=u.positive[t]||0,e>0?(u.positive[t]=n+e,n+e):(u.negative[t]=i+e,i+e)}var i,n,a,r,o,s,l,h=t.data,c=t.context,u=t.stacked?t.stack:!1,d=t.stackingType;for(c.save(),c.lineJoin="round",c.lineWidth=t.lineWidth,c.strokeStyle="rgba(0,0,0,0.5)",c.fillStyle=t.fillStyle,r=0;r<h.length;++r)o=h[r][0],s=h[r][1],u&&("b"==d?t.horizontal?s=e(s,o):o=e(o,s):"a"==d&&(a=u.values[o]||0,u.values[o]=a+s,s=a+s)),l=t.labelFormatter({x:o,y:s,index:r,data:h}),this.plot(t.xScale(o),t.yScale(s),l,t);c.restore()},plot:function(e,i,n,a){a.context;if(t(n)&&!n.complete)throw"Marker image not loaded.";this._plot(e,i,n,a)},_plot:function(e,i,n,a){var r,o=a.context,s=2,l=e,h=i;r=t(n)?{height:n.height,width:n.width}:a.text.canvas(n),r.width=Math.floor(r.width+2*s),r.height=Math.floor(r.height+2*s),-1!=a.position.indexOf("c")?l-=r.width/2+s:-1!=a.position.indexOf("l")&&(l-=r.width),-1!=a.position.indexOf("m")?h-=r.height/2+s:-1!=a.position.indexOf("t")?h-=r.height+a.verticalMargin:h+=a.verticalMargin,l=Math.floor(l)+.5,h=Math.floor(h)+.5,a.fill&&o.fillRect(l,h,r.width,r.height),a.stroke&&o.strokeRect(l,h,r.width,r.height),t(n)?o.drawImage(n,l+s,h+s):Flotr.drawText(o,n,l+s,h+s,{textBaseline:"top",textAlign:"left",size:a.fontSize,color:a.color})}})}(),function(){Flotr._;Flotr.defaultPieLabelFormatter=function(t,e){return(100*e/t).toFixed(2)+"%"},Flotr.addType("pie",{options:{show:!1,lineWidth:1,fill:!0,fillColor:null,fillOpacity:.6,explode:6,sizeRatio:.6,startAngle:Math.PI/4,labelFormatter:Flotr.defaultPieLabelFormatter,pie3D:!1,pie3DviewAngle:Math.PI/2*.8,pie3DspliceThickness:20,epsilon:.1},draw:function(t){var e,i,n,a=t.data,r=t.context,o=r.canvas,s=t.lineWidth,l=t.shadowSize,h=t.sizeRatio,c=t.height,u=t.width,d=t.explode,f=t.color,p=t.fill,g=t.fillStyle,m=Math.min(o.width,o.height)*h/2,v=a[0][1],x=[],y=1,b=2*Math.PI*v/this.total,w=this.startAngle||2*Math.PI*t.startAngle,M=w+b,k=w+b/2,T=t.labelFormatter(this.total,v),S=d+m+4,F=Math.cos(k)*S,_=Math.sin(k)*S,C=0>F?"right":"left",D=_>0?"top":"bottom";if(r.save(),r.translate(u/2,c/2),r.scale(1,y),i=Math.cos(k)*d,n=Math.sin(k)*d,l>0&&(this.plotSlice(i+l,n+l,m,w,M,r),p&&(r.fillStyle="rgba(0,0,0,0.1)",r.fill())),this.plotSlice(i,n,m,w,M,r),p&&(r.fillStyle=g,r.fill()),r.lineWidth=s,r.strokeStyle=f,r.stroke(),e={size:1.2*t.fontSize,color:t.fontColor,weight:1.5},T&&(t.htmlText||!t.textEnabled?(divStyle="position:absolute;"+D+":"+(c/2+("top"===D?_:-_))+"px;",divStyle+=C+":"+(u/2+("right"===C?-F:F))+"px;",x.push('<div style="',divStyle,'" class="flotr-grid-label">',T,"</div>")):(e.textAlign=C,e.textBaseline=D,Flotr.drawText(r,T,F,_,e))),t.htmlText||!t.textEnabled){var A=Flotr.DOM.node('<div style="color:'+t.fontColor+'" class="flotr-labels"></div>');Flotr.DOM.insert(A,x.join("")),Flotr.DOM.insert(t.element,A)}r.restore(),this.startAngle=M,this.slices=this.slices||[],this.slices.push({radius:Math.min(o.width,o.height)*h/2,x:i,y:n,explode:d,start:w,end:M})},plotSlice:function(t,e,i,n,a,r){r.beginPath(),r.moveTo(t,e),r.arc(t,e,i,n,a,!1),r.lineTo(t,e),r.closePath()},hit:function(t){var e=t.data[0],i=t.args,n=t.index,a=i[0],r=i[1],o=this.slices[n],s=a.relX-t.width/2,l=a.relY-t.height/2,h=Math.sqrt(s*s+l*l),c=Math.atan(l/s),u=2*Math.PI,d=o.explode||t.explode,f=o.start%u,p=o.end%u,g=t.epsilon;0>s?c+=Math.PI:s>0&&0>l&&(c+=u),h<o.radius+d&&h>d&&(c>f&&p>c||f>p&&(p>c||c>f)||f===p&&(o.start===o.end&&Math.abs(c-f)<g||o.start!==o.end&&Math.abs(c-f)>g))&&(r.x=e[0],r.y=e[1],r.sAngle=f,r.eAngle=p,r.index=0,r.seriesIndex=n,r.fraction=e[1]/this.total)},drawHit:function(t){var e=t.context,i=this.slices[t.args.seriesIndex];e.save(),e.translate(t.width/2,t.height/2),this.plotSlice(i.x,i.y,i.radius,i.start,i.end,e),e.stroke(),e.restore()},clearHit:function(t){var e=t.context,i=this.slices[t.args.seriesIndex],n=2*t.lineWidth,a=i.radius+n;e.save(),e.translate(t.width/2,t.height/2),e.clearRect(i.x-a,i.y-a,2*a+n,2*a+n),e.restore()},extendYRange:function(t,e){this.total=(this.total||0)+e[0][1]}})}(),Flotr.addType("points",{options:{show:!1,radius:3,lineWidth:2,fill:!0,fillColor:"#FFFFFF",fillOpacity:1,hitRadius:null},draw:function(t){var e=t.context,i=(t.lineWidth,t.shadowSize);e.save(),i>0&&(e.lineWidth=i/2,e.strokeStyle="rgba(0,0,0,0.1)",this.plot(t,i/2+e.lineWidth/2),e.strokeStyle="rgba(0,0,0,0.2)",this.plot(t,e.lineWidth/2)),e.lineWidth=t.lineWidth,e.strokeStyle=t.color,t.fill&&(e.fillStyle=t.fillStyle),this.plot(t),e.restore()},plot:function(t,e){var i,n,a,r=t.data,o=t.context,s=t.xScale,l=t.yScale;for(i=r.length-1;i>-1;--i)a=r[i][1],null!==a&&(n=s(r[i][0]),a=l(a),0>n||n>t.width||0>a||a>t.height||(o.beginPath(),e?o.arc(n,a+e,t.radius,0,Math.PI,!1):(o.arc(n,a,t.radius,0,2*Math.PI,!0),t.fill&&o.fill()),o.stroke(),o.closePath()))}}),Flotr.addType("radar",{options:{show:!1,lineWidth:2,fill:!0,fillOpacity:.4,radiusRatio:.9},draw:function(t){var e=t.context,i=t.shadowSize;e.save(),e.translate(t.width/2,t.height/2),e.lineWidth=t.lineWidth,e.fillStyle="rgba(0,0,0,0.05)",e.strokeStyle="rgba(0,0,0,0.05)",this.plot(t,i/2),e.strokeStyle="rgba(0,0,0,0.1)",this.plot(t,i/4),e.strokeStyle=t.color,e.fillStyle=t.fillStyle,this.plot(t),e.restore()},plot:function(t,e){var i,n,a=t.data,r=t.context,o=Math.min(t.height,t.width)*t.radiusRatio/2,s=2*Math.PI/a.length,l=-Math.PI/2;for(e=e||0,r.beginPath(),i=0;i<a.length;++i)n=a[i][1]/this.max,r[0===i?"moveTo":"lineTo"](Math.cos(i*s+l)*o*n+e,Math.sin(i*s+l)*o*n+e);r.closePath(),t.fill&&r.fill(),r.stroke()},extendYRange:function(t){this.max=Math.max(t.max,this.max||-Number.MAX_VALUE)}}),Flotr.addType("timeline",{options:{show:!1,lineWidth:1,barWidth:.2,fill:!0,fillColor:null,fillOpacity:.4,centered:!0},draw:function(t){var e=t.context;e.save(),e.lineJoin="miter",e.lineWidth=t.lineWidth,e.strokeStyle=t.color,e.fillStyle=t.fillStyle,this.plot(t),e.restore()},plot:function(t){var e=t.data,i=t.context,n=t.xScale,a=t.yScale,r=t.barWidth,o=t.lineWidth;Flotr._.each(e,function(t){var e=t[0],s=t[1],l=t[2],h=r,c=Math.ceil(n(e)),u=Math.ceil(n(e+l))-c,d=Math.round(a(s)),f=Math.round(a(s-h))-d,p=c-o/2,g=Math.round(d-f/2)-o/2;i.strokeRect(p,g,u,f),i.fillRect(p,g,u,f)})},extendRange:function(t){var e=t.data,i=t.xaxis,n=t.yaxis,a=t.timeline.barWidth;if(null===i.options.min&&(i.min=i.datamin-a/2),null===i.options.max){var r=i.max;Flotr._.each(e,function(t){r=Math.max(r,t[0]+t[2])},this),i.max=r+a/2}null===n.options.min&&(n.min=n.datamin-a),null===n.options.min&&(n.max=n.datamax+a)}}),function(){var t=Flotr.DOM;Flotr.addPlugin("crosshair",{options:{mode:null,color:"#FF0000",hideCursor:!0},callbacks:{"flotr:mousemove":function(t,e){this.options.crosshair.mode&&(this.crosshair.clearCrosshair(),this.crosshair.drawCrosshair(e))}},drawCrosshair:function(e){var i=this.octx,n=this.options.crosshair,a=this.plotOffset,r=a.left+Math.round(e.relX)+.5,o=a.top+Math.round(e.relY)+.5;return e.relX<0||e.relY<0||e.relX>this.plotWidth||e.relY>this.plotHeight?(this.el.style.cursor=null,void t.removeClass(this.el,"flotr-crosshair")):(n.hideCursor&&(this.el.style.cursor="none",t.addClass(this.el,"flotr-crosshair")),i.save(),i.strokeStyle=n.color,i.lineWidth=1,i.beginPath(),-1!=n.mode.indexOf("x")&&(i.moveTo(r,a.top),i.lineTo(r,a.top+this.plotHeight)),-1!=n.mode.indexOf("y")&&(i.moveTo(a.left,o),i.lineTo(a.left+this.plotWidth,o)),i.stroke(),i.restore(),void 0)},clearCrosshair:function(){var t=this.plotOffset,e=this.lastMousePos,i=this.octx;e&&(i.clearRect(Math.round(e.relX)+t.left,t.top,1,this.plotHeight+1),i.clearRect(t.left,Math.round(e.relY)+t.top,this.plotWidth+1,1))}})}(),function(){function t(t,e){var i="image/"+t,n=e.toDataURL(i),a=new Image;return a.src=n,a}var e=Flotr.DOM,i=Flotr._;Flotr.addPlugin("download",{saveImage:function(n,a,r,o){var s=null;if(Flotr.isIE&&Flotr.isIE<9)return s="<html><body>"+this.canvas.firstChild.innerHTML+"</body></html>",window.open().document.write(s);if("jpeg"===n||"png"===n)return s=t(n,this.canvas,a,r),i.isElement(s)&&o?(this.download.restoreCanvas(),e.hide(this.canvas),e.hide(this.overlay),e.setStyles({position:"absolute"}),e.insert(this.el,s),this.saveImageElement=s,void 0):window.open(s.src)},restoreCanvas:function(){e.show(this.canvas),e.show(this.overlay),this.saveImageElement&&this.el.removeChild(this.saveImageElement),this.saveImageElement=null}})}(),function(){var t=Flotr.EventAdapter,e=Flotr._;Flotr.addPlugin("graphGrid",{callbacks:{"flotr:beforedraw":function(){this.graphGrid.drawGrid()},"flotr:afterdraw":function(){this.graphGrid.drawOutline()}},drawGrid:function(){function i(t){for(s=0;s<t.length;++s){var e=t[s].v/o.max;for(l=0;y>=l;++l)h[0===l?"moveTo":"lineTo"](Math.cos(l*b+w)*x*e,Math.sin(l*b+w)*x*e)}}function n(t,i){e.each(e.pluck(t,"v"),function(t){t<=o.min||t>=o.max||(t==o.min||t==o.max)&&u.outlineWidth||i(Math.floor(o.d2p(t))+h.lineWidth/2)})}function a(t){h.moveTo(t,0),h.lineTo(t,m)}function r(t){h.moveTo(0,t),h.lineTo(v,t)}var o,s,l,h=this.ctx,c=this.options,u=c.grid,d=u.verticalLines,f=u.horizontalLines,p=u.minorVerticalLines,g=u.minorHorizontalLines,m=this.plotHeight,v=this.plotWidth;if((d||p||f||g)&&t.fire(this.el,"flotr:beforegrid",[this.axes.x,this.axes.y,c,this]),h.save(),h.lineWidth=1,h.strokeStyle=u.tickColor,u.circular){h.translate(this.plotOffset.left+v/2,this.plotOffset.top+m/2);var x=Math.min(m,v)*c.radar.radiusRatio/2,y=this.axes.x.ticks.length,b=2*(Math.PI/y),w=-Math.PI/2;h.beginPath(),o=this.axes.y,f&&i(o.ticks),g&&i(o.minorTicks),d&&e.times(y,function(t){h.moveTo(0,0),h.lineTo(Math.cos(t*b+w)*x,Math.sin(t*b+w)*x)}),h.stroke()}else h.translate(this.plotOffset.left,this.plotOffset.top),u.backgroundColor&&(h.fillStyle=this.processColor(u.backgroundColor,{x1:0,y1:0,x2:v,y2:m}),h.fillRect(0,0,v,m)),h.beginPath(),o=this.axes.x,d&&n(o.ticks,a),p&&n(o.minorTicks,a),o=this.axes.y,f&&n(o.ticks,r),g&&n(o.minorTicks,r),h.stroke();h.restore(),(d||p||f||g)&&t.fire(this.el,"flotr:aftergrid",[this.axes.x,this.axes.y,c,this])},drawOutline:function(){var t,e,n,a,r=this,o=r.options,s=o.grid,l=s.outline,h=r.ctx,c=s.backgroundImage,u=r.plotOffset,d=u.left,f=u.top,p=r.plotWidth,g=r.plotHeight;if(s.outlineWidth){if(h.save(),s.circular){h.translate(d+p/2,f+g/2);var m=Math.min(g,p)*o.radar.radiusRatio/2,v=this.axes.x.ticks.length,x=2*(Math.PI/v),y=-Math.PI/2;for(h.beginPath(),h.lineWidth=s.outlineWidth,h.strokeStyle=s.color,h.lineJoin="round",i=0;i<=v;++i)h[0===i?"moveTo":"lineTo"](Math.cos(i*x+y)*m,Math.sin(i*x+y)*m);h.stroke()}else{h.translate(d,f);var b=s.outlineWidth,w=.5-b+(b+1)%2/2,M="lineTo",k="moveTo";h.lineWidth=b,h.strokeStyle=s.color,h.lineJoin="miter",h.beginPath(),h.moveTo(w,w),p-=b/2%1,g+=b/2,h[-1!==l.indexOf("n")?M:k](p,w),h[-1!==l.indexOf("e")?M:k](p,g),h[-1!==l.indexOf("s")?M:k](w,g),h[-1!==l.indexOf("w")?M:k](w,w),h.stroke(),h.closePath()}h.restore(),c&&(e=c.src||c,n=(parseInt(c.left,10)||0)+u.left,a=(parseInt(c.top,10)||0)+u.top,t=new Image,t.onload=function(){h.save(),c.alpha&&(h.globalAlpha=c.alpha),h.globalCompositeOperation="destination-over",h.drawImage(t,0,0,t.width,t.height,n,a,p,g),h.restore()},t.src=e)}}})}(),function(){var t=Flotr.DOM,e=Flotr._,i=Flotr,n="opacity:0.7;background-color:#000;color:#fff;display:none;position:absolute;padding:2px 8px;-moz-border-radius:4px;border-radius:4px;white-space:nowrap;";Flotr.addPlugin("hit",{callbacks:{"flotr:mousemove":function(t,e){this.hit.track(e)},"flotr:click":function(t){var i=this.hit.track(t);e.defaults(t,i)},"flotr:mouseout":function(){this.hit.clearHit()},"flotr:destroy":function(){this.mouseTrack=null}},track:function(t){return this.options.mouse.track||e.any(this.series,function(t){return t.mouse&&t.mouse.track})?this.hit.hit(t):void 0},executeOnType:function(t,n,a){function r(t,r){e.each(e.keys(i.graphTypes),function(e){t[e]&&t[e].show&&this[e][n]&&(o=this.getOptions(t,e),o.fill=!!t.mouse.fillColor,o.fillStyle=this.processColor(t.mouse.fillColor||"#ffffff",{opacity:t.mouse.fillOpacity}),o.color=t.mouse.lineColor,o.context=this.octx,o.index=r,a&&(o.args=a),this[e][n].call(this[e],o),s=!0)},this)}var o,s=!1;return e.isArray(t)||(t=[t]),e.each(t,r,this),s},drawHit:function(t){var e=this.octx,i=t.series;if(i.mouse.lineColor){if(e.save(),e.lineWidth=i.points?i.points.lineWidth:1,e.strokeStyle=i.mouse.lineColor,e.fillStyle=this.processColor(i.mouse.fillColor||"#ffffff",{opacity:i.mouse.fillOpacity}),e.translate(this.plotOffset.left,this.plotOffset.top),!this.hit.executeOnType(i,"drawHit",t)){var n=t.xaxis,a=t.yaxis;e.beginPath(),e.arc(n.d2p(t.x),a.d2p(t.y),i.points.hitRadius||i.points.radius||i.mouse.radius,0,2*Math.PI,!0),e.fill(),e.stroke(),e.closePath()}e.restore(),this.clip(e)}this.prevHit=t},clearHit:function(){var e=this.prevHit,i=this.octx,n=this.plotOffset;if(i.save(),i.translate(n.left,n.top),e){if(!this.hit.executeOnType(e.series,"clearHit",this.prevHit)){var a=e.series,r=a.points?a.points.lineWidth:1;offset=(a.points.hitRadius||a.points.radius||a.mouse.radius)+r,i.clearRect(e.xaxis.d2p(e.x)-offset,e.yaxis.d2p(e.y)-offset,2*offset,2*offset)}t.hide(this.mouseTrack),this.prevHit=null}i.restore()},hit:function(t){var i,n,a,r,o,s,l,h=this.options,c=this.prevHit;if(0!==this.series.length)return l={relX:t.relX,relY:t.relY,absX:t.absX,absY:t.absY},h.mouse.trackY&&!h.mouse.trackAll&&this.hit.executeOnType(this.series,"hit",[t,l])&&!e.isUndefined(l.seriesIndex)?(r=this.series[l.seriesIndex],l.series=r,l.mouse=r.mouse,l.xaxis=r.xaxis,l.yaxis=r.yaxis):(i=this.hit.closest(t),i&&(i=h.mouse.trackY?i.point:i.x,a=i.seriesIndex,r=this.series[a],o=r.xaxis,s=r.yaxis,n=2*r.mouse.sensibility,(h.mouse.trackAll||i.distanceX<n/o.scale&&(!h.mouse.trackY||i.distanceY<n/s.scale))&&(l.series=r,l.xaxis=r.xaxis,l.yaxis=r.yaxis,l.mouse=r.mouse,l.x=i.x,l.y=i.y,l.dist=i.distance,l.index=i.dataIndex,l.seriesIndex=a))),c&&c.index===l.index&&c.seriesIndex===l.seriesIndex||(this.hit.clearHit(),l.series&&l.mouse&&l.mouse.track&&(this.hit.drawMouseTrack(l),this.hit.drawHit(l),Flotr.EventAdapter.fire(this.el,"flotr:hit",[l,this]))),l},closest:function(t){function e(t){t.distance=a,t.distanceX=r,t.distanceY=o,t.seriesIndex=u,t.dataIndex=d,t.x=h,t.y=c,b=!0}var i,n,a,r,o,s,l,h,c,u,d,f=this.series,p=(this.options,t.relX),g=t.relY,m=Number.MAX_VALUE,v=Number.MAX_VALUE,x={},y={},b=!1;for(u=0;u<f.length;u++)for(i=f[u],n=i.data,s=i.xaxis.p2d(p),l=i.yaxis.p2d(g),d=n.length;d--;)h=n[d][0],c=n[d][1],null!==h&&null!==c&&(h<i.xaxis.min||h>i.xaxis.max||(r=Math.abs(h-s),o=Math.abs(c-l),a=r*r+o*o,m>a&&(m=a,e(x)),v>r&&(v=r,e(y))));return b?{point:x,x:y}:!1},drawMouseTrack:function(e){{var i="",a=e.series,r=e.mouse.position,o=e.mouse.margin,s=e.x,l=e.y,h=n,c=this.mouseTrack,u=this.plotOffset,d=u.left,f=u.right,p=u.bottom,g=u.top,m=e.mouse.trackDecimals;this.options}if(c||(c=t.node('<div class="flotr-mouse-value"></div>'),this.mouseTrack=c,t.insert(this.el,c)),e.mouse.relative)if(a.pie&&a.pie.show){var v={x:this.plotWidth/2,y:this.plotHeight/2},x=Math.min(this.canvasWidth,this.canvasHeight)*a.pie.sizeRatio/2,y=e.sAngle<e.eAngle?(e.sAngle+e.eAngle)/2:(e.sAngle+e.eAngle+2*Math.PI)/2;i+="bottom:"+(o-g-v.y-Math.sin(y)*x/2+this.canvasHeight)+"px;top:auto;",i+="left:"+(o+d+v.x+Math.cos(y)*x/2)+"px;right:auto;"}else i+=/n/.test(r)?"bottom:"+(o-g-e.yaxis.d2p(e.y)+this.canvasHeight)+"px;top:auto;":"top:"+(o+g+e.yaxis.d2p(e.y))+"px;bottom:auto;",i+=/w/.test(r)?"right:"+(o-d-e.xaxis.d2p(e.x)+this.canvasWidth)+"px;left:auto;":"left:"+(o+d+e.xaxis.d2p(e.x))+"px;right:auto;";else"n"==r.charAt(0)?i+="top:"+(o+g)+"px;bottom:auto;":"s"==r.charAt(0)&&(i+="bottom:"+(o+p)+"px;top:auto;"),"e"==r.charAt(1)?i+="right:"+(o+f)+"px;left:auto;":"w"==r.charAt(1)&&(i+="left:"+(o+d)+"px;right:auto;");h+=i,c.style.cssText=h,(!m||0>m)&&(m=0),s&&s.toFixed&&(s=s.toFixed(m)),l&&l.toFixed&&(l=l.toFixed(m)),c.innerHTML=e.mouse.trackFormatter({x:s,y:l,series:e.series,index:e.index,nearest:e,fraction:e.fraction}),t.show(c),e.mouse.relative&&(/[ew]/.test(r)?/[ns]/.test(r)||(c.style.top=g+e.yaxis.d2p(e.y)-t.size(c).height/2+"px"):c.style.left=d+e.xaxis.d2p(e.x)-t.size(c).width/2+"px")}})}(),function(){function t(t){return t.which?1===t.which:0===t.button||1===t.button}function e(t,e){return Math.min(Math.max(0,t),e.plotWidth-1)}function i(t,e){return Math.min(Math.max(0,t),e.plotHeight)}var n=(Flotr.DOM,Flotr.EventAdapter),a=Flotr._;Flotr.addPlugin("selection",{options:{pinchOnly:null,mode:null,color:"#B6D9FF",fps:20},callbacks:{"flotr:mouseup":function(t){var e=this.options.selection,i=this.selection,n=this.getEventPosition(t);e&&e.mode&&(i.interval&&clearInterval(i.interval),this.multitouches?i.updateSelection():e.pinchOnly||i.setSelectionPos(i.selection.second,n),i.clearSelection(),i.selecting&&i.selectionIsSane()&&(i.drawSelection(),i.fireSelectEvent(),this.ignoreClick=!0))},"flotr:mousedown":function(e){var i=this.options.selection,n=this.selection,r=this.getEventPosition(e);i&&i.mode&&(!i.mode||!t(e)&&a.isUndefined(e.touches)||(i.pinchOnly||n.setSelectionPos(n.selection.first,r),n.interval&&clearInterval(n.interval),this.lastMousePos.pageX=null,n.selecting=!1,n.interval=setInterval(a.bind(n.updateSelection,this),1e3/i.fps)))},"flotr:destroy":function(){clearInterval(this.selection.interval)}},getArea:function(){{var t,e,i,n,a=this.selection.selection,r=this.axes;a.first,a.second}return t=r.x.p2d(a.first.x),e=r.x.p2d(a.second.x),i=r.y.p2d(a.first.y),n=r.y.p2d(a.second.y),{x1:Math.min(t,e),y1:Math.min(i,n),x2:Math.max(t,e),y2:Math.max(i,n),xfirst:t,xsecond:e,yfirst:i,ysecond:n}},selection:{first:{x:-1,y:-1},second:{x:-1,y:-1}},prevSelection:null,interval:null,fireSelectEvent:function(t){var e=this.selection.getArea();t=t||"select",e.selection=this.selection.selection,n.fire(this.el,"flotr:"+t,[e,this])},setSelection:function(t,n){var a=this.options,r=this.axes.x,o=this.axes.y,s=o.scale,l=r.scale,h=-1!=a.selection.mode.indexOf("x"),c=-1!=a.selection.mode.indexOf("y"),u=this.selection.selection;this.selection.clearSelection(),u.first.y=i(h&&!c?0:(o.max-t.y1)*s,this),u.second.y=i(h&&!c?this.plotHeight-1:(o.max-t.y2)*s,this),u.first.x=e(c&&!h?0:(t.x1-r.min)*l,this),u.second.x=e(c&&!h?this.plotWidth:(t.x2-r.min)*l,this),this.selection.drawSelection(),n||this.selection.fireSelectEvent()},setSelectionPos:function(t,n){var a=this.options.selection.mode,r=this.selection.selection;t.x=-1==a.indexOf("x")?t==r.first?0:this.plotWidth:e(n.relX,this),t.y=-1==a.indexOf("y")?t==r.first?0:this.plotHeight-1:i(n.relY,this)},drawSelection:function(){this.selection.fireSelectEvent("selecting");var t=this.selection.selection,e=this.octx,i=this.options,n=this.plotOffset,a=this.selection.prevSelection;if(!a||t.first.x!=a.first.x||t.first.y!=a.first.y||t.second.x!=a.second.x||t.second.y!=a.second.y){e.save(),e.strokeStyle=this.processColor(i.selection.color,{opacity:.8}),e.lineWidth=1,e.lineJoin="miter",e.fillStyle=this.processColor(i.selection.color,{opacity:.4}),this.selection.prevSelection={first:{x:t.first.x,y:t.first.y},second:{x:t.second.x,y:t.second.y}};var r=Math.min(t.first.x,t.second.x),o=Math.min(t.first.y,t.second.y),s=Math.abs(t.second.x-t.first.x),l=Math.abs(t.second.y-t.first.y);e.fillRect(r+n.left+.5,o+n.top+.5,s,l),e.strokeRect(r+n.left+.5,o+n.top+.5,s,l),e.restore()}},updateSelection:function(){if(this.lastMousePos.pageX){if(this.selection.selecting=!0,this.multitouches)this.selection.setSelectionPos(this.selection.selection.first,this.getEventPosition(this.multitouches[0])),this.selection.setSelectionPos(this.selection.selection.second,this.getEventPosition(this.multitouches[1]));else{if(this.options.selection.pinchOnly)return;this.selection.setSelectionPos(this.selection.selection.second,this.lastMousePos)}this.selection.clearSelection(),this.selection.selectionIsSane()&&this.selection.drawSelection()}},clearSelection:function(){if(this.selection.prevSelection){var t=this.selection.prevSelection,e=1,i=this.plotOffset,n=Math.min(t.first.x,t.second.x),a=Math.min(t.first.y,t.second.y),r=Math.abs(t.second.x-t.first.x),o=Math.abs(t.second.y-t.first.y);this.octx.clearRect(n+i.left-e+.5,a+i.top-e,r+2*e+.5,o+2*e+.5),this.selection.prevSelection=null}},selectionIsSane:function(){var t=this.selection.selection;return Math.abs(t.second.x-t.first.x)>=5||Math.abs(t.second.y-t.first.y)>=5}})}(),function(){var t=Flotr.DOM;Flotr.addPlugin("labels",{callbacks:{"flotr:afterdraw":function(){this.labels.draw()}},draw:function(){function e(t,e,i){{var n,r=i?e.minorTicks:e.ticks,s=1===e.orientation;1===e.n}for(n={color:e.options.color||p.grid.color,angle:Flotr.toRad(e.options.labelsAngle),textBaseline:"middle"},u=0;u<r.length&&(i?e.options.showMinorLabels:e.options.showLabels);++u)a=r[u],a.label+="",a.label&&a.label.length&&(x=Math.cos(u*l+h)*o,y=Math.sin(u*l+h)*o,n.textAlign=s?Math.abs(x)<.1?"center":x<0?"right":"left":"left",Flotr.drawText(g,a.label,s?x:3,s?y:-(e.ticks[u].v/e.max)*(o-p.fontSize),n))}function i(t,e,i,n){function r(t){return t.options.showLabels&&t.used}function o(t,e,i,n){return t.plotOffset.left+(e?n:i?-p.grid.labelMargin:p.grid.labelMargin+t.plotWidth)}function s(t,e,i,n){return t.plotOffset.top+(e?p.grid.labelMargin:n)+(e&&i?t.plotHeight:0)}var l,h,c=1===e.orientation,d=1===e.n;for(l={color:e.options.color||p.grid.color,textAlign:i,textBaseline:n,angle:Flotr.toRad(e.options.labelsAngle)},l=Flotr.getBestTextAlign(l.angle,l),u=0;u<e.ticks.length&&r(e);++u)a=e.ticks[u],a.label&&a.label.length&&(h=e.d2p(a.v),0>h||h>(c?t.plotWidth:t.plotHeight)||(Flotr.drawText(g,a.label,o(t,c,d,h),s(t,c,d,h),l),!c&&!d&&(g.save(),g.strokeStyle=l.color,g.beginPath(),g.moveTo(t.plotOffset.left+t.plotWidth-8,t.plotOffset.top+e.d2p(a.v)),g.lineTo(t.plotOffset.left+t.plotWidth,t.plotOffset.top+e.d2p(a.v)),g.stroke(),g.restore())))}function n(t,e){var i,n,o=1===e.orientation,s=1===e.n,l="",h=t.plotOffset;if(!o&&!s&&(g.save(),g.strokeStyle=e.options.color||p.grid.color,g.beginPath()),e.options.showLabels&&(s?!0:e.used))for(u=0;u<e.ticks.length;++u)a=e.ticks[u],!a.label||!a.label.length||(o?h.left:h.top)+e.d2p(a.v)<0||(o?h.left:h.top)+e.d2p(a.v)>(o?t.canvasWidth:t.canvasHeight)||(n=h.top+(o?(s?1:-1)*(t.plotHeight+p.grid.labelMargin):e.d2p(a.v)-e.maxLabel.height/2),i=o?h.left+e.d2p(a.v)-r/2:0,l="",0===u?l=" first":u===e.ticks.length-1&&(l=" last"),l+=o?" flotr-grid-label-x":" flotr-grid-label-y",d+=['<div style="position:absolute; text-align:'+(o?"center":"right")+"; ","top:"+n+"px; ",(o||s?"left:":"right:")+i+"px; ","width:"+(o?r:(s?h.left:h.right)-p.grid.labelMargin)+"px; ",e.options.color?"color:"+e.options.color+"; ":" ",'" class="flotr-grid-label'+l+'">'+a.label+"</div>"].join(" "),!o&&!s&&(g.moveTo(h.left+t.plotWidth-8,h.top+e.d2p(a.v)),g.lineTo(h.left+t.plotWidth,h.top+e.d2p(a.v))))
+}{var a,r,o,s,l,h,c,u,d="",f=0,p=this.options,g=this.ctx,m=this.axes;({size:p.fontSize})}for(u=0;u<m.x.ticks.length;++u)m.x.ticks[u].label&&++f;r=this.plotWidth/f,p.grid.circular&&(g.save(),g.translate(this.plotOffset.left+this.plotWidth/2,this.plotOffset.top+this.plotHeight/2),o=this.plotHeight*p.radar.radiusRatio/2+p.fontSize,s=this.axes.x.ticks.length,l=2*(Math.PI/s),h=-Math.PI/2,e(this,m.x,!1),e(this,m.x,!0),e(this,m.y,!1),e(this,m.y,!0),g.restore()),!p.HtmlText&&this.textEnabled?(i(this,m.x,"center","top"),i(this,m.x2,"center","bottom"),i(this,m.y,"right","middle"),i(this,m.y2,"left","middle")):(m.x.options.showLabels||m.x2.options.showLabels||m.y.options.showLabels||m.y2.options.showLabels)&&!p.grid.circular&&(d="",n(this,m.x),n(this,m.x2),n(this,m.y),n(this,m.y2),g.stroke(),g.restore(),c=t.create("div"),t.setStyles(c,{fontSize:"smaller",color:p.grid.color}),c.className="flotr-labels",t.insert(this.el,c),t.insert(c,d))}})}(),function(){var t=Flotr.DOM,e=Flotr._;Flotr.addPlugin("legend",{options:{show:!0,noColumns:1,labelFormatter:function(t){return t},labelBoxBorderColor:"#CCCCCC",labelBoxWidth:14,labelBoxHeight:10,labelBoxMargin:5,container:null,position:"nw",margin:5,backgroundColor:"#F0F0F0",backgroundOpacity:.85},callbacks:{"flotr:afterinit":function(){this.legend.insertLegend()},"flotr:destroy":function(){var e=this.legend.markup;e&&(this.legend.markup=null,t.remove(e))}},insertLegend:function(){if(this.options.legend.show){var i,n,a,r=this.series,o=this.plotOffset,s=this.options,l=s.legend,h=[],c=!1,u=this.ctx,d=e.filter(r,function(t){return t.label&&!t.hide}).length,f=l.position,p=l.margin,g=l.backgroundOpacity;if(d){var m=l.labelBoxWidth,v=l.labelBoxHeight,x=l.labelBoxMargin,y=o.left+p,b=o.top+p,w=0,M={size:1.1*s.fontSize,color:s.grid.color};for(i=r.length-1;i>-1;--i)r[i].label&&!r[i].hide&&(n=l.labelFormatter(r[i].label),w=Math.max(w,this._text.measureText(n,M).width));var k=Math.round(m+3*x+w),T=Math.round(d*(x+v)+x);if(!g&&0!==g&&(g=.1),s.HtmlText||!this.textEnabled||l.container){for(i=0;i<r.length;++i)if(r[i].label&&!r[i].hide){i%l.noColumns===0&&(h.push(c?"</tr><tr>":"<tr>"),c=!0);var S=r[i],F=l.labelBoxWidth,_=l.labelBoxHeight;n=l.labelFormatter(S.label),a="background-color:"+(S.bars&&S.bars.show&&S.bars.fillColor&&S.bars.fill?S.bars.fillColor:S.color)+";",h.push('<td class="flotr-legend-color-box">','<div style="border:1px solid ',l.labelBoxBorderColor,';padding:1px">','<div style="width:',F-1,"px;height:",_-1,"px;border:1px solid ",r[i].color,'">','<div style="width:',F,"px;height:",_,"px;",a,'"></div>',"</div>","</div>","</td>",'<td class="flotr-legend-label">',n,"</td>")}if(c&&h.push("</tr>"),h.length>0){var C='<table style="font-size:smaller;color:'+s.grid.color+'">'+h.join("")+"</table>";if(l.container)C=t.node(C),this.legend.markup=C,t.insert(l.container,C);else{var D={position:"absolute",zIndex:"2",border:"1px solid "+l.labelBoxBorderColor};"n"==f.charAt(0)?(D.top=p+o.top+"px",D.bottom="auto"):"c"==f.charAt(0)?(D.top=p+(this.plotHeight-T)/2+"px",D.bottom="auto"):"s"==f.charAt(0)&&(D.bottom=p+o.bottom+"px",D.top="auto"),"e"==f.charAt(1)?(D.right=p+o.right+"px",D.left="auto"):"w"==f.charAt(1)&&(D.left=p+o.left+"px",D.right="auto");var A=t.create("div");if(A.className="flotr-legend",t.setStyles(A,D),t.insert(A,C),t.insert(this.el,A),!g)return;var O=l.backgroundColor||s.grid.backgroundColor||"#ffffff";e.extend(D,t.size(A),{backgroundColor:O,zIndex:"",border:""}),D.width+="px",D.height+="px",A=t.create("div"),A.className="flotr-legend-bg",t.setStyles(A,D),t.opacity(A,g),t.insert(A," "),t.insert(this.el,A)}}}else{"s"==f.charAt(0)&&(b=o.top+this.plotHeight-(p+T)),"c"==f.charAt(0)&&(b=o.top+this.plotHeight/2-(p+T/2)),"e"==f.charAt(1)&&(y=o.left+this.plotWidth-(p+k)),a=this.processColor(l.backgroundColor,{opacity:g}),u.fillStyle=a,u.fillRect(y,b,k,T),u.strokeStyle=l.labelBoxBorderColor,u.strokeRect(Flotr.toPixel(y),Flotr.toPixel(b),k,T);var z=y+x,L=b+x;for(i=0;i<r.length;i++)r[i].label&&!r[i].hide&&(n=l.labelFormatter(r[i].label),u.fillStyle=r[i].color,u.fillRect(z,L,m-1,v-1),u.strokeStyle=l.labelBoxBorderColor,u.lineWidth=1,u.strokeRect(Math.ceil(z)-1.5,Math.ceil(L)-1.5,m+2,v+2),Flotr.drawText(u,n,z+m+x,L+v,M),L+=v+x)}}}}})}(),function(){function t(t){if(this.options.spreadsheet.tickFormatter)return this.options.spreadsheet.tickFormatter(t);var e=i.find(this.axes.x.ticks,function(e){return e.v==t});return e?e.label:t}var e=Flotr.DOM,i=Flotr._;Flotr.addPlugin("spreadsheet",{options:{show:!1,tabGraphLabel:"Graph",tabDataLabel:"Data",toolbarDownload:"Download CSV",toolbarSelectAll:"Select all",csvFileSeparator:",",decimalSeparator:".",tickFormatter:null,initialTab:"graph"},callbacks:{"flotr:afterconstruct":function(){if(this.options.spreadsheet.show){var t,i=this.spreadsheet,n=e.node('<div class="flotr-tabs-group" style="position:absolute;left:0px;width:'+this.canvasWidth+'px"></div>'),a=e.node('<div style="float:left" class="flotr-tab selected">'+this.options.spreadsheet.tabGraphLabel+"</div>"),r=e.node('<div style="float:left" class="flotr-tab">'+this.options.spreadsheet.tabDataLabel+"</div>");i.tabsContainer=n,i.tabs={graph:a,data:r},e.insert(n,a),e.insert(n,r),e.insert(this.el,n),t=e.size(r).height+2,this.plotOffset.bottom+=t,e.setStyles(n,{top:this.canvasHeight-t+"px"}),this.observe(a,"click",function(){i.showTab("graph")}).observe(r,"click",function(){i.showTab("data")}),"graph"!==this.options.spreadsheet.initialTab&&i.showTab(this.options.spreadsheet.initialTab)}}},loadDataGrid:function(){if(this.seriesData)return this.seriesData;var t=this.series,e={};return i.each(t,function(t,n){i.each(t.data,function(t){var i=t[0],a=t[1],r=e[i];if(r)r[n+1]=a;else{var o=[];o[0]=i,o[n+1]=a,e[i]=o}})}),this.seriesData=i.sortBy(e,function(t,e){return parseInt(e,10)}),this.seriesData},constructDataGrid:function(){if(this.spreadsheet.datagrid)return this.spreadsheet.datagrid;var n,a,r,o=this.series,s=this.spreadsheet.loadDataGrid(),l=["<colgroup><col />"],h=['<table class="flotr-datagrid"><tr class="first-row">'];h.push("<th>&nbsp;</th>"),i.each(o,function(t,e){h.push('<th scope="col">'+(t.label||String.fromCharCode(65+e))+"</th>"),l.push("<col />")}),h.push("</tr>"),i.each(s,function(e){h.push("<tr>"),i.times(o.length+1,function(n){var a="td",r=e[n],o=i.isUndefined(r)?"":Math.round(1e5*r)/1e5;if(0===n){a="th";var s=t.call(this,o);s&&(o=s)}h.push("<"+a+("th"==a?' scope="row"':"")+">"+o+"</"+a+">")},this),h.push("</tr>")},this),l.push("</colgroup>"),r=e.node(h.join("")),n=e.node('<button type="button" class="flotr-datagrid-toolbar-button">'+this.options.spreadsheet.toolbarDownload+"</button>"),a=e.node('<button type="button" class="flotr-datagrid-toolbar-button">'+this.options.spreadsheet.toolbarSelectAll+"</button>"),this.observe(n,"click",i.bind(this.spreadsheet.downloadCSV,this)).observe(a,"click",i.bind(this.spreadsheet.selectAllData,this));var c=e.node('<div class="flotr-datagrid-toolbar"></div>');e.insert(c,n),e.insert(c,a);var u=this.canvasHeight-e.size(this.spreadsheet.tabsContainer).height-2,d=e.node('<div class="flotr-datagrid-container" style="position:absolute;left:0px;top:0px;width:'+this.canvasWidth+"px;height:"+u+'px;overflow:auto;z-index:10"></div>');return e.insert(d,c),e.insert(d,r),e.insert(this.el,d),this.spreadsheet.datagrid=r,this.spreadsheet.container=d,r},showTab:function(t){if(this.spreadsheet.activeTab!==t){switch(t){case"graph":e.hide(this.spreadsheet.container),e.removeClass(this.spreadsheet.tabs.data,"selected"),e.addClass(this.spreadsheet.tabs.graph,"selected");break;case"data":this.spreadsheet.datagrid||this.spreadsheet.constructDataGrid(),e.show(this.spreadsheet.container),e.addClass(this.spreadsheet.tabs.data,"selected"),e.removeClass(this.spreadsheet.tabs.graph,"selected");break;default:throw"Illegal tab name: "+t}this.spreadsheet.activeTab=t}},selectAllData:function(){if(this.spreadsheet.tabs){var t,e,i,n,a=this.spreadsheet.constructDataGrid();return this.spreadsheet.showTab("data"),setTimeout(function(){(i=a.ownerDocument)&&(n=i.defaultView)&&n.getSelection&&i.createRange&&(t=window.getSelection())&&t.removeAllRanges?(e=i.createRange(),e.selectNode(a),t.removeAllRanges(),t.addRange(e)):document.body&&document.body.createTextRange&&(e=document.body.createTextRange())&&(e.moveToElementText(a),e.select())},0),!0}return!1},downloadCSV:function(){var e="",n=this.series,a=this.options,r=this.spreadsheet.loadDataGrid(),o=encodeURIComponent(a.spreadsheet.csvFileSeparator);if(a.spreadsheet.decimalSeparator===a.spreadsheet.csvFileSeparator)throw"The decimal separator is the same as the column separator ("+a.spreadsheet.decimalSeparator+")";i.each(n,function(t,i){e+=o+'"'+(t.label||String.fromCharCode(65+i)).replace(/\"/g,'\\"')+'"'}),e+="%0D%0A",e+=i.reduce(r,function(e,i){var n=t.call(this,i[0])||"";n='"'+(n+"").replace(/\"/g,'\\"')+'"';var r=i.slice(1).join(o);return"."!==a.spreadsheet.decimalSeparator&&(r=r.replace(/\./g,a.spreadsheet.decimalSeparator)),e+n+o+r+"%0D%0A"},"",this),Flotr.isIE&&Flotr.isIE<9?(e=e.replace(new RegExp(o,"g"),decodeURIComponent(o)).replace(/%0A/g,"\n").replace(/%0D/g,"\r"),window.open().document.write(e)):window.open("data:text/csv,"+e)}})}(),function(){var t=Flotr.DOM;Flotr.addPlugin("titles",{callbacks:{"flotr:afterdraw":function(){this.titles.drawTitles()}},drawTitles:function(){var e,i=this.options,n=i.grid.labelMargin,a=this.ctx,r=this.axes;if(!i.HtmlText&&this.textEnabled){var o={size:i.fontSize,color:i.grid.color,textAlign:"center"};i.subtitle&&Flotr.drawText(a,i.subtitle,this.plotOffset.left+this.plotWidth/2,this.titleHeight+this.subtitleHeight-2,o),o.weight=1.5,o.size*=1.5,i.title&&Flotr.drawText(a,i.title,this.plotOffset.left+this.plotWidth/2,this.titleHeight-2,o),o.weight=1.8,o.size*=.8,r.x.options.title&&r.x.used&&(o.textAlign=r.x.options.titleAlign||"center",o.textBaseline="top",o.angle=Flotr.toRad(r.x.options.titleAngle),o=Flotr.getBestTextAlign(o.angle,o),Flotr.drawText(a,r.x.options.title,this.plotOffset.left+this.plotWidth/2,this.plotOffset.top+r.x.maxLabel.height+this.plotHeight+2*n,o)),r.x2.options.title&&r.x2.used&&(o.textAlign=r.x2.options.titleAlign||"center",o.textBaseline="bottom",o.angle=Flotr.toRad(r.x2.options.titleAngle),o=Flotr.getBestTextAlign(o.angle,o),Flotr.drawText(a,r.x2.options.title,this.plotOffset.left+this.plotWidth/2,this.plotOffset.top-r.x2.maxLabel.height-2*n,o)),r.y.options.title&&r.y.used&&(o.textAlign=r.y.options.titleAlign||"right",o.textBaseline="middle",o.angle=Flotr.toRad(r.y.options.titleAngle),o=Flotr.getBestTextAlign(o.angle,o),Flotr.drawText(a,r.y.options.title,this.plotOffset.left-r.y.maxLabel.width-2*n,this.plotOffset.top+this.plotHeight/2,o)),r.y2.options.title&&r.y2.used&&(o.textAlign=r.y2.options.titleAlign||"left",o.textBaseline="middle",o.angle=Flotr.toRad(r.y2.options.titleAngle),o=Flotr.getBestTextAlign(o.angle,o),Flotr.drawText(a,r.y2.options.title,this.plotOffset.left+this.plotWidth+r.y2.maxLabel.width+2*n,this.plotOffset.top+this.plotHeight/2,o))}else{e=[],i.title&&e.push('<div style="position:absolute;top:0;left:',this.plotOffset.left,"px;font-size:1em;font-weight:bold;text-align:center;width:",this.plotWidth,'px;" class="flotr-title">',i.title,"</div>"),i.subtitle&&e.push('<div style="position:absolute;top:',this.titleHeight,"px;left:",this.plotOffset.left,"px;font-size:smaller;text-align:center;width:",this.plotWidth,'px;" class="flotr-subtitle">',i.subtitle,"</div>"),e.push("</div>"),e.push('<div class="flotr-axis-title" style="font-weight:bold;">'),r.x.options.title&&r.x.used&&e.push('<div style="position:absolute;top:',this.plotOffset.top+this.plotHeight+i.grid.labelMargin+r.x.titleSize.height,"px;left:",this.plotOffset.left,"px;width:",this.plotWidth,"px;text-align:",r.x.options.titleAlign,';" class="flotr-axis-title flotr-axis-title-x1">',r.x.options.title,"</div>"),r.x2.options.title&&r.x2.used&&e.push('<div style="position:absolute;top:0;left:',this.plotOffset.left,"px;width:",this.plotWidth,"px;text-align:",r.x2.options.titleAlign,';" class="flotr-axis-title flotr-axis-title-x2">',r.x2.options.title,"</div>"),r.y.options.title&&r.y.used&&e.push('<div style="position:absolute;top:',this.plotOffset.top+this.plotHeight/2-r.y.titleSize.height/2,"px;left:0;text-align:",r.y.options.titleAlign,';" class="flotr-axis-title flotr-axis-title-y1">',r.y.options.title,"</div>"),r.y2.options.title&&r.y2.used&&e.push('<div style="position:absolute;top:',this.plotOffset.top+this.plotHeight/2-r.y.titleSize.height/2,"px;right:0;text-align:",r.y2.options.titleAlign,';" class="flotr-axis-title flotr-axis-title-y2">',r.y2.options.title,"</div>"),e=e.join("");var s=t.create("div");t.setStyles({color:i.grid.color}),s.className="flotr-titles",t.insert(this.el,s),t.insert(s,e)}}})}(),!function($){$.extend({tablesorter:new function(){function log(t){"undefined"!=typeof console&&"undefined"!=typeof console.log?console.log(t):alert(t)}function benchmark(t,e){log(t+" ("+((new Date).getTime()-e.getTime())+"ms)")}function getElementText(t,e,i){var n="",a=t.textExtraction;return e?(t.supportsTextContent||(t.supportsTextContent=e.textContent||!1),n="simple"===a?t.supportsTextContent?e.textContent:e.childNodes[0]&&e.childNodes[0].hasChildNodes()?e.childNodes[0].innerHTML:e.innerHTML:"function"==typeof a?a(e):"object"==typeof a&&a.hasOwnProperty(i)?a[i](e):$(e).text()):""}function getParserById(t){var e,i=g.length;for(e=0;i>e;e++)if(g[e].id.toLowerCase()===t.toLowerCase())return g[e];return!1}function getNodeFromRowAndCellIndex(t,e,i){return t[e].cells[i]}function trimAndGetNodeText(t,e,i){return $.trim(getElementText(t,e,i))}function detectParserForColumn(t,e,i,n){for(var a,r=g.length,o=!1,s="",l=!0;""===s&&l;)i++,e[i]?(o=getNodeFromRowAndCellIndex(e,i,n),s=trimAndGetNodeText(t.config,o,n),t.config.debug&&log("Checking if value was empty on row "+i+", column:"+n+": "+s)):l=!1;for(a=1;r>a;a++)if(g[a].is(s,t,o))return g[a];return g[0]}function buildParserCache(t,e){if(0!==t.tBodies.length){var i,n,a,r,o,s,l=t.tBodies[0].rows,h="";if(l[0])for(i=[],n=l[0].cells,a=n.length,o=0;a>o;o++)s=!1,r=$(e[o]),$.metadata&&r.metadata()&&r.metadata().sorter?s=getParserById(r.metadata().sorter):t.config.headers[o]&&t.config.headers[o].sorter?s=getParserById(t.config.headers[o].sorter):r.attr("class")&&r.attr("class").match("sorter-")&&(s=getParserById(r.attr("class").match(/sorter-(\w+)/)[1]||"")),s||(s=detectParserForColumn(t,l,-1,o)),t.config.debug&&(h+="column:"+o+"; parser:"+s.id+"\n"),i.push(s);return t.config.debug&&log(h),i}}function buildCache(t){var e,i,n,a,r,o,s=t.tBodies[0],l=s&&s.rows.length||0,h=s.rows[0]&&s.rows[0].cells.length||0,c=t.config.parsers,u={row:[],normalized:[]};for(t.config.debug&&(o=new Date),i=0;l>i;++i)if(a=$(s.rows[i]),r=[],a.hasClass(t.config.cssChildRow))u.row[u.row.length-1]=u.row[u.row.length-1].add(a);else{for(u.row.push(a),n=0;h>n;++n)e=trimAndGetNodeText(t.config,a[0].cells[n],n),r.push(""===e?"":c[n].format(e,t,a[0].cells[n],n));r.push(u.normalized.length),u.normalized.push(r)}return t.config.debug&&benchmark("Building cache for "+l+" rows",o),t.config.cache=u,u}function getWidgetById(t){var e,i,n=widgets.length;for(e=0;n>e;e++)if(i=widgets[e],i&&i.hasOwnProperty("id")&&i.id.toLowerCase()===t.toLowerCase())return i}function applyWidget(t,e){var i,n,a=t.config.widgets,r=a.length;for(i=0;r>i;i++)n=getWidgetById(a[i]),n&&(e&&n.hasOwnProperty("init")?n.init(t,widgets,n):!e&&n.hasOwnProperty("format")&&n.format(t))}function appendToTable(t,e){var i,n,a,r,o,s=t.config,l=e.row,h=e.normalized,c=h.length,u=c?h[0].length-1:0,d=[];for(s.debug&&(o=new Date),i=0;c>i;i++)if(r=h[i][u],d.push(l[r]),!s.appender||!s.removeRows)for(a=l[r].length,n=0;a>n;n++)t.tBodies[0].appendChild(l[r][n]);s.appender&&s.appender(t,d),s.debug&&benchmark("Rebuilt table",o),applyWidget(t),setTimeout(function(){$(t).trigger("sortEnd",t)},0)}function computeTableHeaderCellIndexes(t){var e,i,n,a,r,o,s,l,h,c,u,d,f=[],p={},g=t.getElementsByTagName("THEAD")[0],m=g.getElementsByTagName("TR");for(e=0;e<m.length;e++)for(o=m[e].cells,i=0;i<o.length;i++){for(r=o[i],s=r.parentNode.rowIndex,l=s+"-"+r.cellIndex,h=r.rowSpan||1,c=r.colSpan||1,"undefined"==typeof f[s]&&(f[s]=[]),n=0;n<f[s].length+1;n++)if("undefined"==typeof f[s][n]){u=n;break}for(p[l]=u,n=s;s+h>n;n++)for("undefined"==typeof f[n]&&(f[n]=[]),d=f[n],a=u;u+c>a;a++)d[a]="x"}return p}function formatSortingOrder(t){return"number"!=typeof t?"d"===t.toLowerCase().charAt(0)?1:0:1===t?1:0}function checkHeaderMetadata(t){return $.metadata&&$(t).metadata().sorter===!1}function checkHeaderOptions(t,e){return t.config.headers[e]&&t.config.headers[e].sorter===!1}function checkHeaderLocked(t,e){return t.config.headers[e]&&null!==t.config.headers[e].lockedOrder?t.config.headers[e].lockedOrder:!1}function checkHeaderOrder(t,e){return t.config.headers[e]&&t.config.headers[e].sortInitialOrder?t.config.headers[e].sortInitialOrder:t.config.sortInitialOrder}function buildHeaders(t){var e,i,n,a,r=($.metadata?!0:!1,computeTableHeaderCellIndexes(t)),o=t.config;return o.headerList=[],o.debug&&(n=new Date),a=$(o.selectorHeaders,t).wrapInner("<span/>").each(function(n){this.column=r[this.parentNode.rowIndex+"-"+this.cellIndex],this.order=formatSortingOrder(checkHeaderOrder(t,n)),this.count=this.order,(checkHeaderMetadata(this)||checkHeaderOptions(t,n)||$(this).is(".sorter-false"))&&(this.sortDisabled=!0),this.lockedOrder=!1,i=checkHeaderLocked(t,n),"undefined"!=typeof i&&i!==!1&&(this.order=this.lockedOrder=formatSortingOrder(i)),this.sortDisabled||(e=$(this).addClass(o.cssHeader),o.onRenderHeader&&o.onRenderHeader.apply(e,[n])),o.headerList[n]=this}),o.debug&&(benchmark("Built headers",n),log(a)),a}function checkCellColSpan(t,e,i){var n,a,r=[],o=t.tHead.rows,s=o[i].cells;for(n=0;n<s.length;n++)a=s[n],a.colSpan>1?r=r.concat(checkCellColSpan(t,e,i++)):(1===t.tHead.length||a.rowSpan>1||!o[i+1])&&r.push(a);return r}function isValueInArray(t,e){var i,n=e.length;for(i=0;n>i;i++)if(e[i][0]===t)return!0;return!1}function setHeadersCss(t,e,i){var n,a,r=[],o=[t.config.cssDesc,t.config.cssAsc];for(e.removeClass(o[0]).removeClass(o[1]),e.each(function(){this.sortDisabled||(r[this.column]=$(this))}),a=i.length,n=0;a>n;n++)2!==i[n][1]&&r[i[n][0]].addClass(o[i[n][1]])}function fixColumnWidth(t){if(t.config.widthFixed){var e=$("<colgroup>");$("tr:first td",t.tBodies[0]).each(function(){e.append($("<col>").css("width",$(this).width()))}),$(t).prepend(e)}}function updateHeaderSortCount(t,e){var i,n,a,r=t.config,o=e.length;for(i=0;o>i;i++)n=e[i],a=r.headerList[n[0]],a.count=n[1],a.count++}function getCachedSortType(t,e){return t?t[e].type:""}function multisort(a,b,d){var f="var sortWrapper = function(a,b) {",col,mx=0,dir=0,tc=a.config,lc=d.normalized.length,l=b.length,sortTime,i,j,c,s,e,order,orgOrderCol;for(tc.debug&&(sortTime=new Date),i=0;l>i;i++){if(c=b[i][0],order=b[i][1],s="text"===getCachedSortType(tc.parsers,c)?0===order?"sortText":"sortTextDesc":0===order?"sortNumeric":"sortNumericDesc",e="e"+i,/Numeric/.test(s)&&tc.headers[c]&&tc.headers[c].string){for(j=0;lc>j;j++)col=Math.abs(parseFloat(d.normalized[j][c])),mx=Math.max(mx,isNaN(col)?0:col);dir=tc.headers[c]?tc.string[tc.headers[c].string]||0:0}f+="var "+e+" = "+s+"(a["+c+"],b["+c+"],"+mx+","+dir+"); ",f+="if ("+e+") { return "+e+"; } ",f+="else { "}for(orgOrderCol=d.normalized&&d.normalized[0]?d.normalized[0].length-1:0,f+="return a["+orgOrderCol+"]-b["+orgOrderCol+"];",i=0;l>i;i++)f+="}; ";return f+="return 0; ",f+="}; ",eval(f),d.normalized.sort(sortWrapper),tc.debug&&benchmark("Sorting on "+b.toString()+" and dir "+order+" time",sortTime),d}function sortText(t,e){if(""===t)return 1;if(""===e)return-1;if(t===e)return 0;if($.data(tbl[0],"tablesorter").sortLocaleCompare)return t.localeCompare(e);try{for(var i,n=0,a=/^(\.)?\d/,r=Math.min(t.length,e.length)+1;r>n&&t.charAt(n)===e.charAt(n)&&a.test(e.substring(n))===!1&&a.test(t.substring(n))===!1;)n++;if(t=t.substring(n),e=e.substring(n),a.test(t)||a.test(e)){if(a.test(t)===!1)return t?1:-1;if(a.test(e)===!1)return e?-1:1;if(i=parseFloat(t)-parseFloat(e),0!==i)return i;i=t.search(/[^\.\d]/),-1===i&&(i=e.search(/[^\.\d]/)),t=t.substring(i),e=e.substring(i)}return t>e?1:-1}catch(o){return 0}}function sortTextDesc(t,e){return""===t?1:""===e?-1:t===e?0:$.data(tbl[0],"tablesorter").sortLocaleCompare?e.localeCompare(t):-sortText(t,e)}function getTextValue(t,e,i){if(e){var n,a=t.length,r=e+i;for(n=0;a>n;n++)r+=t.charCodeAt(n);return i*r}return 0}function sortNumeric(t,e,i,n){return""===t?1:""===e?-1:(isNaN(t)&&(t=getTextValue(t,i,n)),isNaN(e)&&(e=getTextValue(e,i,n)),t-e)}function sortNumericDesc(t,e,i,n){return""===t?1:""===e?-1:(isNaN(t)&&(t=getTextValue(t,i,n)),isNaN(e)&&(e=getTextValue(e,i,n)),e-t)}var g=[],widgets=[],tbl;this.defaults={cssHeader:"header",cssAsc:"headerSortUp",cssDesc:"headerSortDown",cssChildRow:"expand-child",sortInitialOrder:"asc",sortMultiSortKey:"shiftKey",sortForce:null,sortAppend:null,sortLocaleCompare:!1,sortReset:!1,sortRestart:!1,textExtraction:"simple",parsers:{},widgets:[],widgetZebra:{css:["even","odd"]},headers:{},widthFixed:!1,cancelSelection:!0,sortList:[],headerList:[],dateFormat:"mmddyyyy",onRenderHeader:null,selectorHeaders:"thead th",tableClass:"tablesorter",debug:!1},this.benchmark=benchmark,this.hasInitialized=!1,this.construct=function(t){return this.each(function(){if(this.tHead&&this.tBodies){var e,i,n,a,r,o,s,l,h,c,u,d;this.config={},s=a=$.extend(this.config,$.tablesorter.defaults,t),tbl=e=$(this).addClass(this.config.tableClass),$.data(this,"tablesorter",s),i=buildHeaders(this),this.config.parsers=buildParserCache(this,i),this.config.string={max:1,"max+":1,"max-":-1,none:0},n=buildCache(this),fixColumnWidth(this),i.click(function(t){if(r=e[0].tBodies[0]&&e[0].tBodies[0].rows.length||0,!this.sortDisabled){if(e.trigger("sortStart",tbl[0]),o=$(this),this.order=this.count++%(s.sortReset?3:2),s.sortRestart&&(l=this,i.each(function(){this===l||$(this).is("."+s.cssDesc+",."+s.cssAsc)&&t[s.sortMultiSortKey]||(this.count=0)})),l=this.column,"undefined"!=typeof this.lockedOrder&&this.lockedOrder!==!1&&(this.order=this.lockedOrder),t[s.sortMultiSortKey])if(isValueInArray(l,s.sortList))for(h=0;h<s.sortList.length;h++)u=s.sortList[h],d=s.headerList[u[0]],u[0]===l&&(d.count=u[1],d.count++,u[1]=d.count%(s.sortReset?3:2),u[1]>=2&&(s.sortList.splice(h,1),d.count=0));else this.order<2&&s.sortList.push([l,this.order]);else{if(s.sortList=[],null!==s.sortForce)for(c=s.sortForce,h=0;h<c.length;h++)c[h][0]!==l&&s.sortList.push(c[h]);this.order<2&&s.sortList.push([l,this.order])}if(null!==s.sortAppend)for(c=s.sortAppend,h=0;h<c.length;h++)c[h][0]!==l&&s.sortList.push(c[h]);return e.trigger("sortBegin",tbl[0]),setTimeout(function(){setHeadersCss(e[0],i,s.sortList),appendToTable(e[0],multisort(e[0],s.sortList,n))},1),!1}}).mousedown(function(){return s.cancelSelection?(this.onselectstart=function(){return!1},!1):void 0}),e.bind("update",function(){var t=this;setTimeout(function(){t.config.parsers=buildParserCache(t,i),n=buildCache(t),e.trigger("sorton",[t.config.sortList])},1)}).bind("updateCell",function(t,i){var a=[i.parentNode.rowIndex-1,i.cellIndex];n.normalized[a[0]][a[1]]=s.parsers[a[1]].format(getElementText(s,i,a[1]),i),s.cache=n,e.trigger("sorton",[s.sortList])}).bind("addRows",function(t,i){var a,r=i.filter("tr").length,o=[],l=i[0].cells.length;for(a=0;r>a;a++){for(h=0;l>h;h++)o[h]=s.parsers[h].format(getElementText(s,i[a].cells[h],h),i[a].cells[h]);o.push(n.row.length),n.row.push([i[a]]),n.normalized.push(o),o=[]}s.cache=n,e.trigger("sorton",[s.sortList])}).bind("sorton",function(t,e){$(this).trigger("sortStart",tbl[0]),s.sortList=e;var a=s.sortList;updateHeaderSortCount(this,a),setHeadersCss(this,i,a),appendToTable(this,multisort(this,a,n))}).bind("appendCache",function(){appendToTable(this,n)}).bind("applyWidgetId",function(t,e){getWidgetById(e).format(this)}).bind("applyWidgets",function(){applyWidget(this)}),$.metadata&&$(this).metadata()&&$(this).metadata().sortlist&&(s.sortList=$(this).metadata().sortlist),applyWidget(this,!0),s.sortList.length>0?e.trigger("sorton",[s.sortList]):applyWidget(this),this.hasInitialized=!0}})},this.addParser=function(t){var e,i=g.length,n=!0;for(e=0;i>e;e++)g[e].id.toLowerCase()===t.id.toLowerCase()&&(n=!1);n&&g.push(t)},this.addWidget=function(t){widgets.push(t)},this.formatFloat=function(t){var e=parseFloat(t);return isNaN(e)?$.trim(t):e},this.isDigit=function(t){return/^[\-+]?\d*$/.test($.trim(t.replace(/[,.']/g,"")))},this.clearTableBody=function(t){if($.browser.msie){var e=function(){for(;this.firstChild;)this.removeChild(this.firstChild)};e.apply(t.tBodies[0])}else t.tBodies[0].innerHTML=""}}})(),$.fn.extend({tablesorter:$.tablesorter.construct});var m=$.tablesorter;m.addParser({id:"text",is:function(){return!0},format:function(t){return $.trim(t.toLocaleLowerCase())},type:"text"}),m.addParser({id:"digit",is:function(t){return $.tablesorter.isDigit(t.replace(/,/g,""))},format:function(t){return $.tablesorter.formatFloat(t.replace(/,/g,""))},type:"numeric"}),m.addParser({id:"currency",is:function(t){return/^[\u00a3$\u20ac\u00a4\u00a5\u00a2?.]/.test(t)},format:function(t){return $.tablesorter.formatFloat(t.replace(/\,/g,".").replace(new RegExp(/[^0-9. \-]/g),""))},type:"numeric"}),m.addParser({id:"ipAddress",is:function(t){return/^\d{2,3}[\.]\d{2,3}[\.]\d{2,3}[\.]\d{2,3}$/.test(t)},format:function(t){var e,i,n=t.split("."),a="",r=n.length;for(e=0;r>e;e++)i=n[e],a+=2===i.length?"0"+i:i;return $.tablesorter.formatFloat(a)},type:"numeric"}),m.addParser({id:"url",is:function(t){return/^(https?|ftp|file):\/\/$/.test(t)},format:function(t){return $.trim(t.replace(new RegExp(/(https?|ftp|file):\/\//),""))},type:"text"}),m.addParser({id:"isoDate",is:function(t){return/^\d{4}[\/\-]\d{1,2}[\/\-]\d{1,2}$/.test(t)},format:function(t){return $.tablesorter.formatFloat(""!==t?new Date(t.replace(new RegExp(/-/g),"/")).getTime():"0")},type:"numeric"}),m.addParser({id:"percent",is:function(t){return/\%$/.test($.trim(t))},format:function(t){return $.tablesorter.formatFloat(t.replace(new RegExp(/%/g),""))},type:"numeric"}),m.addParser({id:"usLongDate",is:function(t){return t.match(new RegExp(/^[A-Za-z]{3,10}\.? [0-9]{1,2}, ([0-9]{4}|'?[0-9]{2}) (([0-2]?[0-9]:[0-5][0-9])|([0-1]?[0-9]:[0-5][0-9]\s(AM|PM)))$/))},format:function(t){return $.tablesorter.formatFloat(new Date(t).getTime())},type:"numeric"}),m.addParser({id:"shortDate",is:function(t){return/\d{1,4}[\/\-\,\.\s+]\d{1,4}[\/\-\.\,\s+]\d{1,4}/.test(t)},format:function(t,e,i,n){var a=e.config,r=a.headers&&a.headers[n]?a.headers[n].dateFormat||a.dateFormat:a.dateFormat;return t=t.replace(/\s+/g," ").replace(/[\-|\.|\,|\s]/g,"/"),"mmddyyyy"===r?t=t.replace(/(\d{1,2})\/(\d{1,2})\/(\d{4})/,"$3/$1/$2"):"ddmmyyyy"===r?t=t.replace(/(\d{1,2})\/(\d{1,2})\/(\d{4})/,"$3/$2/$1"):"yyyymmdd"===r&&(t=t.replace(/(\d{4})\/(\d{1,2})\/(\d{1,2})/,"$1/$2/$3")),$.tablesorter.formatFloat(new Date(t).getTime())},type:"numeric"}),m.addParser({id:"time",is:function(t){return/^(([0-2]?[0-9]:[0-5][0-9])|([0-1]?[0-9]:[0-5][0-9]\s(am|pm)))$/.test(t)},format:function(t){return $.tablesorter.formatFloat(new Date("2000/01/01 "+t).getTime())},type:"numeric"}),m.addParser({id:"metadata",is:function(){return!1},format:function(t,e,i){var n=e.config,a=n.parserMetadataName?n.parserMetadataName:"sortValue";return $(i).metadata()[a]},type:"numeric"}),m.addWidget({id:"zebra",format:function(t){var e,i,n,a=0,r=t.config.cssChildRow,o=t.config.widgetZebra.css;t.config.debug&&(n=new Date),$("tr:visible",t.tBodies[0]).each(function(){e=$(this),e.hasClass(r)||a++,i=a%2===0,e.removeClass(o[i?1:0]).addClass(o[i?0:1])}),t.config.debug&&$.tablesorter.benchmark("Applying Zebra widget",n)}})}(jQuery),function(t){var e=new Array,n=new Array;t.fn.doAutosize=function(e){var i=t(this).data("minwidth"),n=t(this).data("maxwidth"),a="",r=t(this),o=t("#"+t(this).data("tester_id"));if(a!==(a=r.val())){var s=a.replace(/&/g,"&").replace(/\s/g," ").replace(/</g,"<").replace(/>/g,">");o.html(s);var l=o.width(),h=l+e.comfortZone>=i?l+e.comfortZone:i,c=r.width(),u=c>h&&h>=i||h>i&&n>h;u&&r.width(h)}},t.fn.resetAutosize=function(e){var i=t(this).data("minwidth")||e.minInputWidth||t(this).width(),n=t(this).data("maxwidth")||e.maxInputWidth||t(this).closest(".tagsinput").width()-e.inputPadding,a=t(this),r=t("<tester/>").css({position:"absolute",top:-9999,left:-9999,width:"auto",fontSize:a.css("fontSize"),fontFamily:a.css("fontFamily"),fontWeight:a.css("fontWeight"),letterSpacing:a.css("letterSpacing"),whiteSpace:"nowrap"}),o=t(this).attr("id")+"_autosize_tester";!t("#"+o).length>0&&(r.attr("id",o),r.appendTo("body")),a.data("minwidth",i),a.data("maxwidth",n),a.data("tester_id",o),a.css("width",i)},t.fn.addTag=function(i,a){return a=jQuery.extend({focus:!1,callback:!0},a),this.each(function(){var r=t(this).attr("id"),o=t(this).val().split(e[r]);if(""==o[0]&&(o=new Array),i=jQuery.trim(i),a.unique){var s=t(o).tagExist(i);1==s&&t("#"+r+"_tag").addClass("not_valid")}else var s=!1;if(""!=i&&1!=s){if(t("<span>").addClass("tag").append(t("<span>").text(i).append("  "),t("<a>",{href:"#",title:"Removing tag",text:"x"}).click(function(){return t("#"+r).removeTag(escape(i))})).insertBefore("#"+r+"_addTag"),o.push(i),t("#"+r+"_tag").val(""),a.focus?t("#"+r+"_tag").focus():t("#"+r+"_tag").blur(),t.fn.tagsInput.updateTagsField(this,o),a.callback&&n[r]&&n[r].onAddTag){var l=n[r].onAddTag;l.call(this,i)}if(n[r]&&n[r].onChange){var h=o.length,l=n[r].onChange;l.call(this,t(this),o[h-1])}}}),!1},t.fn.removeTag=function(a){return a=unescape(a),this.each(function(){var r=t(this).attr("id"),o=t(this).val().split(e[r]);for(t("#"+r+"_tagsinput .tag").remove(),str="",i=0;i<o.length;i++)o[i]!=a&&(str=str+e[r]+o[i]);if(t.fn.tagsInput.importTags(this,str),n[r]&&n[r].onRemoveTag){var s=n[r].onRemoveTag;s.call(this,a)}}),!1},t.fn.tagExist=function(e){return jQuery.inArray(e,t(this))>=0},t.fn.importTags=function(e){id=t(this).attr("id"),t("#"+id+"_tagsinput .tag").remove(),t.fn.tagsInput.importTags(this,e)},t.fn.tagsInput=function(i){var a=jQuery.extend({interactive:!0,defaultText:"add a tag",minChars:0,width:"300px",height:"100px",autocomplete:{selectFirst:!1},hide:!0,delimiter:",",unique:!0,removeWithBackspace:!0,placeholderColor:"#666666",autosize:!0,comfortZone:20,inputPadding:12},i);return this.each(function(){a.hide&&t(this).hide();var i=t(this).attr("id");(!i||e[t(this).attr("id")])&&(i=t(this).attr("id","tags"+(new Date).getTime()).attr("id"));var r=jQuery.extend({pid:i,real_input:"#"+i,holder:"#"+i+"_tagsinput",input_wrapper:"#"+i+"_addTag",fake_input:"#"+i+"_tag"},a);e[i]=r.delimiter,(a.onAddTag||a.onRemoveTag||a.onChange)&&(n[i]=new Array,n[i].onAddTag=a.onAddTag,n[i].onRemoveTag=a.onRemoveTag,n[i].onChange=a.onChange);var o='<div id="'+i+'_tagsinput" class="tagsinput"><div id="'+i+'_addTag">';if(a.interactive&&(o=o+'<input id="'+i+'_tag" value="" data-default="'+a.defaultText+'" />'),o+='</div><div class="tags_clear"></div></div>',t(o).insertAfter(this),t(r.holder).css("width",a.width),t(r.holder).css("height",a.height),""!=t(r.real_input).val()&&t.fn.tagsInput.importTags(t(r.real_input),t(r.real_input).val()),a.interactive){if(t(r.fake_input).val(t(r.fake_input).attr("data-default")),t(r.fake_input).css("color",a.placeholderColor),t(r.fake_input).resetAutosize(a),t(r.holder).bind("click",r,function(e){t(e.data.fake_input).focus()}),t(r.fake_input).bind("focus",r,function(e){t(e.data.fake_input).val()==t(e.data.fake_input).attr("data-default")&&t(e.data.fake_input).val(""),t(e.data.fake_input).css("color","#000000")}),void 0!=a.autocomplete_url){autocomplete_options={source:a.autocomplete_url};for(attrname in a.autocomplete)autocomplete_options[attrname]=a.autocomplete[attrname];void 0!==jQuery.Autocompleter?(t(r.fake_input).autocomplete(a.autocomplete_url,a.autocomplete),t(r.fake_input).bind("result",r,function(e,n){n&&t("#"+i).addTag(n[0]+"",{focus:!0,unique:a.unique})})):void 0!==jQuery.ui.autocomplete&&(t(r.fake_input).autocomplete(autocomplete_options),t(r.fake_input).bind("autocompleteselect",r,function(e,i){return t(e.data.real_input).addTag(i.item.value,{focus:!0,unique:a.unique}),!1
+}))}else t(r.fake_input).bind("blur",r,function(e){var i=t(this).attr("data-default");return""!=t(e.data.fake_input).val()&&t(e.data.fake_input).val()!=i?e.data.minChars<=t(e.data.fake_input).val().length&&(!e.data.maxChars||e.data.maxChars>=t(e.data.fake_input).val().length)&&t(e.data.real_input).addTag(t(e.data.fake_input).val(),{focus:!0,unique:a.unique}):(t(e.data.fake_input).val(t(e.data.fake_input).attr("data-default")),t(e.data.fake_input).css("color",a.placeholderColor)),!1});t(r.fake_input).bind("keypress",r,function(e){return e.which==e.data.delimiter.charCodeAt(0)||13==e.which?(e.preventDefault(),e.data.minChars<=t(e.data.fake_input).val().length&&(!e.data.maxChars||e.data.maxChars>=t(e.data.fake_input).val().length)&&t(e.data.real_input).addTag(t(e.data.fake_input).val(),{focus:!0,unique:a.unique}),t(e.data.fake_input).resetAutosize(a),!1):void(e.data.autosize&&t(e.data.fake_input).doAutosize(a))}),r.removeWithBackspace&&t(r.fake_input).bind("keydown",function(e){if(8==e.keyCode&&""==t(this).val()){e.preventDefault();var i=t(this).closest(".tagsinput").find(".tag:last").text(),n=t(this).attr("id").replace(/_tag$/,"");i=i.replace(/[\s]+x$/,""),t("#"+n).removeTag(escape(i)),t(this).trigger("focus")}}),t(r.fake_input).blur(),r.unique&&t(r.fake_input).keydown(function(e){(8==e.keyCode||String.fromCharCode(e.which).match(/\w+|[áéíóúÁÉÍÓÚñÑ,/]+/))&&t(this).removeClass("not_valid")})}}),this},t.fn.tagsInput.updateTagsField=function(i,n){var a=t(i).attr("id");t(i).val(n.join(e[a]))},t.fn.tagsInput.importTags=function(a,r){t(a).val("");var o=t(a).attr("id"),s=r.split(e[o]);for(i=0;i<s.length;i++)t(a).addTag(s[i],{focus:!1,callback:!1});if(n[o]&&n[o].onChange){var l=n[o].onChange;l.call(a,a,s[i])}}}(jQuery);var activateMenu=function(t){$(".navbar-nav li").eq(t).addClass("active")},nbNewEvents=0,loadedAt=new Date;jQuery(function(t){var e=function(e){t("#favicon").attr("href","/dashboard/img/favicon"+(e?"_red":"")+".ico")},i=function(){t.getJSON("/api/checks/count",function(i){i.total>0?t("#all_up").show().children("strong").text(i.up):t("#all_up").hide(),i.down>0?(t("#all_down").show().children("strong").text(i.down),e(!0)):(t("#all_down").hide(),nbNewEvents>0&&e(!1)),nbNewEvents>0&&(document.title="("+nbNewEvents+") Uptime")})};socket.on("CheckEvent",function(){nbNewEvents++,i(),t("#check_summary").fadeOut().fadeIn().fadeOut().fadeIn(),t("#new_events a").html(nbNewEvents+" new event"+(nbNewEvents>1?"s":"")+" since "+moment(loadedAt).format("LLL"))}),i()}),function(t){function e(t,e,i,n){var a=i.lang();return a[t].call?a[t](i,n):a[t][e]}function i(t,e){return function(i){return s(t.call(this,i),e)}}function n(t){return function(e){var i=t.call(this,e);return i+this.lang().ordinal(i)}}function a(t,e,i){this._d=t,this._isUTC=!!e,this._a=t._a||null,this._lang=i||!1}function r(t){var e=this._data={},i=t.years||t.y||0,n=t.months||t.M||0,a=t.weeks||t.w||0,r=t.days||t.d||0,s=t.hours||t.h||0,l=t.minutes||t.m||0,h=t.seconds||t.s||0,c=t.milliseconds||t.ms||0;this._milliseconds=c+1e3*h+6e4*l+36e5*s,this._days=r+7*a,this._months=n+12*i,e.milliseconds=c%1e3,h+=o(c/1e3),e.seconds=h%60,l+=o(h/60),e.minutes=l%60,s+=o(l/60),e.hours=s%24,r+=o(s/24),r+=7*a,e.days=r%30,n+=o(r/30),e.months=n%12,i+=o(n/12),e.years=i,this._lang=!1}function o(t){return 0>t?Math.ceil(t):Math.floor(t)}function s(t,e){for(var i=t+"";i.length<e;)i="0"+i;return i}function l(t,e,i){var n,a=e._milliseconds,r=e._days,o=e._months;a&&t._d.setTime(+t+a*i),r&&t.date(t.date()+r*i),o&&(n=t.date(),t.date(1).month(t.month()+o*i).date(Math.min(n,t.daysInMonth())))}function h(t){return"[object Array]"===Object.prototype.toString.call(t)}function c(t,e){var i,n=Math.min(t.length,e.length),a=Math.abs(t.length-e.length),r=0;for(i=0;n>i;i++)~~t[i]!==~~e[i]&&r++;return r+a}function u(t,e,i,n){var a,r,o=[];for(a=0;7>a;a++)o[a]=t[a]=null==t[a]?2===a?1:0:t[a];return t[7]=o[7]=e,null!=t[8]&&(o[8]=t[8]),t[3]+=i||0,t[4]+=n||0,r=new Date(0),e?(r.setUTCFullYear(t[0],t[1],t[2]),r.setUTCHours(t[3],t[4],t[5],t[6])):(r.setFullYear(t[0],t[1],t[2]),r.setHours(t[3],t[4],t[5],t[6])),r._a=o,r}function d(t,e){var i,n,a=[];for(!e&&L&&(e=require("./lang/"+t)),i=0;i<I.length;i++)e[I[i]]=e[I[i]]||O.en[I[i]];for(i=0;12>i;i++)n=_([2e3,i]),a[i]=new RegExp("^"+(e.months[i]||e.months(n,""))+"|^"+(e.monthsShort[i]||e.monthsShort(n,"")).replace(".",""),"i");return e.monthsParse=e.monthsParse||a,O[t]=e,e}function f(t){var e="string"==typeof t&&t||t&&t._lang||null;return e?O[e]||d(e):_}function p(t){return t.match(/\[.*\]/)?t.replace(/^\[|\]$/g,""):t.replace(/\\/g,"")}function g(t){var e,i,n=t.match(P);for(e=0,i=n.length;i>e;e++)n[e]=ee[n[e]]?ee[n[e]]:p(n[e]);return function(a){var r="";for(e=0;i>e;e++)r+="function"==typeof n[e].call?n[e].call(a,t):n[e];return r}}function m(t,e){function i(e){return t.lang().longDateFormat[e]||e}for(var n=5;n--&&W.test(e);)e=e.replace(W,i);return Q[e]||(Q[e]=g(e)),Q[e](t)}function v(t){switch(t){case"DDDD":return Y;case"YYYY":return $;case"S":case"SS":case"SSS":case"DDD":return R;case"MMM":case"MMMM":case"dd":case"ddd":case"dddd":case"a":case"A":return B;case"Z":case"ZZ":return j;case"T":return U;case"MM":case"DD":case"YY":case"HH":case"hh":case"mm":case"ss":case"M":case"D":case"d":case"H":case"h":case"m":case"s":return N;default:return new RegExp(t.replace("\\",""))}}function x(t,e,i,n){var a,r;switch(t){case"M":case"MM":i[1]=null==e?0:~~e-1;break;case"MMM":case"MMMM":for(a=0;12>a;a++)if(f().monthsParse[a].test(e)){i[1]=a,r=!0;break}r||(i[8]=!1);break;case"D":case"DD":case"DDD":case"DDDD":null!=e&&(i[2]=~~e);break;case"YY":i[0]=~~e+(~~e>70?1900:2e3);break;case"YYYY":i[0]=~~Math.abs(e);break;case"a":case"A":n.isPm="pm"===(e+"").toLowerCase();break;case"H":case"HH":case"h":case"hh":i[3]=~~e;break;case"m":case"mm":i[4]=~~e;break;case"s":case"ss":i[5]=~~e;break;case"S":case"SS":case"SSS":i[6]=~~(1e3*("0."+e));break;case"Z":case"ZZ":n.isUTC=!0,a=(e+"").match(V),a&&a[1]&&(n.tzh=~~a[1]),a&&a[2]&&(n.tzm=~~a[2]),a&&"+"===a[0]&&(n.tzh=-n.tzh,n.tzm=-n.tzm)}null==e&&(i[8]=!1)}function y(t,e){var i,n,a=[0,0,1,0,0,0,0],r={tzh:0,tzm:0},o=e.match(P);for(i=0;i<o.length;i++)n=(v(o[i]).exec(t)||[])[0],n&&(t=t.slice(t.indexOf(n)+n.length)),ee[o[i]]&&x(o[i],n,a,r);return r.isPm&&a[3]<12&&(a[3]+=12),r.isPm===!1&&12===a[3]&&(a[3]=0),u(a,r.isUTC,r.tzh,r.tzm)}function b(t,e){var i,n,r,o,s,l=t.match(E)||[],h=99;for(r=0;r<e.length;r++)o=y(t,e[r]),n=m(new a(o),e[r]).match(E)||[],s=c(l,n),h>s&&(h=s,i=o);return i}function w(t){var e,i="YYYY-MM-DDT";if(X.exec(t)){for(e=0;4>e;e++)if(G[e][1].exec(t)){i+=G[e][0];break}return j.exec(t)?y(t,i+" Z"):y(t,i)}return new Date(t)}function M(t,e,i,n,a){var r=a.relativeTime[t];return"function"==typeof r?r(e||1,!!i,t,n):r.replace(/%d/i,e||1)}function k(t,e,i){var n=A(Math.abs(t)/1e3),a=A(n/60),r=A(a/60),o=A(r/24),s=A(o/365),l=45>n&&["s",n]||1===a&&["m"]||45>a&&["mm",a]||1===r&&["h"]||22>r&&["hh",r]||1===o&&["d"]||25>=o&&["dd",o]||45>=o&&["M"]||345>o&&["MM",A(o/30)]||1===s&&["y"]||["yy",s];return l[2]=e,l[3]=t>0,l[4]=i,M.apply({},l)}function T(t,e){_.fn[t]=function(t){var i=this._isUTC?"UTC":"";return null!=t?(this._d["set"+i+e](t),this):this._d["get"+i+e]()}}function S(t){_.duration.fn[t]=function(){return this._data[t]}}function F(t,e){_.duration.fn["as"+t]=function(){return+this/e}}for(var _,C,D="1.7.2",A=Math.round,O={},z="en",L="undefined"!=typeof module&&module.exports,I="months|monthsShort|weekdays|weekdaysShort|weekdaysMin|longDateFormat|calendar|relativeTime|ordinal|meridiem".split("|"),H=/^\/?Date\((\-?\d+)/i,P=/(\[[^\[]*\])|(\\)?(Mo|MM?M?M?|Do|DDDo|DD?D?D?|ddd?d?|do?|w[o|w]?|YYYY|YY|a|A|hh?|HH?|mm?|ss?|SS?S?|zz?|ZZ?|.)/g,W=/(\[[^\[]*\])|(\\)?(LT|LL?L?L?)/g,E=/([0-9a-zA-Z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]+)/gi,N=/\d\d?/,R=/\d{1,3}/,Y=/\d{3}/,$=/\d{1,4}/,B=/[0-9a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]+/i,j=/Z|[\+\-]\d\d:?\d\d/i,U=/T/i,X=/^\s*\d{4}-\d\d-\d\d(T(\d\d(:\d\d(:\d\d(\.\d\d?\d?)?)?)?)?([\+\-]\d\d:?\d\d)?)?/,q="YYYY-MM-DDTHH:mm:ssZ",G=[["HH:mm:ss.S",/T\d\d:\d\d:\d\d\.\d{1,3}/],["HH:mm:ss",/T\d\d:\d\d:\d\d/],["HH:mm",/T\d\d:\d\d/],["HH",/T\d\d/]],V=/([\+\-]|\d\d)/gi,Z="Month|Date|Hours|Minutes|Seconds|Milliseconds".split("|"),J={Milliseconds:1,Seconds:1e3,Minutes:6e4,Hours:36e5,Days:864e5,Months:2592e6,Years:31536e6},Q={},K="DDD w M D d".split(" "),te="M D H h m s w".split(" "),ee={M:function(){return this.month()+1},MMM:function(t){return e("monthsShort",this.month(),this,t)},MMMM:function(t){return e("months",this.month(),this,t)},D:function(){return this.date()},DDD:function(){var t=new Date(this.year(),this.month(),this.date()),e=new Date(this.year(),0,1);return~~((t-e)/864e5+1.5)},d:function(){return this.day()},dd:function(t){return e("weekdaysMin",this.day(),this,t)},ddd:function(t){return e("weekdaysShort",this.day(),this,t)},dddd:function(t){return e("weekdays",this.day(),this,t)},w:function(){var t=new Date(this.year(),this.month(),this.date()-this.day()+5),e=new Date(t.getFullYear(),0,4);return~~((t-e)/864e5/7+1.5)},YY:function(){return s(this.year()%100,2)},YYYY:function(){return s(this.year(),4)},a:function(){return this.lang().meridiem(this.hours(),this.minutes(),!0)},A:function(){return this.lang().meridiem(this.hours(),this.minutes(),!1)},H:function(){return this.hours()},h:function(){return this.hours()%12||12},m:function(){return this.minutes()},s:function(){return this.seconds()},S:function(){return~~(this.milliseconds()/100)},SS:function(){return s(~~(this.milliseconds()/10),2)},SSS:function(){return s(this.milliseconds(),3)},Z:function(){var t=-this.zone(),e="+";return 0>t&&(t=-t,e="-"),e+s(~~(t/60),2)+":"+s(~~t%60,2)},ZZ:function(){var t=-this.zone(),e="+";return 0>t&&(t=-t,e="-"),e+s(~~(10*t/6),4)}};K.length;)C=K.pop(),ee[C+"o"]=n(ee[C]);for(;te.length;)C=te.pop(),ee[C+C]=i(ee[C],2);for(ee.DDDD=i(ee.DDD,3),_=function(e,i){if(null===e||""===e)return null;var n,r;return _.isMoment(e)?new a(new Date(+e._d),e._isUTC,e._lang):(i?n=h(i)?b(e,i):y(e,i):(r=H.exec(e),n=e===t?new Date:r?new Date(+r[1]):e instanceof Date?e:h(e)?u(e):"string"==typeof e?w(e):new Date(e)),new a(n))},_.utc=function(t,e){return h(t)?new a(u(t,!0),!0):("string"==typeof t&&!j.exec(t)&&(t+=" +0000",e&&(e+=" Z")),_(t,e).utc())},_.unix=function(t){return _(1e3*t)},_.duration=function(t,e){var i,n=_.isDuration(t),a="number"==typeof t,o=n?t._data:a?{}:t;return a&&(e?o[e]=t:o.milliseconds=t),i=new r(o),n&&(i._lang=t._lang),i},_.humanizeDuration=function(t,e,i){return _.duration(t,e===!0?null:e).humanize(e===!0?!0:i)},_.version=D,_.defaultFormat=q,_.lang=function(t,e){var i;if(!t)return z;if((e||!O[t])&&d(t,e),O[t]){for(i=0;i<I.length;i++)_[I[i]]=O[t][I[i]];_.monthsParse=O[t].monthsParse,z=t}},_.langData=f,_.isMoment=function(t){return t instanceof a},_.isDuration=function(t){return t instanceof r},_.lang("en",{months:"January_February_March_April_May_June_July_August_September_October_November_December".split("_"),monthsShort:"Jan_Feb_Mar_Apr_May_Jun_Jul_Aug_Sep_Oct_Nov_Dec".split("_"),weekdays:"Sunday_Monday_Tuesday_Wednesday_Thursday_Friday_Saturday".split("_"),weekdaysShort:"Sun_Mon_Tue_Wed_Thu_Fri_Sat".split("_"),weekdaysMin:"Su_Mo_Tu_We_Th_Fr_Sa".split("_"),longDateFormat:{LT:"h:mm A",L:"MM/DD/YYYY",LL:"MMMM D YYYY",LLL:"MMMM D YYYY LT",LLLL:"dddd, MMMM D YYYY LT"},meridiem:function(t,e,i){return t>11?i?"pm":"PM":i?"am":"AM"},calendar:{sameDay:"[Today at] LT",nextDay:"[Tomorrow at] LT",nextWeek:"dddd [at] LT",lastDay:"[Yesterday at] LT",lastWeek:"[last] dddd [at] LT",sameElse:"L"},relativeTime:{future:"in %s",past:"%s ago",s:"a few seconds",m:"a minute",mm:"%d minutes",h:"an hour",hh:"%d hours",d:"a day",dd:"%d days",M:"a month",MM:"%d months",y:"a year",yy:"%d years"},ordinal:function(t){var e=t%10;return 1===~~(t%100/10)?"th":1===e?"st":2===e?"nd":3===e?"rd":"th"}}),_.fn=a.prototype={clone:function(){return _(this)},valueOf:function(){return+this._d},unix:function(){return Math.floor(+this._d/1e3)},toString:function(){return this._d.toString()},toDate:function(){return this._d},toArray:function(){var t=this;return[t.year(),t.month(),t.date(),t.hours(),t.minutes(),t.seconds(),t.milliseconds(),!!this._isUTC]},isValid:function(){return this._a?null!=this._a[8]?!!this._a[8]:!c(this._a,(this._a[7]?_.utc(this._a):_(this._a)).toArray()):!isNaN(this._d.getTime())},utc:function(){return this._isUTC=!0,this},local:function(){return this._isUTC=!1,this},format:function(t){return m(this,t?t:_.defaultFormat)},add:function(t,e){var i=e?_.duration(+e,t):_.duration(t);return l(this,i,1),this},subtract:function(t,e){var i=e?_.duration(+e,t):_.duration(t);return l(this,i,-1),this},diff:function(t,e,i){var n,a=this._isUTC?_(t).utc():_(t).local(),r=6e4*(this.zone()-a.zone()),o=this._d-a._d-r,s=this.year()-a.year(),l=this.month()-a.month(),h=this.date()-a.date();return n="months"===e?12*s+l+h/30:"years"===e?s+(l+h/30)/12:"seconds"===e?o/1e3:"minutes"===e?o/6e4:"hours"===e?o/36e5:"days"===e?o/864e5:"weeks"===e?o/6048e5:o,i?n:A(n)},from:function(t,e){return _.duration(this.diff(t)).lang(this._lang).humanize(!e)},fromNow:function(t){return this.from(_(),t)},calendar:function(){var t=this.diff(_().sod(),"days",!0),e=this.lang().calendar,i=e.sameElse,n=-6>t?i:-1>t?e.lastWeek:0>t?e.lastDay:1>t?e.sameDay:2>t?e.nextDay:7>t?e.nextWeek:i;return this.format("function"==typeof n?n.apply(this):n)},isLeapYear:function(){var t=this.year();return t%4===0&&t%100!==0||t%400===0},isDST:function(){return this.zone()<_([this.year()]).zone()||this.zone()<_([this.year(),5]).zone()},day:function(t){var e=this._isUTC?this._d.getUTCDay():this._d.getDay();return null==t?e:this.add({d:t-e})},startOf:function(t){switch(t.replace(/s$/,"")){case"year":this.month(0);case"month":this.date(1);case"day":this.hours(0);case"hour":this.minutes(0);case"minute":this.seconds(0);case"second":this.milliseconds(0)}return this},endOf:function(t){return this.startOf(t).add(t.replace(/s?$/,"s"),1).subtract("ms",1)},sod:function(){return this.clone().startOf("day")},eod:function(){return this.clone().endOf("day")},zone:function(){return this._isUTC?0:this._d.getTimezoneOffset()},daysInMonth:function(){return _.utc([this.year(),this.month()+1,0]).date()},lang:function(e){return e===t?f(this):(this._lang=e,this)}},C=0;C<Z.length;C++)T(Z[C].toLowerCase(),Z[C]);T("year","FullYear"),_.duration.fn=r.prototype={weeks:function(){return o(this.days()/7)},valueOf:function(){return this._milliseconds+864e5*this._days+2592e6*this._months},humanize:function(t){var e=+this,i=this.lang().relativeTime,n=k(e,!t,this.lang()),a=0>=e?i.past:i.future;return t&&(n="function"==typeof a?a(n):a.replace(/%s/i,n)),n},lang:_.fn.lang};for(C in J)J.hasOwnProperty(C)&&(F(C,J[C]),S(C.toLowerCase()));F("Weeks",6048e5),L&&(module.exports=_),"undefined"==typeof ender&&(this.moment=_),"function"==typeof define&&define.amd&&define("moment",[],function(){return _})}.call(this);var StatNavigation=function(t){this.interval=t,this.init()};StatNavigation.prototype.init=function(){this.interval.on("refresh-stat",function(){var t=$("#secondaryNav"),e=this.stat;e&&e.availability&&(t.find(".availability").text(e.availability.replace(".000","")),t.find(".responsiveness").text(e.responsiveness.replace(".000","")),t.find(".avgRespTime").text(e.responseTime),e.downtime?(t.find(".downtime").text(moment.duration(e.downtime,"seconds").humanize()),t.find(".downtime").parentsUntil("li").show(),$("#events").show()):(t.find(".downtime").parentsUntil("li").hide(),$("#events").hide()))}),$("#secondaryNav").affix({offset:$("#dateNavigation").position()})};var uptimeBar=function(){var t=function(t,e,i,n){return'<div class="down" style="left:'+(t-i)/n*100+"%;width:"+(e-t)/n*100+'%" title="'+moment.duration(e-t).humanize()+" outage from "+moment(t).format("LLL")+" to "+moment(e).format("LLL")+'"></div>'},e=function(t,e,i,n){return'<div class="pause" style="left:'+(t-i)/n*100+"%;width:"+(e-t)/n*100+'%" title="'+moment.duration(e-t).humanize()+" pause from "+moment(t).format("LLL")+" to "+moment(e).format("LLL")+'"></div>'},i=function(t,e,i,n,a){if(1==i)return"";var r=1-.8*Math.pow(i,10);return'<div class="down" style="opacity: '+r+";left:"+(t-n)/a*100+"%;width:"+(e-t)/a*100+'%" title="'+(100*i).toFixed(3)+"% availability from "+moment(t).format("LL")+" to "+moment(e).format("LL")+'"></div>'},n=function(t,e){return"check"===t?a(e):"tag"===t?r(e):new Error("unkown type")},a=function(n){var a=n.from,r=n.to,o=n.check,s=n.periods||[],l=Date.now(),h=null,c=r-a,u=new Date(o.firstTested).valueOf(),d=new Date(o.lastChanged).valueOf(),f='<div class="uptimeBar">';if(s)for(var p=s.length,g=0;p>g;g++)s[g][0]>r||s[g][1]<a||(h=s[g][1],s[g][0]=Math.max(s[g][0],a),s[g][1]=Math.min(s[g][1],r),f+=0==s[g][2]||"undefined"==typeof s[g][2]?t(s[g][0],s[g][1],a,c):-1==s[g][2]?e(s[g][0],s[g][1],a,c):i(s[g][0],s[g][1],s[g][2],a,c));return(!o.isUp||o.isPaused)&&(null==h&&(h=Math.max(u,a)),d>h&&(f+=i(h,d,a,c),h=d),f+=o.isPaused?e(h,Math.min(l,r),a,c):t(h,Math.min(l,r),a,c)),u>a&&r>u&&(f+='<div style="background-color:white;left:0;width:'+(u-a)/c*100+'%"></div>'),l>a&&r>l&&(f+='<div style="background-color:white;right:0;width:'+(r-l)/c*100+'%"></div>'),f+="</div>"},r=function(n){var a=n.from,r=n.to,o=n.origin,s=n.periods,l='<div class="uptimeBar">',h=r-a;if(s)for(var c=s.length,u=0;c>u;u++)l+=0==s[u][2]||"undefined"==typeof s[u][2]?t(s[u][0],s[u][1],a,h):-1==s[u][2]?e(s[u][0],s[u][1],a,h):i(s[u][0],s[u][1],s[u][2],a,h);o>a&&r>o&&(l+='<div style="background-color:white;left:0;width:'+(o-a)/h*100+'%"></div>');var d=Date.now();return d>a&&r>d&&(l+='<div style="background-color:white;right:0;width:'+(r-d)/h*100+'%"></div>'),l+="</div>"};return n}();
