+++ app/dashboard/public/js/templates.js
@@ -1,0 +1,4 @@
+window.templates = {};
+templates.checkItem=function anonymous(locals,filters,escape,rethrow){function rethrow(a,e,n,s){var t=e.split("\n"),c=Math.max(s-3,0),i=Math.min(t.length,s+3),o=t.slice(c,i).map(function(a,e){var n=e+c+1;return(n==s?" >> ":"    ")+n+"| "+a}).join("\n");throw a.path=n,a.message=(n||"ejs")+":"+s+"\n"+o+"\n\n"+a.message,a}escape=escape||function(a){return String(a).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/'/g,"&#39;").replace(/"/g,"&quot;")};var __stack={lineno:1,input:'<tr>\n    <td data-sort="<%= check.name %>">\n        <a href="<%=  route + \'/checks/\' + check._id %>"><%= check.name %></a>\n        <a href="<%= check.url %>" target="_blank"><i class="fa fa-external-link"></i></a>\n    </td>\n    <% if (check.lastTested) { %>\n        <% if (check.isPaused) { %>\n    <td data-sort="<%= check.isUp ? check.uptime : -check.downtime %>">\n        <span class="badge alert-info"><i class="fa fa-pause"></i></span>\n        <span class="text-info"> for <span title="<%= new Date(check.lastChanged) %>"><%= moment(check.lastChanged).fromNow(true) %></span>\n    </td>\n        <% } else if (check.isUp) { %>\n    <td data-sort="<%= check.uptime %>">\n        <span class="badge alert-success"><i class="fa fa-arrow-up"></i></span>\n        <span class="text-success"> for <span title="<%= new Date(check.lastChanged) %>"><%= moment(check.lastChanged).fromNow(true) %></span>\n    </td>\n        <% } else { %>\n    <td data-sort="<%= -check.downtime %>">\n        <span class="badge alert-danger"><i class="fa fa-arrow-down"></i></span>\n        <span class="text-danger"> for <span title="<%= new Date(check.lastChanged) %>"><%= moment(check.lastChanged).fromNow(true) %></span>\n    </td>\n        <% } %>\n    <% } else { %>\n    <td data-sort="0">\n        <span class="badge alert-warning"><i class="fa fa-clock-o"></i></span>\n        <span class="text-warning">Just created</span>\n    </td>\n    <% } %>\n    <% if (check.qos) { %>\n    <td data-sort="<%= check.qos.availability %>">\n        <%- uptimeBar(\'check\', {\n            from: Date.now() - 24 * 60 * 60 * 1000,\n            to: Date.now(),\n            periods: check.qos.outages || [],\n            check: check\n         }) %>\n     </td>\n    <td data-sort="<%= check.qos.responseTime %>"><%= Math.round(check.qos.responseTime) %>ms</td>\n    <% } else { %>\n    <td data-sort="0">-</td>\n    <td data-sort="0">-</td>\n    <% } %>\n\n    <td>\n        <a href="<%= route + \'/checks/\' + check._id %>" class="btn btn-primary btn-xs">\n            <i class="fa fa-area-chart"></i>\n        </a>\n        <a href="<%= route + \'/checks/\' + check._id + \'/edit\' %>" class="btn btn-warning btn-xs">\n            <i class="fa fa-pencil"></i>\n        </a>\n        <a href="#" class="btn btn-danger btn-xs">\n            <i class="fa fa-times"></i>\n        </a>\n    </td>\n</tr>',filename:"/Users/an/www/uptime/app/dashboard/src/tpl/checkItem.ejs"};try{var buf=[];with(locals||{})!function(){buf.push('<tr>\n    <td data-sort="',escape((__stack.lineno=2,check.name)),'">\n        <a href="',escape((__stack.lineno=3,route+"/checks/"+check._id)),'">',escape((__stack.lineno=3,check.name)),'</a>\n        <a href="',escape((__stack.lineno=4,check.url)),'" target="_blank"><i class="fa fa-external-link"></i></a>\n    </td>\n    '),__stack.lineno=6,check.lastTested?(buf.push("\n        "),__stack.lineno=7,check.isPaused?(buf.push('\n    <td data-sort="',escape((__stack.lineno=8,check.isUp?check.uptime:-check.downtime)),'">\n        <span class="badge alert-info"><i class="fa fa-pause"></i></span>\n        <span class="text-info"> for <span title="',escape((__stack.lineno=10,new Date(check.lastChanged))),'">',escape((__stack.lineno=10,moment(check.lastChanged).fromNow(!0))),"</span>\n    </td>\n        "),__stack.lineno=12):check.isUp?(buf.push('\n    <td data-sort="',escape((__stack.lineno=13,check.uptime)),'">\n        <span class="badge alert-success"><i class="fa fa-arrow-up"></i></span>\n        <span class="text-success"> for <span title="',escape((__stack.lineno=15,new Date(check.lastChanged))),'">',escape((__stack.lineno=15,moment(check.lastChanged).fromNow(!0))),"</span>\n    </td>\n        "),__stack.lineno=17):(buf.push('\n    <td data-sort="',escape((__stack.lineno=18,-check.downtime)),'">\n        <span class="badge alert-danger"><i class="fa fa-arrow-down"></i></span>\n        <span class="text-danger"> for <span title="',escape((__stack.lineno=20,new Date(check.lastChanged))),'">',escape((__stack.lineno=20,moment(check.lastChanged).fromNow(!0))),"</span>\n    </td>\n        "),__stack.lineno=22),buf.push("\n    "),__stack.lineno=23):(buf.push('\n    <td data-sort="0">\n        <span class="badge alert-warning"><i class="fa fa-clock-o"></i></span>\n        <span class="text-warning">Just created</span>\n    </td>\n    '),__stack.lineno=28),buf.push("\n    "),__stack.lineno=29,check.qos?(buf.push('\n    <td data-sort="',escape((__stack.lineno=30,check.qos.availability)),'">\n        ',(__stack.lineno=31,uptimeBar("check",{from:Date.now()-864e5,to:Date.now(),periods:check.qos.outages||[],check:check})),'\n     </td>\n    <td data-sort="',escape((__stack.lineno=38,check.qos.responseTime)),'">',escape((__stack.lineno=38,Math.round(check.qos.responseTime))),"ms</td>\n    "),__stack.lineno=39):(buf.push('\n    <td data-sort="0">-</td>\n    <td data-sort="0">-</td>\n    '),__stack.lineno=42),buf.push('\n\n    <td>\n        <a href="',escape((__stack.lineno=45,route+"/checks/"+check._id)),'" class="btn btn-primary btn-xs">\n            <i class="fa fa-area-chart"></i>\n        </a>\n        <a href="',escape((__stack.lineno=48,route+"/checks/"+check._id+"/edit")),'" class="btn btn-warning btn-xs">\n            <i class="fa fa-pencil"></i>\n        </a>\n        <a href="#" class="btn btn-danger btn-xs">\n            <i class="fa fa-times"></i>\n        </a>\n    </td>\n</tr>')}();return buf.join("")}catch(err){rethrow(err,__stack.input,__stack.filename,__stack.lineno)}};
+templates.eventItem=function anonymous(locals,filters,escape,rethrow){function rethrow(e,n,a,s){var t=n.split("\n"),c=Math.max(s-3,0),i=Math.min(t.length,s+3),r=t.slice(c,i).map(function(e,n){var a=n+c+1;return(a==s?" >> ":"    ")+a+"| "+e}).join("\n");throw e.path=a,e.message=(a||"ejs")+":"+s+"\n"+r+"\n\n"+e.message,e}escape=escape||function(e){return String(e).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/'/g,"&#39;").replace(/"/g,"&quot;")};var __stack={lineno:1,input:"<% var label = (event.message == 'paused' || event.message == 'restarted') ? 'info' : (event.message == 'up') ? 'success' : 'danger' %>\n<li class=\"<%= highlightFrom && date.valueOf() > highlightFrom ? 'new' : ''%> event-item\">\n    <time class=\"time\" title=\"<%= date %>\"><%= date.format('LT') %></time>\n    <span class=\"badge alert-<%= label %> status\"><i class=\"fa\n    <% \n    switch (event.message) { \n        case 'paused': %>\n            fa-pause\n        <% \n            break; \n            case 'restarted': \n        %>\n            fa-play\n        <% \n            break;\n            case 'down': \n        %>\n            fa-arrow-down\n        <% \n            break;\n            case 'up': \n        %>\n            fa-arrow-up\n        <% \n            break; \n            default: \n        %>\n            fa-question\n    <% } %>\n    \"></i>\n    </span>\n    <span class=\"msg\">\n        <a href=\"<%= route %>/checks/<%= event.check._id %>\"><%= event.check.name %></a>\n        <a href=\"<%= event.check.url %>\" target=\"_blank\"><i class=\"fa fa-external-link\"></i></a>\n        <% \n        switch (event.message) { \n            case 'paused': \n            case 'restarted': \n        %>\n            was <%= event.message %>\n        <% \n            break;\n            case 'down': \n        %>\n            went down (<%= event.details %>)\n        <% \n            break;\n            case 'up': \n                if (event.downtime) { %>\n                    went back up after <%= moment.duration(event.downtime).humanize() %> of downtime\n                <% } else { %>\n                is now up\n\n            <%  }\n            break;\n            default: %>\n                unnown event\n        <% } %>\n    </span>\n</li>",filename:"/Users/an/www/uptime/app/dashboard/src/tpl/checkItem.ejs"};try{var buf=[];with(locals||{})!function(){buf.push(""),__stack.lineno=1;var e="paused"==event.message||"restarted"==event.message?"info":"up"==event.message?"success":"danger";switch(buf.push('\n<li class="',escape((__stack.lineno=2,highlightFrom&&date.valueOf()>highlightFrom?"new":"")),' event-item">\n    <time class="time" title="',escape((__stack.lineno=3,date)),'">',escape((__stack.lineno=3,date.format("LT"))),'</time>\n    <span class="badge alert-',escape((__stack.lineno=4,e)),' status"><i class="fa\n    '),__stack.lineno=5,event.message){case"paused":buf.push("\n            fa-pause\n        "),__stack.lineno=9;break;case"restarted":buf.push("\n            fa-play\n        "),__stack.lineno=14;break;case"down":buf.push("\n            fa-arrow-down\n        "),__stack.lineno=19;break;case"up":buf.push("\n            fa-arrow-up\n        "),__stack.lineno=24;break;default:buf.push("\n            fa-question\n    "),__stack.lineno=29}switch(buf.push('\n    "></i>\n    </span>\n    <span class="msg">\n        <a href="',escape((__stack.lineno=33,route)),"/checks/",escape((__stack.lineno=33,event.check._id)),'">',escape((__stack.lineno=33,event.check.name)),'</a>\n        <a href="',escape((__stack.lineno=34,event.check.url)),'" target="_blank"><i class="fa fa-external-link"></i></a>\n        '),__stack.lineno=35,event.message){case"paused":case"restarted":buf.push("\n            was ",escape((__stack.lineno=40,event.message)),"\n        "),__stack.lineno=41;break;case"down":buf.push("\n            went down (",escape((__stack.lineno=45,event.details)),")\n        "),__stack.lineno=46;break;case"up":event.downtime?(buf.push("\n                    went back up after ",escape((__stack.lineno=50,moment.duration(event.downtime).humanize()))," of downtime\n                "),__stack.lineno=51):(buf.push("\n                is now up\n\n            "),__stack.lineno=54);break;default:buf.push("\n                unnown event\n        "),__stack.lineno=58}buf.push("\n    </span>\n</li>")}();return buf.join("")}catch(err){rethrow(err,__stack.input,__stack.filename,__stack.lineno)}};
+templates.tagItem=function anonymous(locals,filters,escape,rethrow){function rethrow(a,t,e,n){var s=t.split("\n"),o=Math.max(n-3,0),r=Math.min(s.length,n+3),i=s.slice(o,r).map(function(a,t){var e=t+o+1;return(e==n?" >> ":"    ")+e+"| "+a}).join("\n");throw a.path=e,a.message=(e||"ejs")+":"+n+"\n"+i+"\n\n"+a.message,a}escape=escape||function(a){return String(a).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/'/g,"&#39;").replace(/"/g,"&quot;")};var __stack={lineno:1,input:'<tr>\n  <td data-sort="<%= tag.name %>"><a href="<%= route %>/tags/<%= encodeURIComponent(tag.name) %>"><%= tag.name %></a></td>\n  <td data-sort="<%= tag.availability %>"><%= (tag.availability * 100).toFixed(3).replace(\'.000\', \'\') %></td>\n  <td data-sort="<%= tag.availability %>"><%- uptimeBar(\'tag\', {\n        from: Date.now() - 24 * 60 * 60 * 1000,\n        to: Date.now(),\n        origin: new Date(tag.firstTested).valueOf(),\n        periods: tag.outages\n      }) %></td>\n  <td data-sort="<%= tag.responseTime %>"><%= Math.round(tag.responseTime) %>ms</td>\n</tr>',filename:"/Users/an/www/uptime/app/dashboard/src/tpl/checkItem.ejs"};try{var buf=[];with(locals||{})!function(){buf.push('<tr>\n  <td data-sort="',escape((__stack.lineno=2,tag.name)),'"><a href="',escape((__stack.lineno=2,route)),"/tags/",escape((__stack.lineno=2,encodeURIComponent(tag.name))),'">',escape((__stack.lineno=2,tag.name)),'</a></td>\n  <td data-sort="',escape((__stack.lineno=3,tag.availability)),'">',escape((__stack.lineno=3,(100*tag.availability).toFixed(3).replace(".000",""))),'</td>\n  <td data-sort="',escape((__stack.lineno=4,tag.availability)),'">',(__stack.lineno=4,uptimeBar("tag",{from:Date.now()-864e5,to:Date.now(),origin:new Date(tag.firstTested).valueOf(),periods:tag.outages})),'</td>\n  <td data-sort="',escape((__stack.lineno=10,tag.responseTime)),'">',escape((__stack.lineno=10,Math.round(tag.responseTime))),"ms</td>\n</tr>")}();return buf.join("")}catch(err){rethrow(err,__stack.input,__stack.filename,__stack.lineno)}};
