+++ app/dashboard/src/js/checkState.js
