+++ app/dashboard/src/js/dateInterval.js
