+++ app/dashboard/src/js/statNavigation.js
