+++ app/dashboard/src/js/uptimeBar.js
