+++ app/dashboard/src/scss/plugins/jquery.tagsinput.scss
