+++ app/dashboard/src/scss/style.scss
@@ -1,0 +1,5 @@
+@charset "utf-8";
+
+@import 
+  "plugins/jquery.tagsinput.scss",
+  "main.scss";
