+++ app/dashboard/views/events.ejs
@@ -1,5 +1,3 @@
-<script src="<%= route %>/javascripts/ejs.min.js"></script>
-
 <h1>Latest Events</h1>
 
 <div id="events">
